### 2024-03-12 #6472 @editorialbot
[PRE REVIEW]: PyGMI - a python package for geoscience modelling and interpretation
**Submitting author:** @Patrick-Cole (<a href="http://orcid.org/0000-0002-3563-8704">Patrick Cole</a>)
**Repository:** https://github.com/Patrick-Cole/pygmi
**Branch with paper.md** (empty if default branch):
**Version:** v3.2.8.0
