### 2024-03-13 #6472 @Patrick-Cole
Hi,

I chose reviewers on the list based on geophysics, python and earth sciences, but honestly, I am not fussy.

[Matt Hall](https://github.com/kwinkunks)
[leouieda](https://github.com/leouieda)
[TobbeTripitaka](https://github.com/TobbeTripitaka)
[Haipeng Li](https://github.com/Haipeng-ustc)
[margauxmouchene](https://github.com/margauxmouchene)
