### 2024-03-13 2d45b59
Update ratios.py
