+++ pygmi/rsense/ratios.py
@@ -291,7 +291,8 @@ class SatRatios(BasicModule):
         rlist += [r'B5/B3 Used in colour composites',
                   r'B4/B0 Used in colour composites',
                   r'B5/B1 Used in colour composites',
-                  r'B4/B7 Used in colour composites']
+                  r'B4/B7 Used in colour composites',
+                  r'B3/B4 Used in colour composites']
 
         # Landslides
 
