### 2024-03-13 ff386f5
Update ratios.py
