### 2024-03-14 16764c1
Update iodefs.py
ASTER zip import will attempt to assign a UTM crs if none is given.
