### 2024-03-19 f78c5e2
Bugfix
fixed an error loading in kml/kmz
