+++ pygmi/version.py
@@ -1,1 +1,1 @@
-__version__ = '3.2.8.7'
+__version__ = '3.2.8.8'
