### 2024-04-24 #6472 @boisgera
Hi @kthyng,

If @martinfleis was not available, you could consider inviting me as editor. Geoscience is not my area of expertise, but I belong to a "School of Mines" so many of my colleagues would be great potential reviewers.
