### 2024-04-30 78f1b13
bugfix
will skip calculating statistics for a band if it is empty.
