### 2024-05-06 26c3a77
Update iodefs.py
Fixed a bug causing some ASTER hdf data to not import properly in UTM
