### 2024-05-08 ca9b917
Update
SeisComp extended autoloc3 seismic data can be imported
