+++ pygmi/seis/iodefs.py
@@ -131,6 +131,8 @@ def str2int(inp):
     """
     if inp.strip() == '':
         return np.nan
+    if '.' in inp:
+        return int(inp.split('.')[0])
     return int(inp)
 
 
@@ -160,6 +162,7 @@ class ImportSeisan(BasicModule):
         if not nodialog:
             ext = ('SEISAN Format (*.out);;'
                    'SEISAN macroseismic files (*.macro);;'
+                   'SeisComP extended autoloc3 (*.txt);;'
                    'All Files (*.*)')
             self.ifile, _ = QtWidgets.QFileDialog.getOpenFileName(
                 self.parent, 'Open File', '.', ext)
@@ -171,6 +174,9 @@ class ImportSeisan(BasicModule):
         if '.macro' in self.ifile.lower():
             dat = importmacro(self.ifile)
             self.outdata['MacroSeis'] = dat
+        elif '.txt' in self.ifile.lower():
+            dat = importseiscomp(self.ifile, self.showlog)
+            self.outdata['Seis'] = dat
         else:
             dat = importnordic(self.ifile, self.showlog)
             self.outdata['Seis'] = dat
@@ -432,7 +438,7 @@ def importnordic(ifile, showlog=print):
     return dat
 
 
-def importseiscomp(ifile, showlog=print):
+def importseiscomp(ifile, showlog=print, prefmag='MLv'):
     """
     Import SeisComp data.
 
@@ -445,32 +451,218 @@ def importseiscomp(ifile, showlog=print):
 
     Returns
     -------
-    dat : list
+    sdat : list
         SEISAN Data.
 
     """
-    # idir = os.path.dirname(os.path.realpath(__file__))
-    # tfile = os.path.join(idir, r'descriptions.txt')
-
-    # with open(tfile, encoding='utf-8') as inp:
-    #     rnames = inp.read()
-
-    # rnames = rnames.split('\n')
-    breakpoint()
-
     with open(ifile) as pntfile:
         ltmp = pntfile.read()
 
     eventtxt = ltmp.split('Event:')
 
-    for event in eventtxt:
-        if event == '':
+    sdat = []
+
+    for eventrec in eventtxt:
+        event = {}
+        origin = {}
+        netmag = {}
+        phase = []
+        statmag = {}
+        sevent = {}
+        sevent['4'] = []
+
+        if eventrec == '':
             continue
-        elines = event.splitlines()
-        breakpoint()
 
+        elines = eventrec.splitlines()
+        currentsection = 'Event'
 
-    breakpoint()
+        for line in elines:
+            if line == '':
+                continue
+            if 'sta ' in line:
+                continue
+            if 'Origin:' in line:
+                currentsection = 'Origin'
+                continue
+            if 'Network magnitudes:' in line:
+                currentsection = 'Network magnitudes'
+                continue
+            if 'Phase arrivals:' in line:
+                currentsection = 'Phase arrivals'
+                continue
+            if 'Station magnitudes:' in line:
+                currentsection = 'Station magnitudes'
+                continue
+
+            if currentsection == 'Event':
+                ltype = line[:27].strip()
+                val = line[27:].strip()
+                event[ltype] = val
+
+            if currentsection == 'Origin':
+                ltype = line[:27].strip()
+                val = line[27:].strip()
+                origin[ltype] = val
+
+            if currentsection == 'Network magnitudes':
+                ltype = line[:14].strip()
+                val = line[14:].strip()
+                netmag[ltype] = val
+
+            if currentsection == 'Phase arrivals':
+                line2 = line.split()
[…]
