+++ pygmi/version.py
@@ -1,1 +1,1 @@
-__version__ = '3.2.8.9'
+__version__ = '3.2.8.10'
