### 2024-05-21 2e36160
Update
JPEG compression activated for GeoTIFFs
Cluster datasets can be cut
