### 2024-06-06 e1033cf
Update
Added import of EMIT data
Fixed a bug importing SeisComp data
Improved discrete colorbar function
Updated gradients help file.
