+++ pygmi/rsense/emit.py
@@ -1,0 +1,506 @@
+# -----------------------------------------------------------------------------
+# Name:        emit.py (part of PyGMI)
+#
+# Author:      Patrick Cole
+# E-Mail:      pcole@geoscience.org.za
+#
+# Copyright:   (c) 2024 Council for Geoscience
+# Licence:     GPL-3.0
+#
+# This file is part of PyGMI
+#
+# PyGMI is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# PyGMI is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+# -----------------------------------------------------------------------------
+"""
+EMIT is used to import EMIT satellite data into PyGMI.
+
+It uses code by Erik Bolch, ebolch@contractor.usgs.gov
+"""
+
+import os
+import winsound
+import datetime
+
+import numpy as np
+import xarray as xr
+from pyproj.crs import CRS
+
+from pygmi.raster.datatypes import Data
+
+# needs xarray, h5netcdf, rioxarray
+
+
+def emit_xarray(filepath, ortho=False, qmask=None, unpacked_bmask=None):
+    """
+    EMIT xarray.
+
+    This function utilizes other functions in this module to streamline
+    opening an EMIT dataset as an xarray.Dataset.
+
+    Parameters
+    ----------
+    filepath : str
+        a filepath to an EMIT netCDF file.
+    ortho : bool, optional
+        Whether to orthorectify the dataset or leave in crosstrack/downtrack
+        coordinates. The default is False.
+    qmask : numpy array, optional
+        Output from the quality_mask function used to mask
+        pixels based on quality flags selected in that function. Any
+        non-orthorectified array with the proper crosstrack and downtrack
+        dimensions can also be used. The default is None.
+    unpacked_bmask : numpy array, optional
+        From the band_mask function, used to mask band-specific pixels that
+        have been interpolated. The default is None.
+
+    Returns
+    -------
+    out_xr : xarray.Dataset
+        Dataset constructed based on the parameters provided.
+
+    """
+    # Grab granule filename to check product
+    granule_id = os.path.splitext(os.path.basename(filepath))[0]
+
+    # Read in Data as Xarray Datasets
+    engine, wvl_group = "h5netcdf", None
+
+    ds = xr.open_dataset(filepath, engine=engine)
+    loc = xr.open_dataset(filepath, engine=engine, group="location")
+
+    # Check if mineral dataset and read in groups (only ds/loc for minunc)
+
+    if "L2B_MIN_" in granule_id:
+        wvl_group = "mineral_metadata"
+    elif "L2B_MINUNC" not in granule_id:
+        wvl_group = "sensor_band_parameters"
+
+    wvl = None
+
+    if wvl_group:
+        wvl = xr.open_dataset(filepath, engine=engine, group=wvl_group)
+
+    # Building Flat Dataset from Components
+    data_vars = {**ds.variables}
+
[…]
