### 2024-06-06 96e5779
Update
Added EHGA filter to tilt angle module
