### 2024-06-24 #6472 @kthyng
@boisgera Maybe some of the authors of these similar papers could be relevant?
