### 2024-06-24 #6472 @boisgera
> @boisgera Maybe some of the authors of these similar papers could be relevant?

Yes, thank you @kthyng!

I was pushing for reviewers coming from my institution so far because I thought that it was worth introducing some new people to JOSS reviews. I also think that it would be beneficial to these researchers (so far the people which have accepted to review for JOSS on my demand like the process a lot better than the classic review process).

But it's now likely that I'll have to contact external reviewers already familiar with JOSS. I have sent today e-mails to the last 2 "obvious" potential reviewers internally and if its doesn't pan out, the authors on your liste are next!
