### 2024-06-27 ce3e09e
bugfix
Fixed a bug with sunshading, when clicking on interface.
