### 2024-07-02 1560578
Update pfinvert.py
SimPEG API changes
