### 2024-07-07 #6472 @boisgera
> But it's now likely that I'll have to contact external reviewers already familiar with JOSS. I have sent today e-mails to the last 2 "obvious" potential reviewers internally and if its doesn't pan out, the authors on your liste are next!

Started to reach out for JOSS paper authors.
