### 2024-07-22 #7019 @editorialbot
[REVIEW]: PyGMI - a python package for geoscience modelling and interpretation
**Submitting author:** @Patrick-Cole (<a href="http://orcid.org/0000-0002-3563-8704">Patrick Cole</a>)
**Repository:** https://github.com/Patrick-Cole/pygmi
**Branch with paper.md** (empty if default branch):
**Version:** 3.2.9.7
