### 2024-07-26 Patrick-Cole/pygmi#17 @AlexanderJuestel
JOSS Review - Comments on Github Repository
**This is not a complete list yet**

Dear @Patrick-Cole,

please already find some remarks regarding the Repository for your JOSS review at https://github.com/openjournals/joss-reviews/issues/7019

- The Wiki for PyGMI is quite hard to find. I would suggest adding a link in the `About` section on the right hand side of the repository.
- I would include some keywords in the `About` section on the right hand side of the repository.
- A citation file is missing.
- A code of conduct file is missing
- A developers guide and/or contributing file is missing, it should also outline if there are any other dependencies you have to install when you want to contribute to `PyGMI`.
- You provide a `pyproject.toml` file but users may also want to install the package from a `requirements.txt` or `environment.yml` file

Readme:
- The first header should be the name of the package. Currently, it is not machine-readable and limits users with accessibility issues.
- I am missing Badges in the Readme.
- I would add a section with resources explicitly linking to the wiki page of PyGMI and providing links to other necessary files (contributing, developers guide, license)
- I am missing a citation section
- I am missing a reference section
