### 2024-07-26 Patrick-Cole/pygmi#18 @AlexanderJuestel
JOSS Review - Installation of PyGMI
Dear @Patrick-Cole,

this relates to https://github.com/openjournals/joss-reviews/issues/7019

trying to install PyGMI in a fresh environment using `pip install pygmi` results in the following error.

Is Microsoft Visual C++ a must have requirement? I cannot install it on my working laptop due to admin restrictions.

```ruby
      error: Microsoft Visual C++ 14.0 or greater is required. Get it with "Microsoft C++ Build Tools": https://visualstudio.microsoft.com/visual-cpp-build-tools/
      [end of output]

  ERROR: Failed building wheel for gdal
  Building wheel for mtpy (setup.py) ... done
  Created wheel for mtpy: filename=mtpy-1.1.5-py3-none-any.whl size=1183169 sha256=8f14d045df65cbdf9e194c564131d1e2ecd4becd6c775e53bca6587f5e1df508
  Stored in directory: c:\users\ale93371\appdata\local\pip\cache\wheels\d5\2c\cd\da308d41e2f02f4a51457de99ae0ac707475232b9abb6ff6d4
  Building wheel for SimPEG (pyproject.toml) ... done
  Created wheel for SimPEG: filename=simpeg-0.22.1-py3-none-any.whl size=553830 sha256=7e8d36df7ccd3c77cd829665c6ae880fd6f5b2c826eed83736cb787ffd2ff8de
  Stored in directory: c:\users\ale93371\appdata\local\pip\cache\wheels\8a\5f\07\c50be773c0281deac5b6573b4460be2f5bec55da93f31a4088
  Building wheel for pymatsolver (setup.py) ... done
  Created wheel for pymatsolver: filename=pymatsolver-0.2.0-py3-none-any.whl size=13630 sha256=18aa26d52ae822a4f8de418b0568f5e0329fa4dbf1938230658dcb8b213b8aed
  Stored in directory: c:\users\ale93371\appdata\local\pip\cache\wheels\80\86\f9\551506d5369a0d2f8971058df7d9e6eb698f9ad2aeb0d37635
Successfully built mtpy SimPEG pymatsolver
Failed to build gdal
ERROR: Could not build wheels for gdal, which is required to install pyproject.toml-based projects
```
