### 2024-07-26 Patrick-Cole/pygmi#19 @AlexanderJuestel
JOSS Review - Comments on Paper
**This is not a complete list yet**

This relates to https://github.com/openjournals/joss-reviews/issues/7019

Comments:
- Summary: It should be mentioned earlier what kind of data can be processed or modelled. This is currently mentioned in the last sentence of the summary.
- Summary: The scope of the implemented tools seems very broad from "creating voxel based three dimensional potential field models (how does this compare to the GemPy package? - https://github.com/gempy-project/gempy), to working with raster and vector data (rasterio and geopandas should be mentioned and referenced in the main text if you heavily build on them), to gravity, remote sensing, earthquake seismology

- Target audience is very briefly mentioned. Do you have examples where you have applied your package that you could mention here?

- Statement of Need: The statement of need quickly transitions into a description of what the package can do rather than outlining what the actual need for this package is. Available packages such as GemPy (strucutral modeling), PyGimli (Geophysical Inversion and Modeling Library), Harmonica (Forward modeling, inversion, and processing gravity and magnetic data) or SimPEG (Simulation and Parameter Estimation in Geophysics, actually a requirement) leafmap/geemap are not discussed or referenced.

--> relation to other work is missing

- What is the overall purpose of the package? Does it have a specific topic or is it more like a toolbox with different smaller tools (similar to my GemGIS package)? If so, it would be good to emphasize that.

- Statement of Need: If you present so much of the capability of PyGMI, I would recommend using one or two more figures outlining the functionality. Maybe create different sections that outline the functionality.

- Statement of Need: The last sentence mentions that PyGMI works with a GUI which is very nice. I would mention this much earlier as this is a feature many other packages do not have.

- You could also add a section `Resources` linking the wiki package, the repository, etc.

- Figure 1: Figure (with 3 subfigures) with current caption not clear. In the right picture, what doe the lines in the upper plot represent and what do the colored areas in the lower plot represent? (Legends are missing).
- Line 50: Spelling mistake --> solution
