### 2024-07-29 Patrick-Cole/pygmi#18 @Patrick-Cole
This is a requirement for building GDAL, which I struggled with. I found two solutions:
1) Download binaries from https://github.com/cgohlke/geospatial-wheels/
2) Use Anaconda, which solves this.
3) Use WinPython, which already has the GDAL binaries included. https://winpython.github.io/

I have in installation instructions. What do you recommend?

I have also added a new version of PyGMI on PyPi, with some bug fixes.
