### 2024-07-30 7aadb39
Update
Added SAVI, RDVI and GRVI to ratios.
Fixed a bug with seiscomp import.
