+++ pygmi/rsense/ratios.py
@@ -282,6 +282,12 @@ class SatRatios(BasicModule):
                   r'(B3-B4)/(B3+B4) NDWI/NDMI water in leaves',
                   r'(B1-B3)/(B1+B3) NDWI water bodies',
                   r'2.5*(B3-B2)/(B3+6.0*B2-7.5*B0+1) EVI',
+                  r'B3/B1 GRVI',
+                  r'(B3-B2)/sqrt(B3+B2) RDVI',
+                  r'1.5*(B3-B2)/(B3+B2+0.5) SAVI',
+                  r'B3A/B1 GRVI Landslide',
+                  r'(B3A-B2)/sqrt(B3A+B2A) RDVI Landslide',
+                  r'1.5*(B3-B2)/(B3+B2A+0.5) SAVI Landslide',
                   r'0.5*(2*B3+1-sqrt((2*B3+1)**2-8*(B3-B2))) MSAVI2',
                   r'(B3A-B4+B5)/(B3A+B4-B5) NMDI',
                   r'((B4+B2)-(B3+B0))/((B4+B2)+(B3+B0)) BSI']
@@ -908,21 +914,22 @@ def correct_bands(rlist, sensor, bfile=None):
     sdict = {}
 
     sdict['ASTER'] = {'B1': 'B1', 'B2': 'B2', 'B3': 'B3', 'B4': 'B4',
-                      'B3A': 'B3',
+                      'B3A': 'B3', 'B2A': 'B2',
                       'B5': 'B5', 'B6': 'B6', 'B7': 'B7', 'B8': 'B8',
                       'B9': 'B9', 'B10': 'B10', 'B11': 'B11', 'B12': 'B12',
                       'B13': 'B13', 'B14': 'B14'}
     sdict['Landsat 8 and 9 (OLI)'] = {'B0': 'B2', 'B1': 'B3', 'B2': 'B4',
                                       'B3': 'B5', 'B4': 'B6', 'B5': 'B7',
-                                      'B3A': 'B5'}
+                                      'B3A': 'B5', 'B2A': 'B4'}
     sdict['Landsat 7 (ETM+)'] = {'B0': 'B1', 'B1': 'B2', 'B2': 'B3',
                                  'B3': 'B4', 'B4': 'B5', 'B5': 'B7',
-                                 'B3A': 'B4'}
+                                 'B3A': 'B4', 'B2A': 'B3'}
     sdict['Landsat 4 and 5 (TM)'] = sdict['Landsat 7 (ETM+)']
     sdict['Sentinel-2'] = {'B0': 'B2', 'B1': 'B3', 'B2': 'B4', 'B3': 'B8',
-                           'B4': 'B11', 'B5': 'B12', 'B3A': 'B8A'}
+                           'B4': 'B11', 'B5': 'B12', 'B3A': 'B8A',
+                           'B2A': 'B5'}
     sdict['WorldView'] = {'B0': 'B2', 'B1': 'B3', 'B2': 'B5', 'B3': 'B7',
-                          'B3A': 'B7'}
+                          'B3A': 'B7', 'B2A': 'B5'}
     sdict['Unknown'] = {}
 
     if sensor == 'Landsat (All)':
