### 2024-08-02 4d31288
Bugfix
fixed a bug in change detection
