### 2024-08-02 49d860e
Update CITATION.cff
