### 2024-08-02 87f1adf
Update CITATION.cff
