### 2024-08-02 6ea8f20
Update CITATION.cff
