### 2024-08-02 73d413d
Update CITATION.cff
