### 2024-08-02 f01bbc1
Create CODE_OF_CONDUCT.md
