### 2024-08-05 ebe9656
Update misc.py
Bugfix - fixed a bug saving projects
