+++ README.rst
@@ -2,8 +2,8 @@ PyGMI
 =====
 
 
-[![PyPI](https://img.shields.io/badge/python-3-blue.svg)](https://www.python.org/downloads/)
-![PyPI - Version](https://img.shields.io/pypi/v/pygmi)
+.. image:: https://img.shields.io/pypi/v/pygmi
+   :alt: PyPI - Version
 
 
 Overview
