+++ README.rst
@@ -1,10 +1,10 @@
 PyGMI
 =====
 
-
+.. image:: https://img.shields.io/pypi/pyversions/pygmi
+   :alt: PyPI - Python Version
 .. image:: https://img.shields.io/pypi/v/pygmi
    :alt: PyPI - Version
-
 
 Overview
 --------
