+++ requirements.txt
@@ -1,0 +1,31 @@
+# Requirements
+
+discretize>=0.10.0
+fiona>=1.9.5
+gdal>=3.8.4
+geopandas>=0.14.4
+h5netcdf>=1.3.0
+matplotlib>=3.9.0
+mtpy>=1.1.5
+natsort>=8.4.0
+numba>=0.60.0
+numexpr>=2.10.1
+numpy>=1.26.4
+openpyxl>=3.1.2
+pandas>=2.2.2
+psutil>=6.0.0
+pyogrio>=0.9.0
+PyOpenGL>=3.1.7
+pyproj>=3.6.1
+PyQt5>=5.15.10
+pytest>=8.2.2
+rasterio>=1.3.9
+rioxarray>=0.15.6
+scikit-image>=0.24.0
+scikit-learn>=1.5.0
+scipy>=1.13.1
+shapely>=2.0.3
+shapelysmooth>=0.2.0
+simpeg>=0.21.1
+xarray>=2024.6.0
+xlrd>=2.0.1
