### 2024-08-06 1d59de8
Update iodefs.py
Bugfix, fixed a bug importing some ASTER data.
