### 2024-08-06 Patrick-Cole/pygmi#17 @Patrick-Cole
I have made the changes here. Two quick questions:
1) Is a citation section necessary on the readme, if it is on the GitHub about section? I have added references. My hope is that the main citation will become the JOSS reference (if accepted).
2) I already have links to the wiki page and license in the readme overview. Did you have something else in mind? I have added contributors her though as well.

I have changes some of the README.rst so that install instructions are hopefully clearer.
