+++ paper/paper.bib
@@ -223,3 +223,51 @@ journal = {Journal of Machine Learning Research}
    url = {http://repository.ias.ac.in/28174/ http://library.seg.org/doi/abs/10.1190/1.1444942},
    year = {2001},
 }
+
+@software{gillies_2019,
+  author =    {Sean Gillies and others},
+  organization = {Mapbox},
+  title =     {Rasterio: geospatial raster I/O for {Python} programmers},
+  year =      {2013--},
+  url = "https://github.com/rasterio/rasterio"
+}
+
+@software{kelsey_jordahl_2020_3946761,
+  author       = {Kelsey Jordahl and
+                  Joris Van den Bossche and
+                  Martin Fleischmann and
+                  Jacob Wasserman and
+                  James McBride and
+                  Jeffrey Gerard and
+                  Jeff Tratner and
+                  Matthew Perry and
+                  Adrian Garcia Badaracco and
+                  Carson Farmer and
+                  Geir Arne Hjelle and
+                  Alan D. Snow and
+                  Micah Cochran and
+                  Sean Gillies and
+                  Lucas Culbertson and
+                  Matt Bartos and
+                  Nick Eubank and
+                  maxalbert and
+                  Aleksey Bilogur and
+                  Sergio Rey and
+                  Christopher Ren and
+                  Dani Arribas-Bel and
+                  Leah Wasser and
+                  Levi John Wolf and
+                  Martin Journois and
+                  Joshua Wilson and
+                  Adam Greenhall and
+                  Chris Holdgraf and
+                  Filipe and
+                  François Leblanc},
+  title        = {geopandas/geopandas: v0.8.1},
+  month        = jul,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v0.8.1},
+  doi          = {10.5281/zenodo.3946761},
+  url          = {https://doi.org/10.5281/zenodo.3946761}
+}
