+++ paper/paper.md
@@ -18,8 +18,8 @@ bibliography: paper.bib
 
 # Summary
 
-Python Geoscience Modelling and Interpretation (PyGMI) is an open-source software development project, programmed in Python. The main aim of PyGMI is to provide a free geoscientific tool for processing, interpretation and modelling of geoscience data. It also allows for research into better ways to process, analyse and model data and as such is continually updated to implement these advances.
-Although is was originally intended to create voxel based three dimensional potential field models, it expanded to include tools for the processing and visualisation of raster data, import and visualisation of vector data (normally collected in surveys along lines), unsupervised and supervised cluster analysis, gravity processing, remote sensing processing and quality control tools for earthquake seismology.
+Python Geoscience Modelling and Interpretation (PyGMI) is an open-source software development project, working with a GUI, programmed in Python. The main aim of PyGMI is to provide a free geoscientific tool for processing, interpretation and modelling of geoscience data, including magnetic, gravity and remote sensing data. It also allows for research into better ways to process, analyse and model data and as such is continually updated to implement these advances.
+Although is was originally intended to create voxel based three dimensional potential field models, it expanded to include tools for the processing and visualisation of raster data, import and visualisation of vector data (normally collected in surveys along lines), unsupervised and supervised cluster analysis, gravity processing, remote sensing processing and quality control tools for earthquake seismology. It is therefore a toolbox of numerous techniques to acomplish this.
 
 # Statement of Need
 Geophysical data provide geoscientists with the ability to create a picture of the subsurface prior to more expensive endeavours such a drilling or for further research using complementary data sets. The best way to present such a picture is through 3D models. The original concept of PyGMI was a desire to have an effective 3D forward modelling package for magnetic and gravity data. There are a number of basic strategies to do this. One conventional approach follows the modelling of individual profiles, which are then “joined” up into 3D sections. Another is to form bodies from arbitrary polygonal shapes. However, using such methods, the time it takes to construct a 3D model can be prohibitive. Conventional polygonal based 3D modelling programs (3DS Max, Maya, Blender) are akin to CAD packages and require a steep learning curve, which is not always possible or tolerated by many scientists. PyGMI follows a voxel based approach (which is also used commonly in geophysical inversion), allowing the model to ‘drawn’ in on model slices or layers, much like using an art program. Each voxel is assigned a lithology from a user defined list, with its associate geophysical definitions.
@@ -27,9 +27,9 @@ There are many techniques which can be considered here [@Bhattacharyya1964; @Gup
  
 ![An example of a model created with PyGMI.\label{fig:model}](img/figure1.jpg)
 
-Standard raster functions such as equations (making use of the NumExpr library, https://github.com/pydata/numexpr), smoothing, normalisation, merging of datasets and reprojections [courtesy of @gdal2020] are included. Functions unique to potential fields such as tilt angle [@Cooper2006], visibility [@Cooper2005], reduction to the pole, sun shading [@Horn1981], IGRF calculations [@igrf2015, based on code by written by A. Zunde, USGS, S.R.C. Malin & D.R. Barraclough, Institute of Geological Sciences, United Kingdom and maintained by Stefan Maus, NOAA] have all been translated into or developed in python. The sun shading tool in particular allows for sunshade detail and light reflectance to be changed. These two parameters are not normally present in other software packages which normally only allow for the changing of sun dip and azimuth.
[…]
