### 2024-08-08 Patrick-Cole/pygmi#19 @Patrick-Cole
I have started making corrections. One of my challenges is keeping to the recommended paper length, so I have tried to make brief edits. Let me know if you require more once I am complete. One question, regarding adding more figures to statement of need - do you mean an infographic showing functionality, or perhaps a figure of the main interface? For now I have added to images from some of the functions.
