### 2024-08-08 #7019 @Patrick-Cole
I have started making corrections and am responding directly to the comments on my repository above. I hope this is okay.
