+++ CONTRIBUTING.md
@@ -1,0 +1,29 @@
+# Contributing
+
+We are always keen to have new contributions to PyGMI, whether by questions, bug reports, feature requests and new code. 
+Please note the [GNU General Public Licence v3.0](https://github.com/Patrick-Cole/pygmi/blob/pygmi3/LICENSE.txt) will apply to all contributions.
+
+## Issues 
+
+Questions, bug reports and feature requests can be submitted [here](https://github.com/Patrick-Cole/pygmi/issues). Please first check whether  your issue has not already be raised, before adding a new issue.
+
+## Contributing New Code
+
+Code contributions are welcome, either addressing a bug or feature request, or adding a new feature.
+If you contribution is not on the issue tracker, then please [open the issue](https://github.com/Patrick-Cole/pygmi/issues/new/choose) link to that issue in your pull request.
+
+### Python code guidelines
+PyGMI follows the following coding guidelines:
+
+- [Style Guide for Python Code (PEP 8)](https://www.python.org/dev/peps/pep-0008/).
+- [Docstring conventions (PEP 257)](https://www.python.org/dev/peps/pep-0257/).
+
+All code should be tested and linted before submission.
+
+### Testing
+
+PyGMI uses [`pytest`](https://docs.pytest.org/). Please run all tests locally before creating a pull request. 
+All tests are located in the `test` folder and its subfolders.  
+
+### Pull Requests
+Pull request can be done using [the standard process](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request).
