### 2024-08-29 46df92b
Updates
Improved performance of visibility calculation.
Added possibility to specify temporary directory to the mosaic API, as well as assigning a default nodata value where none exists.
Added CRS functionality to reproject API
