+++ _sources/install.rst.txt
@@ -1,17 +1,10 @@
 Installation
-============
+------------
+The simplest installation of PyGMI is on Windows, using a pre-built installer at `64-bit <https://github.com/Patrick-Cole/pygmi/releases>`_.
 
-Instructions on how to install PyGMI can also be found on `Github <https://github.com/Patrick-Cole/pygmi>`_.
+If you prefer building from source, you can use PyPi or Conda. Note that if you are using PyPi,  certain libraries require binaries built. To do so or download binaries manually, please see the Linux or Windows section below.
 
-General (Not Anaconda)
-----------------------
-The easiest way to install pygmi if you are working in a python environment is to use the pip command as follows:
-
-   pip install pygmi
-
-This will download pygmi from PyPI and install it within your python repository. Depending on your operating system, and which libraries you already have installed, you may need to follow instructions in sections below. Please note the use of pip when installing PyGMI may cause Anaconda installations to break. Anaconda users should follow the instructions below.
-
-Running pygmi can be now done at the command prompt as follows:
+Once built, running pygmi can be done at the command prompt as follows:
 
    pygmi
 
@@ -25,28 +18,74 @@ If you prefer not to install pygmi as a library, download the source code and ex
 
    python quickstart.py
 
-Windows Users
--------------
-Installers are available in `64-bit <https://github.com/Patrick-Cole/pygmi/releases>`_
+Requirements
+^^^^^^^^^^^^
+PyGMI will run on both Windows and Linux. It should be noted that the main development is done in Python 3.12 on Windows.
 
-Alternatively, you can use the instructions above to run PyGMI with your local python installation. You may need to install some dependencies using downloaded binaries, because of compilation requirements. Therefore, if you do get an error, you can try installing precompiled binaries before installing PyGMI. You may need to install the `Microsoft Visual C++ Redistributable for Visual Studio 2015, 2017 and 2019 <https://support.microsoft.com/en-us/help/2977003/the-latest-supported-visual-c-downloads>`_.
+PyGMI should still work with Python 3.11.
 
-Examples of binaries you may need to get are:
+PyGMI is developed and has been tested with the following libraries in order to function:
 
-* numexpr
-* numba
-* llvmlite
-* GDAL
-* discretize
+* python 3.12.4
+* discretize 0.10.0
+* fiona 1.9.5
+* gdal 3.8.4
+* geopandas 0.14.4
+* matplotlib 3.9.0
+* mtpy 1.1.5
+* natsort 8.4.0
+* numba 0.60.0
+* numexpr 2.10.1
+* numpy 1.26.4
+* openpyxl 3.1.2
+* pandas 2.2.2
+* psutil 6.0.0
+* pyogrio 0.9.0
+* pyopengl 3.1.7
+* pyproj 3.6.1
+* PyQt5 5.15.10
+* pytest 8.2.2
+* rasterio 1.3.9
+* scikit-image 0.24.0
+* scikit-learn 1.5.0
+* scipy 1.13.1
+* shapely 2.0.3
+* shapelysmooth 0.2.0
+* SimPEG 0.21.1
+* xlrd 2.0.1
+* xarray 2024.6.0
+* h5netcdf 1.3.0
+* rioxarray 0.15.6
+
+PyPi - Windows
+^^^^^^^^^^^^^^
+Windows users can use the `WinPython <https://winpython.github.io/>`_ distribution as an alternative to Anaconda. It comes with most libraries preinstalled, so using pip should be sufficient. 
+
[…]
