+++ paper/paper.md
@@ -27,9 +27,15 @@ There are many techniques which can be considered here [@Bhattacharyya1964; @Gup
  
 ![(a) An example of a model created with PyGMI. (b) Top view of the model, the horizontal line is the profile being modelled. (c) The side view of the profile being modelled. The lines above the model represent the observed and calculated data. \label{fig:model}](img/figure1.jpg)
 
-Standard raster functions such as equations (making use of the NumExpr library, https://github.com/pydata/numexpr), smoothing, normalisation, merging of datasets and reprojections [courtesy of @gdal2020; @gillies_2019] are included. Functions unique to potential fields such as tilt angle [@Cooper2006], visibility [@Cooper2005], reduction to the pole, sun shading [@Horn1981], IGRF calculations [@igrf2015, based on code by written by A. Zunde, USGS, S.R.C. Malin & D.R. Barraclough, Institute of Geological Sciences, United Kingdom and maintained by Stefan Maus, NOAA] have all been translated into or developed in python. The sun shading tool in particular allows for sunshade detail and light reflectance to be changed. These two parameters are not normally present in other software packages which normally only allow for the changing of sun dip and azimuth. Vector data, such as magnetic line data or geochemical data, can be imported courtesy of the GeoPandas library [@kelsey_jordahl_2020_3946761].
+Standard raster functions such as equations (making use of the NumExpr library, https://github.com/pydata/numexpr), smoothing, normalisation, merging of datasets and reprojections [courtesy of @gdal2020; @gillies_2019] are included. Functions unique to potential fields such as tilt angle [@Cooper2006], visibility [@Cooper2005], reduction to the pole, sun shading [@Horn1981], IGRF calculations [@igrf2015, based on code by written by A. Zunde, USGS, S.R.C. Malin & D.R. Barraclough, Institute of Geological Sciences, United Kingdom and maintained by Stefan Maus, NOAA] have all been translated into or developed in python. The sun shading tool in particular allows for sunshade detail and light reflectance to be changed (\autoref{fig:sun}). These two parameters are not normally present in other software packages which normally only allow for the changing of sun dip and azimuth. Vector data, such as magnetic line data or geochemical data, can be imported courtesy of the GeoPandas library [@kelsey_jordahl_2020_3946761].
+
+![Sunshading being displayed on the raster interpretation module. \label{fig:sun}](img/rasterinterp.png)
+
 Gravity data processing has been developed according to the North American gravity database standards, as described by @Hinze2005. It allows for data to be imported from Scintrex CG-5 gravimeters and basic processing to Bouguer anomaly to be performed.
-Earthquake seismology QC tools for the open source SEISAN platform [@Havskov2020] have been implemented. These include various plotting functions including RMS of time residuals, histograms of events vs time, b-values and error ellipses. Filtering of events can be performed via a filtering tool. A fault plane solution tool is also implemented, translated from code written by Andy Michael and Oliver Boyd (http://www.ceri.memphis.edu/people/olboyd/Software/Software.html) and is capable of exporting results to the shapefile format.
+Earthquake seismology QC tools for the open source SEISAN platform [@Havskov2020] have been implemented. These include various plotting functions including RMS of time residuals, histograms of events vs time, b-values and error ellipses (\autoref{fig:seis}). Filtering of events can be performed via a filtering tool. A fault plane solution tool is also implemented, translated from code written by Andy Michael and Oliver Boyd (http://www.ceri.memphis.edu/people/olboyd/Software/Software.html) and is capable of exporting results to the shapefile format.
+
+![Error ellipses of seismic events shown on the QC tool. \label{fig:seis}](img/seisplots.png)
+
[…]
