+++ pygmi/raster/iodefs.py
@@ -36,6 +36,7 @@ import rasterio
 from rasterio.windows import Window
 from pyproj.crs import CRS
 
+from pygmi import menu_default
 from pygmi.raster.datatypes import Data
 from pygmi.raster.dataprep import lstack
 from pygmi.misc import ProgressBarText, ContextModule, BasicModule
@@ -1079,7 +1080,7 @@ def get_geosoft(hfile):
     return dat
 
 
-class ExportData(BasicModule):
+class ExportData(ContextModule):
     """
     Export Data.
 
@@ -1093,6 +1094,53 @@ class ExportData(BasicModule):
         super().__init__(parent)
 
         self.ofile = ''
+        self.ofilt = ''
+        self.exportdata = None
+
+        self.cmb_ofilt = QtWidgets.QComboBox()
+        self.le_ofile = QtWidgets.QLineEdit('')
+        self.cb_bandsort = QtWidgets.QCheckBox('Sort output bands')
+        self.lw_1 = QtWidgets.QListWidget()
+
+        self.setupui()
+
+    def setupui(self):
+        """
+        Set up UI.
+
+        Returns
+        -------
+        None.
+
+        """
+        gl_main = QtWidgets.QGridLayout(self)
+        buttonbox = QtWidgets.QDialogButtonBox()
+        helpdocs = menu_default.HelpButton('pygmi.raster.iodefs.exportraster')
+        pb_ofile = QtWidgets.QPushButton('Output File')
+
+        self.cb_bandsort.setChecked(False)
+        self.lw_1.setSelectionMode(QtWidgets.QAbstractItemView.MultiSelection)
+
+        buttonbox.setOrientation(QtCore.Qt.Horizontal)
+        buttonbox.setCenterButtons(True)
+        buttonbox.setStandardButtons(buttonbox.Cancel | buttonbox.Ok)
+
+        self.setWindowTitle(r'Export Raster Data')
+
+        gl_main.addWidget(self.le_ofile, 0, 0, 1, 1)
+        gl_main.addWidget(pb_ofile, 0, 1, 1, 1)
+
+        gl_main.addWidget(QtWidgets.QLabel('Output Bands:'), 1, 0, 1, 2)
+        gl_main.addWidget(self.lw_1, 2, 0, 1, 2)
+        gl_main.addWidget(self.cb_bandsort, 3, 0, 1, 2)
+
+        gl_main.addWidget(helpdocs, 8, 0, 1, 1)
+        gl_main.addWidget(buttonbox, 8, 1, 1, 3)
+
+        buttonbox.accepted.connect(self.acceptall)
+        buttonbox.rejected.connect(self.reject)
+        pb_ofile.pressed.connect(self.get_ofile)
+        # self.cb_ternary.clicked.connect(self.click_ternary)
 
     def run(self):
         """
@@ -1104,7 +1152,7 @@ class ExportData(BasicModule):
             True if successful, False otherwise.
 
         """
-        self.parent.process_is_active(True)
+        self.process_is_active(True)
 
         if 'Cluster' in self.indata:
             data = self.indata['Cluster']
@@ -1119,37 +1167,48 @@ class ExportData(BasicModule):
                     tmp.dataid = ('Membership of class ' + str(j+1)
                                   + ': '+tmp.dataid)
                     newdat.append(tmp)
-            data = newdat
+            self.exportdata = newdat
 
         elif 'Raster' in self.indata:
-            data = self.indata['Raster']
+            self.exportdata = self.indata['Raster']
         else:
             self.showlog('No raster data')
             self.parent.process_is_active(False)
             return False
 
-        ext = ('GeoTIFF compressed using DEFLATE (*.tif);;'
[…]
