### 2024-09-09 Patrick-Cole/pygmi#18 @boisgera
Does the pre-built installer work on Windows without admin rights? (Since this is advertised as the simplest installation method on Windows).
