### 2024-09-09 Patrick-Cole/pygmi#18 @Patrick-Cole
Honestly, I have admin rights, but it is just a standard installer for windows. If you are allowed to install software on the machine, then I don't see why not. Is there an issue?
