### 2024-09-09 Patrick-Cole/pygmi#19 @boisgera
Hi @Patrick-Cole,

I see that your current word count is 1025 (while the preferred range is 250-1000). As long as your paper does not end up "significantly longer than 1000 words" (say less than 1200 words?), you can give yourself a bit of leeway to improve it.

@AlexanderJuestel, could you chime in on this?
> One question, regarding adding more figures to statement of need - do you mean an infographic showing functionality, or perhaps a figure of the main interface? For now I have added to images from some of the functions.

Best Regards,

Sébastien
