### 2024-09-09 Patrick-Cole/pygmi#17 @boisgera
My two cents on the citation (@AlexanderJuestel feel free to disagree!):

> 1.    Is a citation section necessary on the readme, if it is on the GitHub about section? I have added references. My hope is that the main citation will become the JOSS reference (if accepted).

The `CITATION.cff` file that is automatically picked up as metadata by GitHub is pretty nice! I was not familiar with it, but now I see that some (scikit-learn, pytorch, etc.) are dropping the explicit citation in the README in favor of this metadata. So I'd say that, no, it's not necessary (it may still be beneficial for readers that are not aware of the metadata, but it's a much weaker motivation).

@AlexanderJuestel, can you chime in wrt:
> 2.    I already have links to the wiki page and license in the readme overview. Did you have something else in mind? I have added contributors her though as well.
