### 2024-09-09 Patrick-Cole/pygmi#18 @boisgera
> Honestly, I have admin rights, but it is just a standard installer for windows. If you are allowed to install software on the machine, then I don't see why not. Is there an issue?

Not that I know of! (not a Windows user.)

But I was wondering if @AlexanderJuestel ran into an issue with this method since he decided to go `pip install` instead (maybe that's just the old habits kicking in).

**Update.** Sorry, I misread your answer. Is there not on Windows a difference between installing system-wide (which will require admin rights) and installing *as a user* (which does not)? I guess that Alexander cannot install anything system-wide.
