### 2024-09-10 6635f83
Update
Added kaolin crystallinity and pyrophyllite
