+++ pygmi/rsense/features.py
@@ -33,7 +33,8 @@ feature['f1760'] = [1730, 1790]
 feature['f1800'] = [1740, 1820]
 feature['f2080'] = [2000, 2150]
 feature['f2160'] = [2138, 2188]
-feature['f2200'] = [2120, 2245]
+feature['f2200'] = [2180, 2245]
+# feature['f2200'] = [2120, 2245]
 feature['f2250'] = [2230, 2280]
 feature['f2290'] = [2270, 2330]
 feature['f2320'] = [2295, 2345]
@@ -45,21 +46,27 @@ ratio = {}
 ratio['NDVI'] = '(R860-R687)/(R860+R687)'
 ratio['dryveg'] = '(R2006+R2153)/(R2081+R2100)'
 ratio['albedo'] = 'R1650'
+ratio['rkgrpcrys'] = '((R2138+R2173)/R2156)/((R2156+R2190)/R2173)'
 
-ratio['r2350De'] = '(R2326+R2376)/(R2343+R2359)'
-ratio['r2160D2190'] = '(R2136+R2188)/(R2153+R2171)'  # Kaolin from non kaolin
-ratio['r2250D'] = '(R2227+R2275)/(R2241+R2259)'  # Chlorite epidote biotite
-ratio['r2380D'] = '(R2365+R2415)/(R2381+R2390)'  # Amphibole, talc
-ratio['r2320D'] = '(R2295+R2345)/(R2305+R2322)'
-ratio['r2330D'] = '(R2300+R2349)/(R2316+R2333)'  # (R2265+R2349)/(R2316+R2333)'  # MgOH and CO3
 ratio['r1100D'] = '(R921+R1650)/(R1020+R1231)'   # Ferrous iron
+ratio['r1550D'] = '(R1518+R1572)/(R1532+R1545)'  # Epidote
 ratio['r1750D'] = '(R1720+R1790)/(R1740+R1760)'  # Gypsum
 ratio['r1950D'] = '(R1900+R1980)/(R1930+R1960)'   # Water feature
-ratio['r1550D'] = '(R1518+R1572)/(R1532+R1545)'  # Epidote
+ratio['r2160D2190'] = '(R2136+R2188)/(R2153+R2171)'  # Kaolin from non kaolin
 ratio['r2200D'] = '(R2120+R2245)/(R2175+R2220)'
+ratio['r2250D'] = '(R2227+R2275)/(R2241+R2259)'  # Chlorite epidote biotite
+ratio['r2320D'] = '(R2295+R2345)/(R2305+R2322)'
+ratio['r2330D'] = '(R2300+R2349)/(R2316+R2333)'  # (R2265+R2349)/(R2316+R2333)'  # MgOH and CO3
+ratio['r2350De'] = '(R2326+R2376)/(R2343+R2359)'
+ratio['r2380D'] = '(R2365+R2415)/(R2381+R2390)'  # Amphibole, talc
 
 product = {}
 product['filter'] = ['NDVI < .25', 'dryveg < 1.015', 'albedo > 1000']
+product['pyrophyllite'] = ['f2160', 'f2160 > f2200',
+                           'f2080 > 0.01', 'f2200 > 0.01']
+
+
+
 product['white mica'] = ['f2200', 'r2350De > 1.02', 'r2160D2190 < 1.005',
                          'r2200D > 1.01']
 product['smectite'] = ['f2200', 'r2350De < 1.0', 'r2160D2190 < 1.005',
@@ -74,6 +81,7 @@ product['epidote'] = ['f1550', 'r1550D > 1.01', 'r2330D > 1.0', 'r2250D > 1.0']
 product['amphibole'] = ['f2390', 'r2380D > 1.02', 'r2320D > 1.01',
[…]
