+++ pygmi/rsense/hyperspec.py
@@ -43,6 +43,7 @@ from pygmi.raster.iodefs import get_raster
 from pygmi.raster.datatypes import numpy_to_pygmi
 from pygmi.raster.iodefs import export_raster
 from pygmi.rsense import features
+from pygmi.raster.modest_image import imshow
 
 
 class GraphMap(FigureCanvasQTAgg):
@@ -60,6 +61,7 @@ class GraphMap(FigureCanvasQTAgg):
 
         super().__init__(self.figure)
         self.setParent(parent)
+        self.rgb = True
 
         self.parent = parent
         self.datarr = []
@@ -87,7 +89,7 @@ class GraphMap(FigureCanvasQTAgg):
         None.
 
         """
-        dat = self.datarr[self.mindx]
+        dat = self.datarr[self.mindx].data
 
         if self.refl != 1.:
             dat = dat/self.refl
@@ -98,33 +100,35 @@ class GraphMap(FigureCanvasQTAgg):
         ax1 = self.figure.add_subplot(211)
         self.ax1 = ax1
 
-        ymin = dat.mean()-2*dat.std()
-        ymax = dat.mean()+2*dat.std()
+        self.compute_initial_figure()
 
-        if self.rotate is True:
-            self.csp = ax1.imshow(dat.T, vmin=ymin, vmax=ymax,
-                                  interpolation='none')
-            rows, cols = cols, rows
-        else:
-            self.csp = ax1.imshow(dat, vmin=ymin, vmax=ymax,
-                                  interpolation='none')
+        # ymin = dat.mean()-2*dat.std()
+        # ymax = dat.mean()+2*dat.std()
 
-        ax1.set_xlim((0, cols))
-        ax1.set_ylim((0, rows))
-        ax1.xaxis.set_visible(False)
-        ax1.yaxis.set_visible(False)
+        # if self.rotate is True:
+        #     self.csp = ax1.imshow(dat.T, vmin=ymin, vmax=ymax,
+        #                           interpolation='none')
+        #     rows, cols = cols, rows
+        # else:
+        #     self.csp = ax1.imshow(dat, vmin=ymin, vmax=ymax,
+        #                           interpolation='none')
 
-        ax1.xaxis.set_major_formatter(frm)
-        ax1.yaxis.set_major_formatter(frm)
+        # ax1.set_xlim((0, cols))
+        # ax1.set_ylim((0, rows))
+        # ax1.xaxis.set_visible(False)
+        # ax1.yaxis.set_visible(False)
 
-        if self.rotate is True:
-            ax1.plot(self.row, self.col, '+w')
-        else:
-            ax1.plot(self.col, self.row, '+w')
+        # ax1.xaxis.set_major_formatter(frm)
+        # ax1.yaxis.set_major_formatter(frm)
+
+        # if self.rotate is True:
+        #     ax1.plot(self.row, self.col, '+w')
+        # else:
+        #     ax1.plot(self.col, self.row, '+w')
 
         ax2 = self.figure.add_subplot(212)
 
-        prof = [i[self.row, self.col] for i in self.datarr]
+        prof = [i.data[self.row, self.col] for i in self.datarr]
 
         prof = np.ma.stack(prof).filled(0)/self.refl
 
@@ -166,6 +170,60 @@ class GraphMap(FigureCanvasQTAgg):
 
         self.figure.canvas.draw()
 
+    def compute_initial_figure(self):
+        """Compute initial figure."""
+        clippercu = 1
+        clippercl = 1
+        dat = self.datarr
+
+        redidx = (np.abs(self.wvl - 630)).argmin()
[…]
