+++ pygmi/rsense/features.py
@@ -31,10 +31,10 @@ feature['f1480'] = [1440, 1520]
 feature['f1550'] = [1510, 1610]
 feature['f1760'] = [1730, 1790]
 feature['f1800'] = [1740, 1820]
-feature['f2080'] = [2000, 2150]
-feature['f2160'] = [2138, 2188]
-feature['f2200'] = [2180, 2245]
-# feature['f2200'] = [2120, 2245]
+feature['f2080'] = [2060, 2100]
+feature['f2160'] = [2138, 2179]
+feature['f2200p'] = [2180, 2245]
+feature['f2200'] = [2120, 2245]
 feature['f2250'] = [2230, 2280]
 feature['f2290'] = [2270, 2330]
 feature['f2320'] = [2295, 2345]
@@ -62,17 +62,18 @@ ratio['r2380D'] = '(R2365+R2415)/(R2381+R2390)'  # Amphibole, talc
 
 product = {}
 product['filter'] = ['NDVI < .25', 'dryveg < 1.015', 'albedo > 1000']
-product['pyrophyllite'] = ['f2160', 'f2160 > f2200',
+product['pyrophyllite'] = ['f2160', 'f2160 > f2200p',
                            'f2080 > 0.01', 'f2200 > 0.01']
-
-
-
 product['white mica'] = ['f2200', 'r2350De > 1.02', 'r2160D2190 < 1.005',
                          'r2200D > 1.01']
+# product['white micanew'] = ['f2200', 'f2350 > 0.1', 'f2160 < f2200',
+#                             'f2200 > .01']
 product['smectite'] = ['f2200', 'r2350De < 1.0', 'r2160D2190 < 1.005',
                        'r2200D > 1.01']
 product['kaolin-2200'] = ['f2200', 'r2160D2190 > 1.005']
-product['kaolin-2160'] = ['f2160', 'r2160D2190 > 1.005']
+product['kaolin-2200new'] = ['f2200', 'f2160 < f2200p',
+                             'f2200 > 0.01', 'r2160D2190 > 1.005']
+# product['kaolin-2160'] = ['f2160', 'r2160D2190 > 1.005']
 product['chlorite-2250'] = ['f2250', 'r2330D > 1.0', 'r2250D > 1.005',
                             'r1550D < 1.01']
 product['chlorite-2330'] = ['f2330', 'r2330D > 1.0', 'r2250D > 1.005',
@@ -81,7 +82,7 @@ product['epidote'] = ['f1550', 'r1550D > 1.01', 'r2330D > 1.0', 'r2250D > 1.0']
 product['amphibole'] = ['f2390', 'r2380D > 1.02', 'r2320D > 1.01',
                         'r2160D2190 < 1.005', 'r2200D < 1.1', 'r2250D < 1.01']
 product['ferrous iron'] = ['r1100D']
-product['koalin group crystallinity'] = ['rkgrpcrys']
+product['kaolin group crystallinity'] = ['rkgrpcrys']
 product['ferric iron'] = ['f900']
 product['carbonate'] = ['f2320', 'r2250D < 1.009', 'r2380D < 1.0',
                         'r2200D < 1.0']
