+++ pyproject.toml
@@ -32,33 +32,26 @@ classifiers = [
     'Topic :: Software Development :: Libraries :: Python Modules',
 ]
 
-dependencies = ['discretize',
-                'gdal',
-                'fiona',
-                'geopandas',
-                'matplotlib',
-                'mtpy',
-                'natsort',
-                'numba',
-                'numexpr',
-                'numpy',
-                'openpyxl',
-                'pandas',
-                'pillow',
-                'psutil',
-                'pyogrio',
-                'pyopengl',
-                'pyproj',
-                'PyQt5',
-                'pytest',
-                'rasterio',
-                'scikit-image',
-                'scikit-learn',
-                'scipy',
-                'shapely',
-                'shapelysmooth',
-                'SimPEG',
-                'xlrd']
+dependencies = [
+                'fiona>=1.9.5',
+                'geopandas>=0.14.4',
+                'h5netcdf>=1.3.0',
+                'matplotlib>=3.9.0',
+                'mtpy>=1.1.5',
+                'natsort>=8.4.0',
+                'numexpr>=2.10.1',
+                'openpyxl>=3.1.2',
+                'psutil>=6.0.0',
+                'pyopengl>=3.1.7',
+                'pyqt5>=5.15.10',
+                'pytest>=8.2.2',
+                'rasterio>=1.3.9',
+                'rioxarray>=0.15.6',
+                'scikit-image>=0.24.0',
+                'shapelysmooth>=0.2.0',
+                'simpeg>=0.21.1',
+                'xarray>=2024.6.0',
+                ]
 
 [project.urls]
 homepage = 'http://patrick-cole.github.io/pygmi/'

[…]
