### 2024-09-18 2cc8119
Update
Updated requirements for simpler install
