+++ pyproject.toml
@@ -50,7 +50,6 @@ dependencies = [
                 'scikit-image>=0.24.0',
                 'shapelysmooth>=0.2.0',
                 'simpeg>=0.21.1',
-                'xarray>=2024.6.0',
                 ]
 
 [project.urls]
