### 2024-09-18 Patrick-Cole/pygmi#18 @Patrick-Cole
I am not sure that admin rights would have caused Alexander's problem. I am aware that the GDAL binary had to be built (or pre-built somewhere), so I think that caused the error. With that in mind, I have simplified the installation, since I noticed that GDAL binaries are included in the rasterio library. It should work now via pip, with the new version I uploaded today. I hope it helps.
