### 2024-09-30 Patrick-Cole/pygmi#18 @boisgera
I have installed the new version of pygmi with pip in verbose mode and AFAICT the Windows compiler is not used.
@AlexanderJuestel, could you `pip install pygmi` again and see how it goes?
