### 2024-10-01 e0bcdc0
Bugfix
Fixed a crash caused by an update of SimPEG API
