### 2024-10-01 6b47c36
Update
Changed default font size for graphs from 10 to 8.
