### 2024-10-04 e0a7578
Update
Added Clip to Zoom for raster data.
