+++ pygmi/raster/cliptozoom.py
@@ -1,0 +1,227 @@
+# -----------------------------------------------------------------------------
+# Name:        raster/cliptozoom.py (part of PyGMI)
+#
+# Author:      Patrick Cole
+# E-Mail:      pcole@geoscience.org.za
+#
+# Copyright:   (c) 2024 Council for Geoscience
+# Licence:     GPL-3.0
+#
+# This file is part of PyGMI
+#
+# PyGMI is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# PyGMI is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+# -----------------------------------------------------------------------------
+"""
+Clip to Zoom.
+
+This module allows a raster dataset to be clipped to the current zoomed
+extents.
+"""
+
+import numpy as np
+from PyQt5 import QtWidgets, QtCore
+from matplotlib import colormaps
+from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
+from matplotlib.figure import Figure
+from matplotlib.backends.backend_qt5 import NavigationToolbar2QT
+import matplotlib.colors as mcolors
+from shapely import Polygon
+import geopandas as gpd
+
+from pygmi.misc import frm, BasicModule
+from pygmi.raster.modest_image import imshow
+from pygmi.raster.dataprep import cut_raster
+
+
+class MyMplCanvas(FigureCanvasQTAgg):
+    """
+    Canvas for the actual plot.
+
+    Attributes
+    ----------
+    axes : matplotlib subplot
+    parent : parent
+        reference to the parent routine
+    """
+
+    def __init__(self, parent=None):
+        fig = Figure(layout='constrained', dpi=150)
+        self.axes = fig.add_subplot(111)
+        super().__init__(fig)
+
+    def update_raster(self, data1, cmap):
+        """
+        Update the raster plot.
+
+        Parameters
+        ----------
+        data1 : PyGMI raster Data
+            raster dataset to be used in contouring
+        cmap : str
+            Matplotlib colormap description
+
+        Returns
+        -------
+        None.
+
+        """
+        self.figure.clear()
+        self.axes = self.figure.add_subplot(111)
+
+        rdata = imshow(self.axes, data1.data, extent=data1.extent,
+                       cmap=colormaps[cmap], interpolation='nearest')
+
+        if not data1.isrgb:
+            rdata.set_clim_std(2.5)
+            cbar = self.figure.colorbar(rdata, format=frm)
+            cbar.set_label(data1.units)
+
+        if data1.crs is not None and data1.crs.is_geographic:
+            self.axes.set_xlabel('Longitude')
+            self.axes.set_ylabel('Latitude')
+        else:
+            self.axes.set_xlabel('Eastings')
+            self.axes.set_ylabel('Northings')
+
+        self.axes.xaxis.set_major_formatter(frm)
+        self.axes.yaxis.set_major_formatter(frm)
+
+        bbox = self.axes.get_window_extent()
+        dpi = self.figure.dpi
+        awidth = bbox.width / dpi
[…]
