### 2024-10-04 ab8252a
Update cliptozoom.py
