### 2024-10-08 ae5ee75
Update cliptozoom.py
bugfix for cliptozoom
