### 2024-10-11 454fbc3
Update
Comboboxes will make sure their lists are clear before adding items
ASTER, Landsat, Sentinel-2 and Worldview files now store azimuth and zenith information when imported
