+++ pygmi/raster/dataprep.py
@@ -1610,6 +1610,7 @@ class Metadata(ContextModule):
             tmp.datetime = i.datetime
 
         self.cmb_bandid.currentIndexChanged.disconnect()
+        self.cmb_bandid.clear()
         self.cmb_bandid.addItems(bandid)
         indx = self.cmb_bandid.currentIndex()
         self.oldtxt = self.cmb_bandid.itemText(indx)
