+++ pygmi/raster/show_table.py
@@ -118,6 +118,7 @@ class BasicStats(ContextModule):
         if len(self.data) == 1:
             self.cmb_1.hide()
 
+        self.cmb_1.clear()
         self.cmb_1.addItems(self.bands)
         self.tablewidget.setRowCount(data.shape[0])
         self.tablewidget.setColumnCount(data.shape[1])
@@ -308,6 +309,7 @@ class ClusterStats(ContextModule):
         if len(self.data) == 1:
             self.cmb_1.hide()
 
+        self.cmb_1.clear()
         self.cmb_1.addItems(self.bands)
         self.tablewidget.setRowCount(len(data))
         self.tablewidget.setColumnCount(len(data[0]))
