+++ pygmi/raster/dataprep.py
@@ -28,25 +28,25 @@ import tempfile
 import math
 import os
 import glob
-from collections import Counter
 from PyQt5 import QtWidgets, QtCore
 import numpy as np
 import pandas as pd
 from scipy.signal.windows import tukey
 import rasterio
 import rasterio.merge
-import pyproj
-from pyproj.crs import CRS, ProjectedCRS
-from pyproj.crs.coordinate_operation import TransverseMercatorConversion
+from pyproj.crs import CRS
 from rasterio.warp import calculate_default_transform, reproject
 from rasterio.mask import mask as riomask
 import geopandas as gpd
-from shapely.geometry import LineString
+from shapely import LineString
 
 from pygmi import menu_default
 from pygmi.raster.datatypes import Data
-from pygmi.misc import ProgressBarText, ContextModule, BasicModule
+from pygmi.misc import ContextModule, BasicModule
 from pygmi.raster.datatypes import numpy_to_pygmi
+from pygmi.raster.iodefs import get_raster, export_raster
+from pygmi.vector.dataprep import reprojxy
+from pygmi.raster.misc import GroupProj, lstack
 
 
 class Continuation(BasicModule):
@@ -431,6 +431,7 @@ class DataMerge(BasicModule):
         self.tmpdir = None
         self.is_import = True
         self.method = merge_median
+
         self.rb_first = QtWidgets.QRadioButton('First - copy first file over '
                                                'last file at overlap.')
         self.rb_last = QtWidgets.QRadioButton('Last - copy last file over '
@@ -474,7 +475,7 @@ class DataMerge(BasicModule):
         buttonbox = QtWidgets.QDialogButtonBox()
         helpdocs = menu_default.HelpButton('pygmi.raster.dataprep.datamerge')
         pb_idirlist = QtWidgets.QPushButton('Batch Directory')
-        pb_sfile = QtWidgets.QPushButton('Shapefile for boundary (optional)')
+        pb_sfile = QtWidgets.QPushButton('Shapefile or Raster for boundary (optional)')
 
         pixmapi = QtWidgets.QStyle.SP_DialogOpenButton
         icon = self.style().standardIcon(pixmapi)
@@ -482,7 +483,6 @@ class DataMerge(BasicModule):
         pb_idirlist.setIcon(icon)
         pb_sfile.setStyleSheet('text-align:left;')
         pb_idirlist.setStyleSheet('text-align:left;')
-
 
         self.cb_files_diff.setChecked(True)
         self.cb_shift_to_median.setChecked(False)
@@ -603,7 +603,7 @@ class DataMerge(BasicModule):
         None.
 
         """
-        ext = 'ESRI Shapefile (*.shp);;'
+        ext = 'Common formats (*.shp *.hdr *.tif);;'
 
         sfile, _ = QtWidgets.QFileDialog.getOpenFileName(
             self.parent, 'Open File', '.', ext)
@@ -695,11 +695,6 @@ class DataMerge(BasicModule):
             Success of routine.
 
         """
-        # The next line is only to avoid circular dependancies with merge
-        # function.
-
-        from pygmi.raster.iodefs import get_raster, export_raster
-
         indata = []
         if 'Raster' in self.indata:
             for i in self.indata['Raster']:
@@ -716,6 +711,10 @@ class DataMerge(BasicModule):
 
             for ifile in self.piter(ifiles):
                 indata += get_raster(ifile, piter=iter, metaonly=True)
+            
+            if len(indata) == len(ifiles):
+                self.singleband = True
+    
 
         if indata is None:
             self.showlog('No input datasets')
@@ -744,7 +743,19 @@ class DataMerge(BasicModule):
         else:
             crs = indata[0].crs
 
[…]
