+++ pygmi/seis/beachball.py
@@ -900,8 +900,6 @@ def _testfn2():
 
     plt.show()
 
-    breakpoint()
-
-
+    
 if __name__ == "__main__":
     _testfn()
