+++ pygmi/clust/segmentation.py
@@ -567,8 +567,6 @@ def _testfn2():
     plt.colorbar()
     plt.show()
 
-    breakpoint()
-
 
 if __name__ == "__main__":
     _testfn2()
