+++ pygmi/rsense/ratios.py
@@ -34,7 +34,7 @@ from PyQt5 import QtWidgets, QtCore
 from pygmi import menu_default
 from pygmi.rsense.iodefs import get_from_rastermeta, set_export_filename
 from pygmi.raster.iodefs import export_raster
-from pygmi.raster.dataprep import lstack
+from pygmi.raster.misc import lstack
 from pygmi.misc import BasicModule
 
 

[…]
