### 2024-10-14 a21b589
Bugfix
Fixed a memory leak with data on main interface
