+++ pygmi/rsense/topo.py
@@ -1,0 +1,309 @@
+# -----------------------------------------------------------------------------
+# Name:        topo.py (part of PyGMI)
+#
+# Author:      Patrick Cole
+# E-Mail:      pcole@geoscience.org.za
+#
+# Copyright:   (c) 2024 Council for Geoscience
+# Licence:     GPL-3.0
+#
+# This file is part of PyGMI
+#
+# PyGMI is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# PyGMI is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+# -----------------------------------------------------------------------------
+"""Calculate topographic correction for satellite data."""
+
+import os
+import sys
+import numpy as np
+from PyQt5 import QtWidgets, QtCore
+import matplotlib.pyplot as plt
+
+from pygmi.raster.misc import norm2, lstack, aspect2
+from pygmi.raster.iodefs import get_raster
+from pygmi import menu_default
+from pygmi.misc import BasicModule
+
+
+class TopoCorrect(BasicModule):
+    """Calculate topographic correction."""
+
+    def __init__(self, parent=None):
+        super().__init__(parent)
+
+        self.lw_indices = QtWidgets.QListWidget()
+
+        self.setupui()
+
+    def setupui(self):
+        """
+        Set up UI.
+
+        Returns
+        -------
+        None.
+
+        """
+        gl_main = QtWidgets.QGridLayout(self)
+        btn_invert = QtWidgets.QPushButton('Invert Selection')
+        buttonbox = QtWidgets.QDialogButtonBox()
+        helpdocs = menu_default.HelpButton('pygmi.rsense.change')
+        lbl_ratios = QtWidgets.QLabel('Indices:')
+
+        self.lw_indices.setSelectionMode(self.lw_indices.MultiSelection)
+
+        buttonbox.setOrientation(QtCore.Qt.Horizontal)
+        buttonbox.setCenterButtons(True)
+        buttonbox.setStandardButtons(buttonbox.Cancel | buttonbox.Ok)
+
+        self.setWindowTitle('Topographic Correction')
+
+        gl_main.addWidget(lbl_ratios, 1, 0, 1, 1)
+        gl_main.addWidget(self.lw_indices, 1, 1, 1, 1)
+        gl_main.addWidget(btn_invert, 2, 0, 1, 2)
+
+        gl_main.addWidget(helpdocs, 6, 0, 1, 1)
+        gl_main.addWidget(buttonbox, 6, 1, 1, 3)
+
+        buttonbox.accepted.connect(self.accept)
+        buttonbox.rejected.connect(self.reject)
+        self.lw_indices.clicked.connect(self.set_selected_indices)
+        btn_invert.clicked.connect(self.invert_selection)
+
+    def settings(self, nodialog=False):
+        """
+        Entry point into item.
+
+        Parameters
+        ----------
+        nodialog : bool, optional
+            Run settings without a dialog. The default is False.
+
[…]
