+++ pygmi/rsense/iodefs.py
@@ -40,123 +40,23 @@ import numpy as np
 import numexpr as ne
 import pandas as pd
 import geopandas as gpd
-from geopandas import GeoDataFrame
 from shapely.geometry import Point
 import rasterio
-from rasterio import Affine
 from natsort import natsorted
 from pyproj.crs import CRS
 
 from pygmi import menu_default
+from pygmi.misc import ProgressBarText, ContextModule, BasicModule
 from pygmi.raster.datatypes import Data, RasterMeta
 from pygmi.raster.iodefs import get_raster, export_raster
-from pygmi.raster.ginterp import histcomp, norm255, norm2, currentshader
-from pygmi.misc import ProgressBarText, ContextModule, BasicModule
-from pygmi.raster.dataprep import data_reproject
+from pygmi.raster.misc import histcomp, norm255, norm2, currentshader
+from pygmi.raster.misc import data_reproject, lstack
 from pygmi.vector.dataprep import reprojxy
-from pygmi.raster.misc import lstack
 from pygmi.rsense.emit import xr_to_pygmi, emit_xarray
 
 
 warnings.filterwarnings("ignore",
                         category=rasterio.errors.NotGeoreferencedWarning)
-
-EDIST = {1: 0.98331, 2: 0.9833, 3: 0.9833, 4: 0.9833,
-         5: 0.9833, 6: 0.98332, 7: 0.98333, 8: 0.98335,
-         9: 0.98338, 10: 0.98341, 11: 0.98345, 12: 0.98349,
-         13: 0.98354, 14: 0.98359, 15: 0.98365, 16: 0.98371,
-         17: 0.98378, 18: 0.98385, 19: 0.98393, 20: 0.98401,
-         21: 0.9841, 22: 0.98419, 23: 0.98428, 24: 0.98439,
-         25: 0.98449, 26: 0.9846, 27: 0.98472, 28: 0.98484,
-         29: 0.98496, 30: 0.98509, 31: 0.98523, 32: 0.98536,
-         33: 0.98551, 34: 0.98565, 35: 0.9858, 36: 0.98596,
-         37: 0.98612, 38: 0.98628, 39: 0.98645, 40: 0.98662,
-         41: 0.9868, 42: 0.98698, 43: 0.98717, 44: 0.98735,
-         45: 0.98755, 46: 0.98774, 47: 0.98794, 48: 0.98814,
-         49: 0.98835, 50: 0.98856, 51: 0.98877, 52: 0.98899,
-         53: 0.98921, 54: 0.98944, 55: 0.98966, 56: 0.98989,
-         57: 0.99012, 58: 0.99036, 59: 0.9906, 60: 0.99084,
-         61: 0.99108, 62: 0.99133, 63: 0.99158, 64: 0.99183,
-         65: 0.99208, 66: 0.99234, 67: 0.9926, 68: 0.99286,
-         69: 0.99312, 70: 0.99339, 71: 0.99365, 72: 0.99392,
-         73: 0.99419, 74: 0.99446, 75: 0.99474, 76: 0.99501,
-         77: 0.99529, 78: 0.99556, 79: 0.99584, 80: 0.99612,
-         81: 0.9964, 82: 0.99669, 83: 0.99697, 84: 0.99725,
-         85: 0.99754, 86: 0.99782, 87: 0.99811, 88: 0.9984,
-         89: 0.99868, 90: 0.99897, 91: 0.99926, 92: 0.99954,
-         93: 0.99983, 94: 1.00012, 95: 1.00041, 96: 1.00069,
-         97: 1.00098, 98: 1.00127, 99: 1.00155, 100: 1.00184,
[…]
