### 2024-11-25 Patrick-Cole/pygmi#17 @boisgera
My checks for the issues that have been addressed so far ; only a superficial analysis, @AlexanderJuestel feel free to disagree.

- [X] The Wiki for PyGMI is quite hard to find. I would suggest adding a link in the About section on the right hand side of the repository.
- [X]  I would include some keywords in the About section on the right hand side of the repository.
- [X]  A citation file is missing.
- [X] A code of conduct file is missing
- [X] A developers guide and/or contributing file is missing, it should also outline if there are any other dependencies you have to install when you want to contribute to PyGMI.
**AFAICT the requirements for the installation are also what contributors would need?**
- [X] You provide a pyproject.toml file but users may also want to install the package from a requirements.txt or environment.yml file
**Found a requirements.txt file**

Readme:

- [X] The first header should be the name of the package. Currently, it is not machine-readable and limits users with accessibility issues.
- [X] I am missing Badges in the Readme.
- [x] I would add a section with resources explicitly linking to the wiki page of PyGMI and providing links to other necessary files (contributing, developers guide, license)
- [ ] I am missing a citation section
**Is it still a requirement now that CITATION.cff files are picked up by GitHub and linked to in the project metadata? (with a built-in "Cite this repository" widget)**
- [X] I am missing a reference section
