### 2024-11-25 #7019 @boisgera
Hello everyone,

Let me try to recap the state of the review and see how to proceed.

- wrt [Installation issues](https://github.com/Patrick-Cole/pygmi/issues/18): @AlexanderJuestel  initially experienced some problems (the standard `pip install pygmi` on Windows was not enough) and after a discussion, @Patrick-Cole made some changes that should make it work. Can you confirm that @AlexanderJuestel? @AnkitBarik if by any chance you are a Windows user, could you give some feedback on this?

- I have made a quick check of the state of the [Comments on the GitHub repo issue](https://github.com/Patrick-Cole/pygmi/issues/17). Do we agree on this @AlexanderJuestel?

- wrt the [paper issue](https://github.com/Patrick-Cole/pygmi/issues/19). The latest changes have been made on Sept. 2 and [I then mentioned](https://github.com/Patrick-Cole/pygmi/issues/19#issuecomment-2337679303) that you still have a bit of wiggle room to make some changes. I have not check the latest version of the paper. @Patrick-Cole, did you manage to adress the comments made by @AlexanderJuestel? Otherwise could you say what's still in the works and if some extra discussions are needed on some points?

@AnkitBarik feel free to chime in of course, about those issues or any other feedback that you may have!

More generally, for everyone, do not hesitate to contact me if you need my help to push this review forward.

Best regards,

Sébastien
