### 2024-11-26 Patrick-Cole/pygmi#17 @Patrick-Cole
Hi,

Thanks for the comments, I have some questions:

_I would add a section with resources explicitly linking to the wiki page of PyGMI and providing links to other necessary files (contributing, developers guide, license)_
In the overview, I have "The PyGMI [Wiki](http://patrick-cole.github.io/pygmi/index.html) pages, include installation and full usage! Contributors can check this [link](https://github.com/Patrick-Cole/pygmi/blob/pygmi3/CONTRIBUTING.md) for ways to contribute." Is this acceptable.

_I am missing a citation section_
I have a references section. Should this be relabeled?
