### 2024-11-26 #7019 @Patrick-Cole
I think I made most of the changes. I have the following comments:

_Statement of Need: If you present so much of the capability of PyGMI, I would recommend using one or two more figures outlining the functionality. Maybe create different sections that outline the functionality._

I avoided adding sections given paper size limitations, and how much functionality PyGMI has. If you still prefer more sections, I can put them in and then get advice on paper length issues which result.

I see I missed adding a resources section, I will add that.

It is possible that  I have not given enough detail on other corrections, and will fix that once I get some feedback.

Once again, I appreciate the feedback. I have definitely learned a lot from it.
