### 2024-11-27 Patrick-Cole/pygmi#19 @Patrick-Cole
I have completed all changes to the paper, and posted a new version, thanks.
