### 2025-01-09 4c71447
Bugfix
Fixed a bug causing a crash in magnetic calculation
