### 2025-01-09 a110474
Bugfix
Negative null values can now be specified for gridding, Blanking distance now displays properly.
