### 2025-01-09 5945e9c
Update graphs.py
bugfix - fixed a display problem on correlation coefficient graph.
