### 2025-01-10 9855759
Update dataprep.py
Fixed a bug where raster band units could be set to the string 'None' in the metadata interface.
