### 2025-01-10 2a75823
Update cooper.py
Derivative ratio now accepts float order.
