### 2025-01-10 d701544
Bugfix
MarineCopper colormap now works for vector display
