+++ pygmi/vector/graphs.py
@@ -36,6 +36,7 @@ from matplotlib.figure import Figure
 from matplotlib.backends.backend_qt5 import NavigationToolbar2QT
 from matplotlib.ticker import StrMethodFormatter
 from matplotlib import rcParams
+from matplotlib.colors import ListedColormap
 import pandas as pd
 from sklearn.cluster import KMeans
 
@@ -43,6 +44,262 @@ from pygmi.misc import frm, ContextModule, discrete_colorbar
 
 rcParams['savefig.dpi'] = 300
 
+copper = np.array([[255., 236., 184.],
+                   [255., 230., 172.],
+                   [255., 225., 164.],
+                   [255., 221., 158.],
+                   [255., 218., 153.],
+                   [254., 216., 149.],
+                   [253., 213., 146.],
+                   [252., 212., 143.],
+                   [251., 210., 140.],
+                   [250., 208., 138.],
+                   [249., 206., 136.],
+                   [248., 205., 133.],
+                   [247., 203., 131.],
+                   [246., 202., 129.],
+                   [245., 201., 128.],
+                   [245., 199., 126.],
+                   [244., 198., 124.],
+                   [243., 197., 122.],
+                   [242., 196., 121.],
+                   [241., 195., 120.],
+                   [240., 194., 118.],
+                   [239., 193., 117.],
+                   [238., 192., 115.],
+                   [238., 191., 114.],
+                   [237., 190., 113.],
+                   [236., 189., 112.],
+                   [235., 188., 110.],
+                   [234., 187., 109.],
+                   [234., 186., 108.],
+                   [233., 185., 107.],
+                   [232., 184., 106.],
+                   [231., 183., 105.],
+                   [230., 182., 104.],
+                   [230., 181., 103.],
+                   [229., 181., 102.],
+                   [228., 180., 101.],
+                   [227., 179., 100.],
+                   [227., 178.,  99.],
+                   [226., 177.,  98.],
+                   [225., 177.,  97.],
+                   [224., 176.,  96.],
+                   [224., 175.,  95.],
+                   [223., 174.,  94.],
+                   [222., 174.,  93.],
+                   [222., 173.,  93.],
+                   [221., 172.,  92.],
+                   [220., 171.,  91.],
+                   [220., 171.,  90.],
+                   [219., 170.,  89.],
+                   [218., 169.,  88.],
+                   [217., 168.,  88.],
+                   [217., 168.,  87.],
+                   [216., 167.,  86.],
+                   [215., 166.,  85.],
+                   [215., 166.,  85.],
+                   [214., 165.,  84.],
+                   [213., 164.,  83.],
+                   [213., 164.,  82.],
+                   [212., 163.,  82.],
+                   [211., 162.,  81.],
+                   [211., 162.,  80.],
+                   [210., 161.,  80.],
+                   [210., 160.,  79.],
+                   [209., 160.,  78.],
+                   [208., 159.,  77.],
+                   [208., 158.,  77.],
+                   [207., 158.,  76.],
+                   [206., 157.,  75.],
+                   [206., 156.,  75.],
+                   [205., 156.,  74.],
+                   [204., 155.,  74.],
+                   [204., 154.,  73.],
+                   [203., 154.,  72.],
+                   [203., 153.,  72.],
+                   [202., 153.,  71.],
+                   [201., 152.,  70.],
+                   [201., 151.,  70.],
+                   [200., 151.,  69.],
+                   [199., 150.,  68.],
+                   [199., 149.,  68.],
+                   [198., 149.,  67.],
+                   [197., 148.,  67.],
[…]
