### 2025-01-14 a26f67a
bugfix
fixed a bug on the graph tool where ptp calculation should have been ma.ptp. Other locations updated as well.
