+++ pygmi/pfmod/tab_mext.py
@@ -308,7 +308,7 @@ class MextDisplay(QtWidgets.QDialog):
             curgrid = self.parent.inraster[ctxt]
 
             self.dsb_utlz.setValue(curgrid.data.max())
-            zextent = np.ptp(curgrid.data)+self.dsb_zcell.value()
+            zextent = np.ma.ptp(curgrid.data)+self.dsb_zcell.value()
             if zextent > self.dsb_zextent.value():
                 self.dsb_zextent.setValue(zextent)
 
