+++ pygmi/pfmod/pfinvert.py
@@ -344,7 +344,7 @@ class MagInvert(BasicModule):
             curgrid = self.inraster[ctxt]
 
             self.dsb_utlz.setValue(curgrid.data.max())
-            zextent = np.ptp(curgrid.data)+self.dsb_zcell.value()
+            zextent = np.ma.ptp(curgrid.data)+self.dsb_zcell.value()
             if zextent > self.dsb_zextent.value():
                 self.dsb_zextent.setValue(zextent)
 
