### 2025-01-15 4718f9c
Update
obj graph will no longer apply to supervised classification
bugfix - fixed scaling bug in segmentation.
