### 2025-01-16 54a9ee4
Update
Results from supervised classification will      have their class names correctly displayed on class maps and statistics.
