### 2025-01-17 #7019 @boisgera
Hi @Patrick-Cole, @AnkitBarik,

For your information, I am in contact with potential replacements for Alexander (since the JOSS process mandates two reviewers). I'll keep you informed!

Best regards,

Sébastien
