### 2025-01-27 0a521b5
Updates
Equation editor shows warning about negative logs and sqrt
Gradients now accepts float order
help files updated and added to fuzzy and crisp clustering
Membership plot has better labels for choices
