### 2025-01-31 591234f
Update
Clipping a vector using a polygon no longer needs identical projections.
Fixed a big where Feature Intersection Density geometries could not be used.
