### 2025-02-03 #7019 @boisgera
Jean-Marc Temmos (of [SEMANTIC TS](https://semantic-ts.com/)) has kindly accepted to review the project.

Thanks! 🙏
