+++ pygmi/mag/tiltdepth.py
@@ -38,10 +38,12 @@ import numpy as np
 
 from PyQt5 import QtWidgets
 from matplotlib import colormaps
+from matplotlib import cm
 from matplotlib.figure import Figure
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
 from matplotlib.backends.backend_qt5 import NavigationToolbar2QT
 from numba import jit
+import geopandas as gpd
 
 from pygmi.raster.datatypes import Data
 from pygmi.raster.cooper import vertical
@@ -201,6 +203,47 @@ class TiltDepth(BasicModule):
 
         return True
 
+    def change_cbar1(self):
+        """
+        Change the colour map for the colour bar.
+
+        Returns
+        -------
+        None.
+
+        """
+        zout = self.indata['Raster'][0]
+        txt = str(self.cmb_cbar.currentText())
+
+        self.figure.clear()
+        self.axes = self.figure.add_subplot(111)
+
+        self.axes.contour(self.X, self.Y, self.Z, [0])
+        self.axes.contour(self.X, self.Y, self.Z, [45], linestyles='dashed')
+        self.axes.contour(self.X, self.Y, self.Z, [-45], linestyles='dashed')
+
+        cmap = colormaps.get_cmap(txt)
+        cmap2 = np.array([cmap(i) for i in range(cmap.N)])
+        low = int(cmap.N*(45/180))
+        high = int(cmap.N*(135/180))
+        cmap2[low:high] = cmap2[int(cmap.N/2)]
+
+        cmap3 = cm.colors.ListedColormap(cmap2)
+        ims = self.axes.imshow(self.Z, extent=zout.extent, cmap=cmap3)
+
+        if self.x0 is not None:
+            i = 0
+            self.axes.plot(self.x1[i], self.y1[i], 'oy')
+            self.axes.plot(self.x0[i], self.y0[i], 'sy')
+            self.axes.plot(self.x2[i], self.y2[i], 'oy')
+
+        self.axes.xaxis.set_major_formatter(frm)
+        self.axes.yaxis.set_major_formatter(frm)
+
+        self.figure.colorbar(ims, format=frm)
+
+        self.figure.canvas.draw()
+
     def change_cbar(self):
         """
         Change the colour map for the colour bar.
@@ -210,10 +253,11 @@ class TiltDepth(BasicModule):
         None.
 
         """
-        if 'Raster' not in self.outdata:
+        if 'Vector' not in self.outdata:
             return
 
-        zout = self.outdata['Raster'][0]
+        gdf = self.outdata['Vector'][0]
+        zout = self.indata['Raster'][0]
         txt = str(self.cmb_cbar.currentText())
 
         self.figure.clear()
@@ -224,13 +268,31 @@ class TiltDepth(BasicModule):
         vmin = zout.data.mean() - 2.5*zout.data.std()
         vmax = zout.data.mean() + 2.5*zout.data.std()
 
-        ims = self.axes.imshow(zout.data, extent=zout.extent, cmap=cmap,
-                               interpolation='nearest', vmin=vmin, vmax=vmax)
+        self.axes.imshow(zout.data, extent=zout.extent, cmap='gray',
+                         interpolation='nearest', vmin=vmin, vmax=vmax)
+
+        cmap = colormaps.get_cmap(txt)
+        cmap2 = np.array([cmap(i) for i in range(cmap.N)])
+        low = int(cmap.N*(45/180))
+        high = int(cmap.N*(135/180))
[…]
