+++ pygmi/mt/iodefs.py
@@ -148,7 +148,7 @@ class ExportEDI(ContextModule):
 
     def run(self):
         """
-        Run.
+        Entry point into the routine, used to run context menu item.
 
         Returns
         -------
