### 2025-02-18 763bbbf
Bugfix
Updated docstrings
Fixed a bug in mosaic relating to bounds.
Fixed a bug in mosaic relating to float32 nodata.
