+++ pygmi/mt/__init__.py
@@ -1,4 +1,4 @@
 """
-MT Routines
+A set of routines to assist in magnetotelluric (MT) data.
 """
 
