+++ pygmi/mag/__init__.py
@@ -1,4 +1,4 @@
 """
-Magnetic Routines
+A collection of routines performed specifically on magnetic data.
 """
 
