### 2025-02-19 627887b
docstring updates
