+++ pygmi/rsense/change_viewer.py
@@ -43,8 +43,8 @@ class MyMplCanvas(FigureCanvasQTAgg):
 
     Parameters
     ----------
-    parent : parent
-        Reference to the parent routine.
+    parent : parent, optional
+        Reference to the parent routine. The default is None.
     width : float
         Width of the plot.
     height : float
@@ -179,7 +179,15 @@ class MyMplCanvas(FigureCanvasQTAgg):
 
 
 class SceneViewer(BasicModule):
-    """Scene Viewer."""
+    """
+    Scene viewer GUI.
+
+    Parameters
+    ----------
+    parent : parent, optional
+        Reference to the parent routine. The default is None.
+
+    """
 
     def __init__(self, parent=None):
         super().__init__(parent)
