+++ pygmi/clust/crisp_clust.py
@@ -43,8 +43,9 @@ class CrispClust(BasicModule):
 
     Parameters
     ----------
-    parent : parent
-        Reference to the parent routine.
+    parent : parent, optional
+        Reference to the parent routine. The default is None.
+
     """
 
     def __init__(self, parent=None):
