+++ pygmi/rsense/menu.py
@@ -43,8 +43,9 @@ class MenuWidget():
 
     Parameters
     ----------
-    parent : pygmi.main.MainWidget
-        Reference to MainWidget class found in main.py
+    parent : pygmi.main.MainWidget, optional
+        Reference to MainWidget class found in main.py. Default is None.
+
     """
 
     def __init__(self, parent=None):
