+++ pygmi/em/__init__.py
@@ -1,4 +1,4 @@
 """
-EM Routines
+Electromagnetic (EM) routines, including  some inversion capability.
 """
 
