+++ pygmi/raster/cliptozoom.py
@@ -48,8 +48,9 @@ class MyMplCanvas(FigureCanvasQTAgg):
 
     Parameters
     ----------
-    parent : parent
-        Reference to the parent routine.
+    parent : parent, optional
+        Reference to the parent routine. The default is None.
+
     """
 
     def __init__(self, parent=None):
@@ -111,7 +112,15 @@ class MyMplCanvas(FigureCanvasQTAgg):
 
 
 class ClipToZoom(BasicModule):
-    """Clip to zoom Class."""
+    """
+    Clip to zoom GUI Class.
+
+    Parameters
+    ----------
+    parent : parent, optional
+        Reference to the parent routine. The default is None.
+
+    """
 
     def __init__(self, parent=None):
         super().__init__(parent)
