### 2025-02-25 #7019 @boisgera
Hi @AnkitBarik, @jmtemmos,

Do you identify some "low-hanging fruits" in your respective checklists? Some blockers?

Anything I can do to help move the process forward?

Best regards,

Sébastien
