### 2025-02-26 9d6d031
Bugfixes
fixed a bug where reprojecting crashed for data without nodata.
Fixed a bug where a missing profile picture on a model cashed a crash.
