+++ pygmi/seis/graphs.py
@@ -31,7 +31,7 @@ menu.
 
 import os
 import numpy as np
-from PyQt5 import QtWidgets, QtCore
+from PyQt5 import QtWidgets, QtCore, QtGui
 import geopandas as gpd
 from shapely.geometry import Polygon
 from scipy.spatial.distance import cdist
@@ -43,7 +43,6 @@ from matplotlib.figure import Figure
 from matplotlib.backends.backend_qt5 import NavigationToolbar2QT
 from matplotlib.patches import Ellipse
 from shapelysmooth import catmull_rom_smooth
-from sklearn.linear_model import HuberRegressor
 
 from pygmi.vector.dataprep import gridxyz
 from pygmi.misc import ContextModule
@@ -670,6 +669,34 @@ class MyMplCanvas(FigureCanvasQTAgg):
 
         self.isolines = gdf
 
+    def update_tempb(self, btot, datetot):
+        """
+        Update temporal b value plot.
+
+        Parameters
+        ----------
+        btot : numpy array
+            Array of b values.
+        datetot : list
+            list of dates.
+
+        Returns
+        -------
+        None.
+
+        """
+        self.figure.clear()
+        self.axes = self.figure.add_subplot(111)
+
+        self.axes.plot(datetot, btot, 'g',
+                       label='b-value (maximum likelihood)')
+        self.axes.legend()
+        self.axes.set_xlabel('Date', fontsize=8)
+        self.axes.set_ylabel('b-value', fontsize=8)
+        self.axes.tick_params(axis='x', rotation=90)
+
+        self.figure.canvas.draw()
+
 
 class GraphWindow(ContextModule):
     """
@@ -745,12 +772,6 @@ class PlotQC(GraphWindow):
     parent : parent, optional
         Reference to the parent routine. The default is None.
 
-    Attributes
-    ----------
-    lbl_2 : QLabel
-        reference to GraphWindow's lbl_2
-    cmb_2 : QComboBox
-        reference to GraphWindow's cmb_2
     """
 
     def __init__(self, parent=None):
@@ -926,12 +947,6 @@ class PlotIso(GraphWindow):
     parent : parent, optional
         Reference to the parent routine. The default is None.
 
-    Attributes
-    ----------
-    lbl_2 : QLabel
-        reference to GraphWindow's lbl_2
-    cmb_2 : QComboBox
-        reference to GraphWindow's cmb_2
     """
 
     def __init__(self, parent=None):
@@ -977,6 +992,151 @@ class PlotIso(GraphWindow):
         self.lbl_1.setText('Product:')
         self.cmb_1.setCurrentIndex(0)
         self.change_band()
+
+    def save_shp(self):
+        """
+        Save shapefile.
+
+        Returns
+        -------
+        bool
+            True if successful, False otherwise.
+
+        """
+        ext = 'Shape file (*.shp)'
+
+        filename, _ = QtWidgets.QFileDialog.getSaveFileName(
+            self.parent, 'Save Shape File', '.', ext)
+        if filename == '':
+            return False
+        os.chdir(os.path.dirname(filename))
+
+        ifile = str(filename)
+
+        if os.path.isfile(ifile):
+            tmp = ifile[:-4]
+            os.remove(tmp+'.shp')
+            os.remove(tmp+'.shx')
+            os.remove(tmp+'.prj')
+            os.remove(tmp+'.dbf')
+
+        gdf = self.mmc.isolines
+        gdf = gdf.set_crs(4326)
+
+        gdf.to_file(filename)
+
+        return True
+
+
[…]
