+++ pygmi/raster/anaglyph.py
@@ -557,140 +557,6 @@ def sunshade(data, *, azim=-np.pi/4., elev=np.pi/4., alpha=1, cell=100,
     return colormap
 
 
-# def norm2(dat, datmin=None, datmax=None):
-#     """
-#     Normalise vector.
-
-#     Parameters
-#     ----------
-#     dat : numpy array
-#         Vector to be normalised.
-#     datmin : float, optional
-#         Minimum dat value. The default is None.
-#     datmax : float, optional
-#         Maximum dat value. The default is None.
-
-#     Returns
-#     -------
-#     numpy array
-#         Normalised output data.
-
-#     """
-#     if datmin is None:
-#         datmin = np.min(dat)
-#     if datmax is None:
-#         datmax = np.max(dat)
-#     return (dat-datmin)/(datmax-datmin)
-
-
-# def currentshader(data, cell, theta, phi, alpha):
-#     """
-#     Blinn shader.
-
-#     Parameters
-#     ----------
-#     data : numpy array
-#         input MxN data to be imaged.
-#     cell : float
-#         between 1 and 100 - controls sunshade detail.
-#     theta : float
-#         sun elevation (also called g in code below).
-#     phi : float
-#         azimuth.
-#     alpha : float
-#         how much incident light is reflected.
-
-#     Returns
-#     -------
-#     R : numpy array
-#         Output data.
-
-#     """
-#     cdy = np.array([[1., 2., 1.], [0., 0., 0.], [-1., -2., -1.]])
-#     cdx = np.array([[1., 0., -1.], [2., 0., -2.], [1., 0., -1.]])
-
-#     dzdx = ndimage.convolve(data, cdx)  # Use convolve: matrix filtering
-#     dzdy = ndimage.convolve(data, cdy)  # 'valid' gets reduced array
-
-#     dzdx = dzdx/8.
-#     dzdy = dzdy/8.
-
-#     pinit = dzdx
-#     qinit = dzdy
-
-#     # Update cell
-#     p = pinit/cell
-#     q = qinit/cell
-#     sqrt_1p2q2 = np.sqrt(1+p**2+q**2)
-
-#     # Update angle
-#     cosg2 = np.cos(theta/2)
-#     p0 = -np.cos(phi)*np.tan(theta)
-#     q0 = -np.sin(phi)*np.tan(theta)
-#     sqrttmp = 1+np.sqrt(1+p0**2+q0**2)
-#     p1 = p0 / sqrttmp
-#     q1 = q0 / sqrttmp
-
-#     n = 2.0
-
-#     cosi = ((1+p0*p+q0*q)/(sqrt_1p2q2*np.sqrt(1+p0**2+q0**2)))
-#     coss = ((1+p1*p+q1*q)/(sqrt_1p2q2*np.sqrt(1+p1**2+q1**2)))
-#     Ps = coss**n
-#     R = ((1-alpha)+alpha*Ps)*cosi/cosg2
-#     return R
-
-
-# def histcomp(img, nbr_bins=256, perc=5.):
-#     """
-#     Histogram compaction.
-
-#     Parameters
-#     ----------
-#     img : numpy masked array
-#         Input data.
-#     nbr_bins : int, optional
-#         Number of bins. The default is 256.
-#     perc : float, optional
-#         Percentage to clip. The default is 5.
-
-#     Returns
-#     -------
-#     img2 : numpy array
-#         Output data.
-
-#     """
-#     tmp = img.compressed()
-
-#     imhist, bins = np.histogram(tmp, nbr_bins)
-
[…]
