+++ pygmi/seis/menu.py
@@ -99,6 +99,11 @@ class MenuWidget():
         context_menu['Seis'].addAction(self.action_show_QC_plots)
         self.action_show_QC_plots.triggered.connect(self.show_QC_plots)
 
+        self.action_show_TP_plots = QtWidgets.QAction('Show Temporal b-value '
+                                                      'Plots')
+        context_menu['Seis'].addAction(self.action_show_TP_plots)
+        self.action_show_TP_plots.triggered.connect(self.show_TP_plots)
+
         self.action_export_seisan = QtWidgets.QAction('Export SEISAN Data')
         context_menu['Seis'].addAction(self.action_export_seisan)
         self.action_export_seisan.triggered.connect(self.export_seisan)
@@ -164,3 +169,7 @@ class MenuWidget():
     def show_iso_plots(self):
         """Show QC plots."""
         self.parent.launch_context_item(graphs.PlotIso)
+
+    def show_TP_plots(self):
+        """Show Temporal b-value plots."""
+        self.parent.launch_context_item(graphs.PlotTempB)
