### 2025-02-27 009533c
Update misc.py
fixed a bug causing a crash while merging models
