### 2025-02-27 f38fbdc
Update
Magnetic residual and gravity residual now appears immediately in overview combo of modelling after calculation.
