### 2025-03-06 6a457f4
Update
Added borehole menu
Added Local to Base station label in gravity
Saving EDI file will not give overwrite warning,
Test contours are thinner in 3D modelling.
Ranged copy now references layer and profile view.
Plots now use tight layout and not constrained.
