+++ pygmi/bholes/graphs.py
@@ -1,0 +1,619 @@
+# -----------------------------------------------------------------------------
+# Name:        boreholes/graphs.py (part of PyGMI)
+#
+# Author:      Patrick Cole
+# E-Mail:      pcole@geoscience.org.za
+#
+# Copyright:   (c) 2018 Council for Geoscience
+# Licence:     GPL-3.0
+#
+# This file is part of PyGMI
+#
+# PyGMI is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# PyGMI is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+# -----------------------------------------------------------------------------
+"""
+This module provides a variety of methods to plot borehole data via the context
+menu.
+"""
+
+import xml
+import re
+import textwrap
+import numpy as np
+import pandas as pd
+from PyQt5 import QtWidgets, QtCore
+from matplotlib.path import Path
+import matplotlib.patches as mpatches
+from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
+from matplotlib.figure import Figure
+
+from pygmi.misc import ContextModule
+
+
+class MyMplCanvas(FigureCanvasQTAgg):
+    """
+    Matplotlib canvas widget for the actual plot.
+
+    Parameters
+    ----------
+    parent : parent, optional
+        Reference to the parent routine. The default is None.
+
+    """
+
+    def __init__(self, parent=None):
+        fig = Figure()
+        super().__init__(fig)
+
+    def update_legend(self, data1):
+        """
+        Update the plot legend.
+
+        Parameters
+        ----------
+        data1 : dictionary
+            Dictionary containing the data.
+
+        Returns
+        -------
+        None.
+
+        """
+        fig = self.figure
+        fig.clear()
+        ax = fig.gca()
+
+        idir = __file__.rpartition('\\')[0]
+        logfile = idir+'\\logplot.xlsx'
+
+        pagewidth = 8
+        pageheight = 4
+        dpp = 14  # depth per page
+        wpp = dpp*pagewidth/pageheight
+
+    # #########################################################################
+    # Load in logplot init
+        xl = pd.ExcelFile(logfile)
+        usgs = xl.parse('USGS')
+        cgs = xl.parse('CGS')
+        cgslookup = xl.parse('250K Lookup')
+        colours = xl.parse('Colours')
+        xl.close()
+
+        usgs = usgs.set_index('code').to_dict()['description']
+        cgslookup['COLOR_CODE'] = cgslookup['COLOR_CODE'].astype(str)
+        cgslookup['COLOR_CODE'] = cgslookup['COLOR_CODE'].apply('{0:0>3}'.format)
+        stratcol = cgslookup.set_index('LITHO_NAME').to_dict()['COLOR_CODE']
+        col = colours.set_index('code').to_dict()['colour']
+        clith = cgs.set_index('lithology').to_dict()['lithology description']
+        cgs = cgs.set_index('lithology').to_dict()['code']
+        col['none'] = 'ffffff'
+        col['nan'] = 'ffffff'
[…]
