+++ pygmi/mt/graphs.py
@@ -83,7 +83,7 @@ class MyMplCanvas(FigureCanvasQTAgg):
     """
 
     def __init__(self, parent=None):
-        fig = Figure(layout='constrained')
+        fig = Figure(layout='tight')
         self.axes = fig.add_subplot(111)
         self.line = None
         self.ind = None
