### 2025-03-06 9e2de17
Update iodefs.py
modis 6.0 support removed.
