### 2025-03-06 be4ecc8
Updates
Added feature for analyse spectra
Changed 'Create Batch' to 'Import Batch' in remote sensing menu.
L1C sentinel data can be topographically corrected.
