+++ pygmi/rsense/features.py
@@ -71,8 +71,8 @@ product['white mica'] = ['f2200', 'r2350De > 1.02', 'r2160D2190 < 1.005',
 product['smectite'] = ['f2200', 'r2350De < 1.0', 'r2160D2190 < 1.005',
                        'r2200D > 1.01']
 # product['kaolin-2200'] = ['f2200', 'r2160D2190 > 1.005']
-product['kaolin-2200'] = ['f2200', 'f2160 < f2200p',
-                          'f2200 > 0.01', 'r2160D2190 > 1.005']
+product['kaolin'] = ['f2200', 'f2160 < f2200p',
+                     'f2200 > 0.01', 'r2160D2190 > 1.005']
 # product['kaolin-2160'] = ['f2160', 'r2160D2190 > 1.005']
 product['chlorite-2250'] = ['f2250', 'r2330D > 1.0', 'r2250D > 1.005',
                             'r1550D < 1.01']
