### 2025-03-06 bfdfa4b
Update graphs.py
Changed borehole font size and log size
