### 2025-03-06 Patrick-Cole/pygmi#22 @jmtemmos
Linux Installation documentation
related to JOSS Review https://github.com/openjournals/joss-reviews/issues/7019

Dear @Patrick-Cole,

please find below some remarks that might be useful to clarify software installation

I'm using virtualenv and python3.13, under Manjaro Linux (rolling release)

It could be useful to list the install of the following lib to prevent the warning in the gui :

`pip install numba`
`pip install scikit-learn`  sklearn has been deprecated in pypy :)

Best regards
