### 2025-03-06 Patrick-Cole/pygmi#23 @jmtemmos
Wayland artifacts
Hello @Patrick-Cole

This relates to https://github.com/openjournals/joss-reviews/issues/7019

I'm using manjaro/wayland /plasma/kde6.3 desktop.

When launching pygmi-gui, the following error message is displayed :

`qt.qpa.wayland: Wayland does not support QWindow::requestActivate()`

App is launching ok, but some graphical bugs are present : menu location are shifted from the initial mouse click location.

Best regards.
