### 2025-03-06 Patrick-Cole/pygmi#24 @jmtemmos
Crash when open TIF file.
Hello @Patrick-Cole

This relates to https://github.com/openjournals/joss-reviews/issues/7019

When i import a geotiff file using the GUI, I get the following error displayed along with a crash :

```Traceback (most recent call last):
 File "/home/jmt/dev/pygmi/.venv/lib/python3.13/site-packages/pygmi/raster/menu.py", line 285, in import_data
   self.parent.item_insert('Io', 'Import Raster Data', iodefs.ImportData)
   ~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 File "/home/jmt/dev/pygmi/.venv/lib/python3.13/site-packages/pygmi/main.py", line 926, in item_insert
   item.np_poly *= 1.5*text_width/item.np_poly[:, 0].ptp()
                                  ^^^^^^^^^^^^^^^^^^^^^^
AttributeError: `ptp` was removed from the ndarray class in NumPy 2.0. Use np.ptp(arr, ...) instead.
[1]    201171 IOT instruction (core dumped)  pygmi-gui
```

Best regards
Jean-Marc
