### 2025-03-06 #7019 @jmtemmos
Hi @Patrick-Cole. I've just added 3 issues.
I hope this will help.
Third one is the most important since it prevents me to open any Geotiff files (I've tried several).

* [Linux Installation documentation Patrick-Cole/pygmi#22](https://github.com/Patrick-Cole/pygmi/issues/22)
* [Wayland artifacts Patrick-Cole/pygmi#23](https://github.com/Patrick-Cole/pygmi/issues/23)
* [Crash when open TIF file. Patrick-Cole/pygmi#24](https://github.com/Patrick-Cole/pygmi/issues/24)

Best regards,
Jean-Marc
