### 2025-03-07 Patrick-Cole/pygmi#22 @Patrick-Cole
Hi Jean-Marc,

I thought I was being smart by excluding those packages from the pygmi requirements,  since I thought some of my other packages installed them. I will re add them to my next PyGMI commit to pypi so hopefully that will fix the issue.

Regards,

Patrick
