+++ pyproject.toml
@@ -39,6 +39,7 @@ dependencies = [
                 'matplotlib>=3.9.0',
                 'mtpy>=1.1.5',
                 'natsort>=8.4.0',
+                'numba>=0.61.0',
                 'numexpr>=2.10.1',
                 'openpyxl>=3.1.2',
                 'psutil>=6.0.0',
@@ -47,6 +48,7 @@ dependencies = [
                 'pytest>=8.2.2',
                 'rasterio>=1.3.9',
                 'rioxarray>=0.15.6',
+                'scikit-learn>=1.6.0',
                 'scikit-image>=0.24.0',
                 'shapelysmooth>=0.2.0',
                 'simpeg>=0.21.1',
