### 2025-03-07 f5b5f8f
Bugfixes
Fixed a bug causing strange colours on change viewer.
Fixed bugs causing matplotlib save image to not work.
Importing satellite data in batch mode will select bands starting with B
