### 2025-03-07 Patrick-Cole/pygmi#24 @Patrick-Cole
Hi,

I have posted a new version on pypi. It should fix this problem. I am revamping the help, so for now that will not work. It will be fixed by next week though.

Regards,

Patrick
