+++ pygmi/html/_sources/mt.dm.removesshift.rst.txt
@@ -1,18 +1,0 @@
-Remove Static Shift
--------------------
-This module allows for the static shift correction to EDI data. Apparent resistivity and phase versus period are both displayed as graphs.
-
-The interface for this module has the following aspects and options:
-
-1. Windows showing the apparent resistivity and phase versus period for the selected station.
-2. **Apply to all stations** – Select this option to apply the rotation to all the stations that were loaded. If it is not selected it will only apply to the currently selected station.
-3. **Station Name** – If data for multiple stations were imported the dropdown list can be used to select a specific station.
-4. **Graph Type** – This can be either ‘**xy, yx**’ or ‘**xx, yy**’.
-5. **Shift X** – Correction to X.
-6. **Shift Y** – Correction to Y. 
-7. **Remove Static Shift** – Apply the correction to data for the current station or for all the imported stations.
-8. **Reset data** - Restores data for pre-shifted status.
-9. Standard image display settings that allow the user to zoom into specific areas of the image, move the zoomed in area around, return to the full image, save the image, etc.
-
-.. image:: _images/mtstatic.png
-
