+++ pygmi/helpdocs/html/_sources/autoapi/pygmi/bholes/index.rst
@@ -1,0 +1,23 @@
+pygmi.bholes
+============
+
+.. py:module:: pygmi.bholes
+
+.. autoapi-nested-parse::
+
+   These routines cover the import and display of Council for Geoscience borehole
+   data. This data can also be used in modelling in other modules.
+
+
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 1
+
+   /autoapi/pygmi/bholes/graphs/index
+   /autoapi/pygmi/bholes/iodefs/index
+   /autoapi/pygmi/bholes/menu/index
+
+
