+++ pygmi/helpdocs/html/_sources/autoapi/pygmi/vector/index.rst
@@ -1,0 +1,26 @@
+pygmi.vector
+============
+
+.. py:module:: pygmi.vector
+
+.. autoapi-nested-parse::
+
+   General routines related to vector data.
+
+
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 1
+
+   /autoapi/pygmi/vector/dataprep/index
+   /autoapi/pygmi/vector/graphs/index
+   /autoapi/pygmi/vector/iodefs/index
+   /autoapi/pygmi/vector/menu/index
+   /autoapi/pygmi/vector/minc/index
+   /autoapi/pygmi/vector/show_table/index
+   /autoapi/pygmi/vector/structure/index
+
+

[…]
