### 2025-03-14 Patrick-Cole/pygmi#23 @Patrick-Cole
Hi,

I do know that Wayland does not always work well with Qt. I will see if I can do anything from my side, otherwise you may need to test using X11. I will keep you posted.

Regards

Patrick
