+++ pygmi/helpdocs/html/_sources/bholes.cm.showlog.rst
@@ -1,0 +1,13 @@
+Show Borehole Log
+-----------------
+This function displays the borehole log. It has the following aspects:
+
+1. The locality and core number.
+2. An image of the lithology of the log.
+3. Information about the borehole.
+4. Legend for the lithologies in the boreholes.
+5. **BoreholeID** – The user can select the borehole to display using a dropdown list.
+
+.. image:: _images/bholedisplay.png
+
+Show Borehole Log interface.
