+++ pygmi/helpdocs/html/_sources/pfmod.dm.mc.editor.rst
@@ -5,16 +5,15 @@ The Potential Field Modelling interface appears once the Model Extent Parameters
 
 .. toctree::
     :titlesonly:
-    :numbered:
-    
-    pfmod.dm.mc.toolbar
-    pfmod.dm.mc.lpviews
-    pfmod.dm.mc.lpcontrols
-    pfmod.dm.mc.modeltools
-    pfmod.dm.mc.profexport
-    pfmod.dm.mc.profview
-    pfmod.dm.mc.calcinfo
-    pfmod.dm.mc.imgset
+
+    1. Top Toolbar<pfmod.dm.mc.toolbar>
+    2. Layer and profile views<pfmod.dm.mc.lpviews>
+    3. Profile and layer control<pfmod.dm.mc.lpcontrols>
+    4. Modelling tools<pfmod.dm.mc.modeltools>
+    5. Profile export<pfmod.dm.mc.profexport>
+    6. Profile viewing options<pfmod.dm.mc.profview>
+    7. Calculation progess<pfmod.dm.mc.calcinfo>
+    8. Standard image display settings<pfmod.dm.mc.imgset>
 
 .. figure:: _images/pfmodedit.png
    
