+++ pygmi/helpdocs/html/_sources/cluster.dm.fuzzy.rst
@@ -1,6 +1,6 @@
 Fuzzy Clustering
 ----------------
-In crisp clustering a sample can only belong to one cluster whereas in fuzzy clustering a sample is assigned to all clusters but it has probability assigned to it of belonging to a certain cluster (Paasche and Eberle, 2009, 2011). This is referred to as the membership of a sample to a cluster. Higher membership means the sample is closer to the centre of the cluster (Paasche and Eberle, 2009). If a sample falls in between two classes, it may have a 50% probability that it belongs to either of the two classes. There are different ways of defining the classes that are similar to crisp clustering, e.g. **fuzzy C-means**, **advanced fuzzy C-means** and **Gustafson Kessel**. The figure below illustrates the relation between the crisp clustering algorithms and the fuzzy clustering algorithms. 
+In crisp clustering a sample can only belong to one cluster whereas in fuzzy clustering a sample is assigned to all clusters but it has probability assigned to it of belonging to a certain cluster (Paasche and Eberle, 2009, 2011). This is referred to as the membership of a sample to a cluster. Higher membership means the sample is closer to the centre of the cluster (Paasche and Eberle, 2009). If a sample falls in between two classes, it may have a 50% probability that it belongs to either of the two classes. There are different ways of defining the classes that are similar to crisp clustering, e.g. **fuzzy C-means**, **advanced fuzzy C-means** and **Gustafson Kessel**. The figure below illustrates the relation between the :doc:`crisp clustering<cluster.dm.crisp>` algorithms and the fuzzy clustering algorithms. 
 
 .. figure:: _images/clustfuzzy1.png
 
