+++ pygmi/helpdocs/html/_sources/rsense.dm.change.rst
@@ -2,42 +2,52 @@ Change Detection
 ----------------
 Change detection is an important technique used in the analysis of remote sensing data. Ordinarily, satellite data has a fixed number of bands (between 4 and 12 on average). These bands represent a light spectrum and can be utilised to study various terrestrial phenomena, from vegetation to minerals. The limits of the data are normally due to the number of bands in the data. One way to overcome this is to examine the temporal dimensions of the data. By studying the change in the data across multiple scenes, we can leverage more information.
 
-Change detection relies on the changing nature of the earth. This can be from extreme environmental disasters such as fires and landslides, or as simple as studying the effect a seasonal change has on the target being studied.
+Change detection relies on the changing nature of the earth. This can be from extreme environmental disasters such as fires and landslides, or as simple as studying the effect a seasonal change has on the target being studied, land use and land cover change and mineral outcrops by using change detection as a process of elimination (Cole and Coetzee, 2022).
 
-This includes:
-1) Landslides.
-2) Environmental studies.
-3) Land use and land cover.
-4) Mineral outcrop (by using change detection as a process of elimination).
+The **Change Detection** menu has two functions, namely **Calculate Change Indices** and **View Change Data**.
+
+.. figure:: _images/changefuncs.png
+
+   Change Detection functions.
+
 
 Calculate Change Indices
 ^^^^^^^^^^^^^^^^^^^^^^^^
-Change Detection Indices give a measure of change in an area. The input to this module is a list of datasets loaded using the batch importer. Each dataset must also have an associated date. The import tool for satellite scenes should assign a date automatically, but should a manual date need to be assigned, this can be done within the metadata context menu. A variety of change indices can be calculated:
+**Change Detection Indices** give a measure of change in an area. The input to this module is a list of datasets loaded using the batch importer. It does not have to be satellite images, but can be derived products such as classifications. Each dataset must have an associated date. The import tool for satellite scenes should assign a date automatically, but should a manual date need to be assigned, this can be done within the :doc:`metadata<raster.cm.meta>` context menu. 
 
-Indices:
+The following indices can be calculated:
 
-* Difference - calculates the difference between two dates.
-* Mean - average pixel value image over all dates.
-* Standard Deviation - shows pixel variation though standard deviation.
-* Coefficient of Variation - measure of variation, taking mean into account.
-* Spectral Angle Mapper - find spectral angle between two scenes.
+* **Difference** – Calculates the difference between two dates (latter date minus earlier date).
+* **Mean** – Average pixel value image over all dates.
+* **Standard Deviation** – Shows pixel variation through standard deviation.
+* **Coefficient of Variation** – Measure of variation, taking mean into account.
+* **Spectral Angle Mapper** – Find spectral angle between two scenes.
+
+Note that for **Difference** and **Spectral Angle Mapper** calculations only two datasets are allowed.
 
 .. figure:: _images/changeindices.png
 
+   Change indices that can be calculated.
+
+The result is a multiband raster file consisting of all the indices calculated for all bands. It can be exported by right-clicking on the **Calculate Change Indices** module and selecting :doc:`Export Raster Data<raster.cm.export>`.
+
 View Change Data
 ^^^^^^^^^^^^^^^^
[…]
