+++ pygmi/helpdocs/html/_sources/mag.rst
@@ -3,7 +3,7 @@ Magnetics
 
 Description of Modules
 ----------------------
-The Magnetics Menu is a collection of routines performed specifically on magnetic data. The user first imports the magnetic dataset and then connects it with the required function.
+The Magnetics Menu is a collection of routines performed specifically on magnetic data. The user first :doc:`imports<raster.dm.importrasterdata>` the magnetic dataset and then connects it with the required function.
 
 .. toctree::
     :titlesonly:
