+++ pygmi/helpdocs/html/_sources/seis.cm.exportseisan.rst
@@ -1,3 +1,3 @@
 Export SEISAN Data
 ------------------
-This option allows for the export of seismology data in NORDIC format.
+This option allows for the export of seismology data in **NORDIC** or **NORDIC2** format (**OUT** file).
