### 2025-03-19 a734e97
Update
Updated help files
Fixed a bug where ASTER files were duplicated in batch mode.
Fixed a bug in equation editor where a leading space caused errors
