+++ pygmi/rsense/iodefs.py
@@ -186,7 +186,7 @@ class ImportData(BasicModule):
         self.le_sfile.setText('')
         self.lbl_ftype.setText('File Type:')
 
-        ext = 'Common formats (*.zip *.hdf *.tar *.tar.gz *.xml *.h5 *.nc);;'
+        ext = 'Common formats (*.zip *.tar *.tar.gz *.xml *.h5 *.nc);;'
 
         self.ifile, _ = QtWidgets.QFileDialog.getOpenFileName(
             self.parent, 'Open File', '.', ext)
@@ -377,7 +377,7 @@ class ImportBatch(BasicModule):
 
         self.le_sfile.setText(self.idir)
 
-        types = ['*.tif', '*.hdr', '*.hdf', '*.zip', '*.tar', '*.tar.gz',
+        types = ['*.tif', '*.hdr', '*.zip', '*.tar', '*.tar.gz',
                  '*[!aux].xml', '*.h5', '*.SAFE/MTD_MSIL*.xml']
 
         allfiles = []
@@ -399,7 +399,6 @@ class ImportBatch(BasicModule):
             if tmp:
                 self.tnames[sensor] = tmp
 
-        # breakpoint()
         self.cmb_sensor.disconnect()
         self.cmb_sensor.clear()
         self.cmb_sensor.addItems(self.bands.keys())
@@ -1246,7 +1245,7 @@ def consolidate_aster_list(flist):
         List of filenames.
 
     """
-    asterhfiles = []
+    # asterhfiles = []
     asterzfiles = []
     otherfiles = []
 
@@ -1254,8 +1253,8 @@ def consolidate_aster_list(flist):
         bfile = os.path.basename(ifile)
         ext = os.path.splitext(ifile)[1].lower()
 
-        if 'AST_' in bfile and ext == '.hdf':
-            asterhfiles.append(ifile)
+        # if 'AST_' in bfile and ext == '.hdf':
+        #     asterhfiles.append(ifile)
         if 'AST_' in bfile and ext == '.zip':
             asterzfiles.append(ifile)
         else:
@@ -1269,11 +1268,11 @@ def consolidate_aster_list(flist):
         tmp[adate] = ifile
     asterfiles += list(tmp.values())
 
-    for ifile in asterhfiles:
-        adate = os.path.basename(ifile).split('_')[2]
-        tmp[adate] = ifile
+    # for ifile in asterhfiles:
+    #     adate = os.path.basename(ifile).split('_')[2]
+    #     tmp[adate] = ifile
 
-    asterfiles += list(tmp.values())
+    # asterfiles += list(tmp.values())
 
     flist = asterfiles+otherfiles
 
@@ -1526,18 +1525,18 @@ def get_data(ifile, *, piter=None, showlog=print, tnames=None, metaonly=False,
     if '.xml' in bfile.lower():
         dtree = etree_to_dict(ET.parse(ifile).getroot())
 
-    if 'AST_' in bfile and ext == '.hdf':
-        idir = os.path.dirname(ifile)
-        adate = os.path.basename(ifile).split('_')[2]
-        ifiles = glob.glob(os.path.join(idir, '*'+adate+'*.hdf'))
-        dat = []
-        for afile in ifiles:
-            tmp = get_aster_hdf(afile, piter, showlog, tnames, metaonly)
-            if tmp is not None:
-                dat += tmp
-            if dat == []:
-                dat = None
-    elif 'AST_' in bfile and ext == '.zip':
+    # if 'AST_' in bfile and ext == '.hdf':
+    #     idir = os.path.dirname(ifile)
+    #     adate = os.path.basename(ifile).split('_')[2]
+    #     ifiles = glob.glob(os.path.join(idir, '*'+adate+'*.hdf'))
+    #     dat = []
+    #     for afile in ifiles:
[…]
