+++ pygmi/helpdocs/html/_sources/faq.rst
@@ -1,0 +1,8 @@
+FAQs and Troubleshooting
+========================
+This section contains common mistakes which users make, and ways to avoid them.
+
+.. toctree::
+    :titlesonly:
+    
+    faq.importxyz
