### 2025-03-21 #7019 @boisgera
@jmtemmos, I see that you have completed [your checklist](https://github.com/openjournals/joss-reviews/issues/7019#issuecomment-2650962728), thanks! 🙏

@AnkitBarik, do you think that you could complete [your own checklist](https://github.com/openjournals/joss-reviews/issues/7019#issuecomment-2243292342) in the next weeks? (or at least to identify blockers if there are any?)
