+++ pygmi/pfmod/tab_prof.py
@@ -27,12 +27,12 @@
 import copy
 import os
 import random
-from PyQt5 import QtWidgets, QtCore, QtGui
+from PyQt6 import QtWidgets, QtCore, QtGui
 import numpy as np
 from scipy import ndimage
 from scipy import interpolate
-from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
-from matplotlib.backends.backend_qt5 import NavigationToolbar2QT
+from matplotlib.backends.backend_qtagg import FigureCanvasQTAgg
+from matplotlib.backends.backend_qt import NavigationToolbar2QT
 from matplotlib.figure import Figure
 from matplotlib import colormaps
 import pandas as pd
@@ -119,10 +119,10 @@ class ProfileDisplay(QtWidgets.QWidget):
         None.
 
         """
-        sizepolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred,
-                                           QtWidgets.QSizePolicy.Fixed)
-        sizepolicy2 = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Ignored,
-                                            QtWidgets.QSizePolicy.Fixed)
+        sizepolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Policy.Preferred,
+                                           QtWidgets.QSizePolicy.Policy.Fixed)
+        sizepolicy2 = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Policy.Ignored,
+                                            QtWidgets.QSizePolicy.Policy.Fixed)
 
         self.lw_prof_defs.setFixedWidth(220)
 
@@ -148,23 +148,23 @@ class ProfileDisplay(QtWidgets.QWidget):
 
         self.hs_overview.setMaximum(100)
         self.hs_overview.setProperty('value', 0)
-        self.hs_overview.setOrientation(QtCore.Qt.Horizontal)
+        self.hs_overview.setOrientation(QtCore.Qt.Orientation.Horizontal)
 
         self.hs_sideview.setMaximum(100)
         self.hs_sideview.setProperty('value', 0)
-        self.hs_sideview.setOrientation(QtCore.Qt.Horizontal)
+        self.hs_sideview.setOrientation(QtCore.Qt.Orientation.Horizontal)
 
         self.hs_layer.setSizePolicy(sizepolicy)
-        self.hs_layer.setOrientation(QtCore.Qt.Horizontal)
+        self.hs_layer.setOrientation(QtCore.Qt.Orientation.Horizontal)
 
         self.hs_profnum.setSizePolicy(sizepolicy)
-        self.hs_profnum.setOrientation(QtCore.Qt.Horizontal)
+        self.hs_profnum.setOrientation(QtCore.Qt.Orientation.Horizontal)
 
         self.sb_profnum.setPrefix('Profile: ')
         self.sb_profnum.setWrapping(True)
 
         self.hs_cprofnum.setSizePolicy(sizepolicy)
-        self.hs_cprofnum.setOrientation(QtCore.Qt.Horizontal)
+        self.hs_cprofnum.setOrientation(QtCore.Qt.Orientation.Horizontal)
         self.hs_cprofnum.setHidden(True)
 
         self.sb_cprofnum.setPrefix('Custom: ')
@@ -1372,7 +1372,7 @@ class ProfileDisplay(QtWidgets.QWidget):
         if curtext is None:
             curtext = self.cmb_overview.currentText()
         cindex = self.cmb_overview.findText(curtext,
-                                            QtCore.Qt.MatchFixedString)
[…]
