+++ pygmi/pfmod/show_table.py
@@ -24,8 +24,8 @@
 # -----------------------------------------------------------------------------
 """Routine which displays a table graphically with various stats."""
 
-from PyQt5 import QtWidgets
-from PyQt5.QtCore import Qt
+from PyQt6 import QtWidgets
+from PyQt6.QtCore import Qt
 import numpy as np
 
 from pygmi.misc import ContextModule
@@ -143,7 +143,7 @@ class BasicStats3D(ContextModule):
             for col in range(data.shape[1]):
                 txt = f' {data[row, col]:,.5f}'
                 txt = QtWidgets.QLabel(txt)
-                txt.setAlignment(Qt.AlignRight)
+                txt.setAlignment(Qt.AlignmentFlag.AlignRight)
 
                 self.tablewidget.setCellWidget(row, col, txt)
 

[…]
