+++ _sources/pfmod.dm.mc.editor.rst
@@ -1,0 +1,20 @@
+Editor overview
+^^^^^^^^^^^^^^^
+
+The Potential Field Modelling interface appears once the Model Extent Parameters have been applied. The features of this interface will be discussed under the following headings:
+
+.. toctree::
+    :titlesonly:
+
+    1. Top Toolbar<pfmod.dm.mc.toolbar>
+    2. Layer and profile views<pfmod.dm.mc.lpviews>
+    3. Profile and layer control<pfmod.dm.mc.lpcontrols>
+    4. Modelling tools<pfmod.dm.mc.modeltools>
+    5. Profile export<pfmod.dm.mc.profexport>
+    6. Profile viewing options<pfmod.dm.mc.profview>
+    7. Calculation progess<pfmod.dm.mc.calcinfo>
+    8. Standard image display settings<pfmod.dm.mc.imgset>
+
+.. figure:: _images/pfmodedit.png
+   
+   Overview of the Potential Field Modelling Editor.
