+++ _sources/vector.cm.showmapprof.rst
@@ -1,0 +1,15 @@
+Show Profiles on a Map
+----------------------
+This module displays line data in in map form. The line data is displayed in black, superimposed over the x,y location of the data (gray line). This is only applicable to point data collected along lines.
+
+The interface has the following features:
+
+1. **Show Line Labels** – Select to show the line number at the end of a line.
+2. **Column** – The channel to display.
+3. **Scale** – This allows the user to change the anomaly scale.
+4. Standard image display settings that allow the user to zoom into specific areas of the image, move the zoomed in area around, return to the full image, save the image, etc.
+
+.. figure:: _images/vectorlinedatamap.png
+
+   Show Profiles on Map interface. A zoomed in section shows the selected channel in black superimposed on the positional data in blue.
+
