+++ _sources/seis.ref.rst.txt
@@ -1,8 +1,0 @@
-References
-----------
-
-  Wiemer, S. & Baer, M. (2000) Mapping and removing quarry blast events from seismicity catalogues. Bulletin of the Seismological Society of America. (April), 525–530. [online]. Available from: http://www.bssaonline.org/content/90/2/525.short (Accessed 8 November 2013).
- 
-  http://www.ceri.memphis.edu/people/olboyd/Software/Software.html
-  
-  SEISAN Earthquake Analysis Software, version 11, July 2018, 554 pages, Accessed on 13 Sep. 2018, [Online]. Available: http://seisan.info/
