### 2025-03-27 ad3df84
Bugfixes
Fixed a bug when exporting amplitudes in Nordic2.
Fixed Qt6 Enum.
