+++ requirements.txt
@@ -10,7 +10,7 @@ numexpr>=2.10.1
 openpyxl>=3.1.2
 psutil>=6.0.0
 pyopengl>=3.1.7
-pyqt5>=5.15.10
+pyqt6>=6.8.1
 pytest>=8.2.2
 rasterio>=1.3.9
 rioxarray>=0.15.6
