### 2025-03-27 3547183
Bugfix
PyQt6 enum
