+++ pygmi/grav/dataprep.py
@@ -330,14 +330,14 @@ class ProcessData(BasicModule):
         if self.le_knownstat.text() == 'None':
             txt = ('Invalid base station number.')
             QtWidgets.QMessageBox.warning(self.parent, 'Error',
-                                          txt, QtWidgets.QMessageBox.Ok)
+                                          txt, QtWidgets.QMessageBox.StandardButton.Ok)
             return
 
         kstat = float(self.le_knownstat.text())
         if kstat not in pdat['STATION'].values:
             txt = ('Invalid base station number.')
             QtWidgets.QMessageBox.warning(self.parent, 'Error',
-                                          txt, QtWidgets.QMessageBox.Ok)
+                                          txt, QtWidgets.QMessageBox.StandardButton.Ok)
             return
 
         # Drift Correction, to abs base value
@@ -361,7 +361,7 @@ class ProcessData(BasicModule):
             txt = ('Your known base values need to be before and after at '
                    'least one local base station value.')
             QtWidgets.QMessageBox.warning(self.parent, 'Error',
-                                          txt, QtWidgets.QMessageBox.Ok)
+                                          txt, QtWidgets.QMessageBox.StandardButton.Ok)
             return
 
         absbase = grv-np.interp(x, xp, fp) + float(self.le_knownbase.text())
