### 2025-04-03 cfb6e13
Updates
Added helpfiles
Fixed Qt6 glitches
