+++ pygmi/vector/show_table.py
@@ -29,6 +29,7 @@ from PyQt6.QtCore import Qt
 import pandas as pd
 
 from pygmi.misc import ContextModule
+from pygmi import menu_default
 
 
 class BasicStats(ContextModule):
@@ -65,8 +66,10 @@ class BasicStats(ContextModule):
         """
         hbl = QtWidgets.QHBoxLayout(self)
         vbl = QtWidgets.QVBoxLayout()
+        helpdocs = menu_default.HelpButton('vector.cm.stats')
 
         vbl.addWidget(self.pushbutton_save)
+        vbl.addWidget(helpdocs)
         hbl.addWidget(self.tablewidget)
         hbl.addLayout(vbl)
 
