+++ pygmi/raster/equation_editor.py
@@ -30,6 +30,7 @@ import numexpr as ne
 
 from pygmi.misc import BasicModule
 from pygmi.raster.misc import lstack
+from pygmi import menu_default
 
 
 class EquationEditor(BasicModule):
@@ -83,6 +84,7 @@ class EquationEditor(BasicModule):
         lbl_3 = QtWidgets.QLabel('Output Data Type:')
         self.cmb_dtype.addItems(['auto', 'uint8', 'int16', 'int32',
                                  'float32', 'float64'])
+        helpdocs = menu_default.HelpButton('raster.dm.equationeditor')
 
         self.textbrowser.setEnabled(True)
         self.resize(600, 480)
@@ -151,7 +153,8 @@ class EquationEditor(BasicModule):
         gl_1.addWidget(self.cmb_dtype, 6, 0, 1, 1)
         gl_1.addWidget(lbl_3, 5, 0, 1, 1)
         gl_1.addWidget(self.textbrowser2, 7, 0, 1, 2)
-        gl_1.addWidget(buttonbox, 8, 0, 1, 2)
+        gl_1.addWidget(helpdocs, 8, 0, 1, 1)
+        gl_1.addWidget(buttonbox, 8, 1, 1, 1)
 
         self.cmb_1.currentIndexChanged.connect(self.combo)
         buttonbox.accepted.connect(self.accept)
