+++ pygmi/clust/graphs.py
@@ -108,6 +108,68 @@ class MyMplCanvas(FigureCanvasQTAgg):
         self.axes.yaxis.set_major_formatter(frm)
         self.figure.canvas.draw()
 
+    def update_bars(self, data1, rdata):
+        """
+        Update the class plot.
+
+        Parameters
+        ----------
+        data1 : pygmi.raster.datatypes.Data
+            Input raster dataset containing classes.
+        rdata : pygmi.raster.datatypes.Data
+            Input raster dataset containing data.
+
+        Returns
+        -------
+        None.
+
+        """
+        cnr = data1.metadata['Cluster']['no_clusters']
+        x = range(cnr)
+
+        if 'labels' in data1.metadata['Cluster']:
+            lbls = data1.metadata['Cluster']['labels']
+        else:
+            lbls = [f'{i+1}' for i in x]
+
+        self.figure.clear()
+        self.axes = self.figure.add_subplot(111, label='map')
+
+        cdata = {}
+        cdatab = {}
+        for i in x:
+            cdata[lbls[i]] = []
+            cdatab[lbls[i]] = []
+
+        for rdat in rdata:
+            for i in x:
+                cmin = rdat.data[data1.data == (i+1)].min()
+                cmax = rdat.data[data1.data == (i+1)].max()
+                cdata[lbls[i]].append(cmax-cmin)
+                cdatab[lbls[i]].append(cmin)
+
+        dataids = [i.dataid for i in rdata]
+
+        x = np.arange(len(dataids))
+        width = .8/cnr
+        multiplier = 0
+
+        for attribute, measurement in cdata.items():
+            bottom = cdatab[attribute]
+            offset = width * multiplier
+            rects = self.axes.bar(x + offset, measurement, width, bottom,
+                                  label=attribute)
+            # ax.bar_label(rects, padding=5)
+            multiplier += 1
+
+        # Add some text for labels, title and custom x-axis tick labels, etc.
+        self.axes.set_ylabel('Value')
+        self.axes.set_title(f'Dataset ranges for {cnr} classes')
+        self.axes.set_xticks(x+.4, dataids)
+        self.axes.legend(loc='upper left')
+
+        self.figure.canvas.draw()
+
     def update_scatter(self, x, y):
         """
         Update the scatter plot.
@@ -272,6 +334,79 @@ class PlotRaster(ContextModule):
 
         self.show()
         self.change_band()
+
+
+class PlotBars(ContextModule):
+    """
+    Plot Bar Class GUI.
+
+    Parameters
+    ----------
+    parent : parent, optional
+        Reference to the parent routine. The default is None.
+
+    """
+
+    def __init__(self, parent=None):
+        super().__init__(parent)
+
+        self.setAttribute(QtCore.Qt.WidgetAttribute.WA_DeleteOnClose)
+        self.setWindowTitle('Class Dataset Ranges')
+
+        vbl = QtWidgets.QVBoxLayout(self)
+        hbl = QtWidgets.QHBoxLayout()
+        self.mmc = MyMplCanvas(self)
+        mpl_toolbar = NavigationToolbar2QT(self.mmc, self.parent)
+
+        self.cmb_1 = QtWidgets.QComboBox()
+        lbl_1 = QtWidgets.QLabel('Bands:')
+
+        hbl.addWidget(menu_default.HelpButton('cluster.cm.showbars'), 0,
[…]
