### 2025-04-24 409118b
Updates
Updated show surface to PyVista
Updated 3D model display to PyVista
Linked help to raster, vector, boreholes, classification, EM, gravity, magnetics, and MT menus
