+++ pygmi/raster/graphs.py
@@ -41,10 +41,11 @@ from matplotlib.backends.backend_qtagg import FigureCanvasQTAgg
 from matplotlib.figure import Figure
 from matplotlib.backends.backend_qt import NavigationToolbar2QT
 import matplotlib.colors as mcolors
+import pyvista as pv
+from pyvistaqt import QtInteractor
 
 from pygmi.misc import frm, ContextModule
 from pygmi.raster.modest_image import imshow
-from pygmi import menu_default
 
 
 class MyMplCanvas(FigureCanvasQTAgg):
@@ -313,11 +314,12 @@ class PlotCCoef(ContextModule):
 
         self.mmc = MyMplCanvas(self)
         mpl_toolbar = NavigationToolbar2QT(self.mmc, self.parent)
-        helpdocs = menu_default.HelpButton('raster.cm.stats')
+        self.buttonbox.buttonbox.hide()
+        self.buttonbox.htmlfile = 'raster.cm.showcorr'
 
         vbl.addWidget(self.mmc)
         hbl.addWidget(mpl_toolbar)
-        hbl.addWidget(menu_default.HelpButton('raster.cm.showcorr'))
+        hbl.addWidget(self.buttonbox)
         vbl.addLayout(hbl)
 
         self.setMinimumSize(600, 600)
@@ -343,11 +345,10 @@ class PlotCCoef(ContextModule):
             return
 
         if not check_bands(data):
-            QtWidgets.QMessageBox.warning(self, 'Warning',
-                                          'Different size input datasets. '
-                                          'Merge and resample your input data '
-                                          'to fix this.',
-                                          QtWidgets.QMessageBox.StandardButton.Ok)
+            QtWidgets.QMessageBox.warning(
+                self, 'Warning', 'Different size input datasets. '
+                'Merge and resample your input data to fix this.',
+                QtWidgets.QMessageBox.StandardButton.Ok)
             return
 
         self.show()
@@ -379,7 +380,10 @@ class PlotRaster(ContextModule):
         self.mmc = MyMplCanvas(self)
         mpl_toolbar = NavigationToolbar2QT(self.mmc)
 
-        hbl.addWidget(menu_default.HelpButton('raster.cm.showsimple'))
+        self.buttonbox.buttonbox.hide()
+        self.buttonbox.htmlfile = 'raster.cm.showsimple'
+
+        hbl.addWidget(self.buttonbox)
 
         self.cmb_1 = QtWidgets.QComboBox()
         lbl_1 = QtWidgets.QLabel('Bands:')
@@ -453,10 +457,14 @@ class PlotSurface(ContextModule):
 
         vbl = QtWidgets.QVBoxLayout(self)  # self is where layout is assigned
         hbl = QtWidgets.QHBoxLayout()
-        self.mmc = MyMplCanvas(self)
-        mpl_toolbar = NavigationToolbar2QT(self.mmc, self.parent)
+        hbl2 = QtWidgets.QHBoxLayout()
 
-        hbl.addWidget(menu_default.HelpButton('raster.cm.showsurface'))
+        self.plotter = QtInteractor(self)
+        self.vslider = QtWidgets.QSlider()
+
+        self.buttonbox.buttonbox.hide()
+        self.buttonbox.htmlfile = 'raster.cm.showsurface'
+        hbl.addWidget(self.buttonbox)
 
         self.cmb_1 = QtWidgets.QComboBox()
         lbl_1 = QtWidgets.QLabel('Bands:')
@@ -469,14 +477,17 @@ class PlotSurface(ContextModule):
         hbl.addWidget(self.cmb_2)
         self.cmb_2.addItems(['viridis', 'jet', 'gray', 'terrain'])
 
-        vbl.addWidget(self.mmc)
-        vbl.addWidget(mpl_toolbar)
[…]
