+++ pygmi/rsense/hyperspec.py
@@ -1406,4 +1406,4 @@ def _testfn2():
 
 
 if __name__ == "__main__":
-    _testfn2()
+    _testfn()
