### 2025-04-24 a64c037
Update
Connected help to pfmod, rsense and seis.
Fixed a bug importing some macroseismic data.
