+++ pygmi/rsense/iodefs.py
@@ -35,7 +35,7 @@ import datetime
 from collections import defaultdict
 import warnings
 
-from PyQt6 import QtWidgets, QtCore
+from PyQt6 import QtWidgets
 import numpy as np
 import numexpr as ne
 import pandas as pd
@@ -45,7 +45,6 @@ import rasterio
 from natsort import natsorted
 from pyproj.crs import CRS
 
-from pygmi import menu_default
 from pygmi.misc import ProgressBarText, ContextModule, BasicModule
 from pygmi.raster.datatypes import Data, RasterMeta, bounds_intersection
 from pygmi.raster.iodefs import get_raster, export_raster
@@ -94,6 +93,7 @@ class ImportData(BasicModule):
         None.
 
         """
+        self.buttonbox.htmlfile = 'rsense.dm.importdata'
         pb_sfile = QtWidgets.QPushButton(' Filename')
 
         pixmapi = QtWidgets.QStyle.StandardPixmap.SP_DialogOpenButton
@@ -106,25 +106,16 @@ class ImportData(BasicModule):
 
         gl_1 = QtWidgets.QGridLayout(self)
 
-        self.lw_tnames.setSelectionMode(QtWidgets.QAbstractItemView.SelectionMode.MultiSelection)
+        self.lw_tnames.setSelectionMode(
+            QtWidgets.QAbstractItemView.SelectionMode.MultiSelection)
 
         gl_1.addWidget(pb_sfile, 1, 0, 1, 1)
         gl_1.addWidget(self.le_sfile, 1, 1, 1, 1)
         gl_1.addWidget(self.lbl_ftype, 2, 0, 1, 2)
         gl_1.addWidget(self.lw_tnames, 3, 0, 1, 2)
         gl_1.addWidget(self.cb_ensuresutm, 4, 0, 1, 2)
+        gl_1.addWidget(self.buttonbox, 9, 0, 1, 2)
 
-        buttonbox = QtWidgets.QDialogButtonBox()
-        buttonbox.setOrientation(QtCore.Qt.Orientation.Horizontal)
-        buttonbox.setCenterButtons(True)
-        buttonbox.setStandardButtons(
-            buttonbox.StandardButton.Cancel |
-            buttonbox.StandardButton.Ok)
-
-        gl_1.addWidget(buttonbox, 9, 0, 1, 2)
-
-        buttonbox.accepted.connect(self.accept)
-        buttonbox.rejected.connect(self.reject)
         pb_sfile.pressed.connect(self.get_sfile)
 
     def settings(self, nodialog=False):
@@ -168,9 +159,9 @@ class ImportData(BasicModule):
                        tnames=tnames)
 
         if dat is None:
-            QtWidgets.QMessageBox.warning(self.parent, 'Error',
-                                          'Could not import the data.',
-                                          QtWidgets.QMessageBox.StandardButton.Ok)
+            QtWidgets.QMessageBox.warning(
+                self.parent, 'Error', 'Could not import the data.',
+                QtWidgets.QMessageBox.StandardButton.Ok)
             return False
 
         if self.cb_ensuresutm.isChecked() is True:
@@ -294,6 +285,7 @@ class ImportBatch(BasicModule):
         None.
 
         """
+        self.buttonbox.htmlfile = 'rsense.dm.createbatchlist'
         pb_sfile = QtWidgets.QPushButton(' Directory')
 
         pixmapi = QtWidgets.QStyle.StandardPixmap.SP_DialogOpenButton
@@ -306,7 +298,8 @@ class ImportBatch(BasicModule):
 
         gl_1 = QtWidgets.QGridLayout(self)
 
[…]
