+++ pyproject.toml
@@ -33,25 +33,27 @@ classifiers = [
 ]
 
 dependencies = [
-                'fiona>=1.9.5',
-                'geopandas>=0.14.4',
-                'h5netcdf>=1.3.0',
-                'matplotlib>=3.9.0',
+                'fiona>=1.10.1',
+                'geopandas>=1.0.1',
+                'h5netcdf>=1.6.1',
+                'matplotlib>=3.10.1',
                 'mtpy>=1.1.5',
                 'natsort>=8.4.0',
-                'numba>=0.61.0',
-                'numexpr>=2.10.1',
-                'openpyxl>=3.1.2',
-                'psutil>=6.0.0',
-                'pyopengl>=3.1.7',
-                'pyqt6>=6.8.1',
-                'pytest>=8.2.2',
-                'rasterio>=1.3.9',
-                'rioxarray>=0.15.6',
-                'scikit-learn>=1.6.0',
-                'scikit-image>=0.24.0',
-                'shapelysmooth>=0.2.0',
-                'simpeg>=0.21.1',
+                'numba>=0.61.2',
+                'numexpr>=2.10.2',
+                'openpyxl>=3.1.5',
+                'psutil>=7.0.0',
+                'pyqt6>=6.9.0',
+                'pytest>=8.3.5',
+                'pyvista>=0.45.0',
+                'pyvistaqt>=0.11.2',
+                'pywin32>=310; os_name == "nt"',
+                'rasterio>=1.4.3',
+                'rioxarray>=0.19.0',
+                'scikit-learn>=1.6.1',
+                'scikit-image>=0.25.2',
+                'shapelysmooth>=0.2.1',
+                'simpeg>=0.24.0',
                 ]
 
 [project.urls]
