### 2025-05-07 #7019 @Patrick-Cole
Hi,

I am just enquiring about progress with the review. On a related  note, I have released a new version of PyGMI which should address outstanding issues. It uses Qt6 to deal with the Wayland issue and has vastly better integrated help.

Regards

Patrick
