+++ pygmi/rsense/usgs.py
@@ -1,0 +1,413 @@
+# -----------------------------------------------------------------------------
+# Name:        usgs.py (part of PyGMI)
+#
+# Author:      Patrick Cole
+# E-Mail:      pcole@geoscience.org.za
+#
+# Copyright:   (c) 2025 Council for Geoscience
+# Licence:     GPL-3.0
+#
+# This file is part of PyGMI
+#
+# PyGMI is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# PyGMI is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+# -----------------------------------------------------------------------------
+"""
+USGS SPECPR import.
+
+
+Variable                  Description                             (bytes)
+-------------------------------------------------------------------------
+icflag          32 one bit flags:                                     4
+                (low bit = 0, high bit = 31)
+                bit 00 continuation data flag
+                        =0 first record of a spectrum consists of:
+                                header then 256 data channels
+                        =1 continuation data record consisting of:
+                                bit flags followed by 1532 bytes of
+                                real data (bit 1=0) (383 channels)
+                                or 1532 bytes of text (bit 1=1).
+                                A maximum of 12 continuation records
+                                are allowed for a total of 4852
+                                channels (limited by arrays of 4864)
+                                or 19860 characters of text (bit 1=1).
+                bit 01 text/data flag:
+                       =0 the data in the array "data" is data
+                       =1 the data in the array "data" is ascii text
+                                 as is most of the header info.
+                bit 02 flag to indicate whether or not the data for
+                       the error bar (1 sigma standard deviation of
+                       the mean) is in the next record set.
+                       =0: no errors, =1: errors in next record set.
+                bit 03 RA, Dec / Long., Lat flag:
+                       =0 the array "ira" and "idec" corresponds to
+                          the right ascension and declination of an
+                          astronomical object.
+                       =1 the array "ira" and "idec" correspond to
+                          the longitude and latitude of a spot on a
+                          planetary surface.
+                bit 04 variable iscta universal time/civil time flag
+                       =0 cta is civil time
+                       =1 cta is universal time
+                bit 05 variable isctb universal time/civil time flag
+                       =0 ctb is civil time
+                       =1 ctb is universal time
+                bit 06 unused
+                bit 07 unused
+                bit 08 unused
+                bit 09 unused
+                bit 10 unused
+                bit 11 unused
+                bit 12 unused
+                bit 13 unused
+                bit 14 unused
+                bit 15 unused
+                bit 16 unused
+                bit 17 unused
+                bit 18 unused
+                bit 19 unused
+                bit 20 unused
+                bit 21 unused
+                bit 22 unused
+                bit 23 unused
+                bit 24 unused
+                bit 25 unused
+                bit 26 unused
[…]
