### 2025-05-07 #7019 @AnkitBarik
Hi, I'm really really sorry for taking this long for the review. One limitation is that I am running Ubuntu 22.04 which only ships with Python 3.10 and an older version of `libgdal` (3.4.1) which seems incompatible with the requirements. Thus, I have not tested the CLI installation, but I tried out the GUI. Firstly, excellent job! I have a few comments:

#### Functionality

1. Is it possible to have some level of backward compatibility? Most users will not be running Python 3.11 or 3.12. Will it completely break the software if you use Python 3.10 for example?

#### Documentation

1. When I click on the API reference links, it shows me a 404 not found page. Could you check if the docs are getting built properly?
2. In the example section - could you please add examples of analyzing datasets already present in the repository/install directory? It would be very useful for new users without having to try to find a dataset for testing out.

I'm almost done with my checklist. Again, apologies for taking this long.
