### 2025-05-07 #7019 @boisgera
Hi @AnkitBarik,

If you need a recent version of python and libgdal, you may want to try the (isolated) development environments provided by pixi (from https://prefix.dev/). The commands

```
$ pixi init pygmi
$ cd pygmi
$ pixi add python
$ pixi add libgdal
# ... extra dependencies
$ pixi shell
```

should put you in a suitable environment for testing purposes. Just delete the pygmi directory when you're done!
