### 2025-05-07 #7019 @AnkitBarik
@boisgera Thank you very much! I was able to test out both the Linux and Windows installations and they work fine. I was also able to test out some of the features using the gravity data in the test directory. Couple of comments for @Patrick-Cole :

1. When I open the UI on Windows, the UI takes on the "dark mode" of the OS, hence any symbol in black (like the joining line icon and play arrow) cannot be seen. This is purely cosmetic and does not need to be fixed for the publication.
2. It would be good to add a github workflow for code testing, but I'm not sure how memory intensive that is. Is it possible to add a subset of really small test cases to test out individual components? This usually helps new users to see if a certain commit broke the package or not. I'm not sure if this is a necessity for publication though - @boisgera ?
