### 2025-05-08 #7019 @Patrick-Cole
Hi,

No worries about the delays - I am learning a lot through this process its all good. To answer your questions:

1. I have not tested it recently with python 3.10, but I can do so. You can also use Anaconda on Ubuntu, there are instructions for that on the wiki. Your point is noted however. I am not sure how much backwards compatibility is recommended, so any guidance is appreciated. Since I am currently the sole developer, I tend to recommend the latest version, but I can amend that (with testing). Gdal is an annoying library, since binaries are not always available. However, you no longer need it for the latest version since rasterio comes with its own gdal, and I use that. I know Anaconda always has a  new gdal though.
2. Thanks for the API heads up. I fixed that this morning.

I will also look into you other two points and get back to you.

Regards

Patrick
