+++ _sources/autoapi/pygmi/grav/menu/index.rst
@@ -1,0 +1,45 @@
+pygmi.grav.menu
+===============
+
+.. py:module:: pygmi.grav.menu
+
+.. autoapi-nested-parse::
+
+   Gravity Menu Routines.
+
+
+
+Classes
+-------
+
+.. autoapisummary::
+
+   pygmi.grav.menu.MenuWidget
+
+
+Module Contents
+---------------
+
+.. py:class:: MenuWidget(parent=None)
+
+   Widget class to call the main interface.
+
+   This widget class creates menus to be found on the main
+   interface. Normal as well as context menus are defined here.
+
+   :param parent: Reference to MainWidget class found in main.py. Default is None.
+   :type parent: pygmi.main.MainWidget, optional
+
+
+   .. py:method:: import_data()
+
+      Import data.
+
+
+
+   .. py:method:: process_data()
+
+      Process data.
+
+
+
