+++ _sources/autoapi/pygmi/test/test_cluster/index.rst
@@ -1,0 +1,40 @@
+pygmi.test.test_cluster
+=======================
+
+.. py:module:: pygmi.test.test_cluster
+
+.. autoapi-nested-parse::
+
+   These are tests. Run pytest on this file from within this directory to do
+   the tests.
+
+
+
+Functions
+---------
+
+.. autoapisummary::
+
+   pygmi.test.test_cluster.test_cluster
+   pygmi.test.test_cluster.test_crisp
+   pygmi.test.test_cluster.test_fuzzy
+
+
+Module Contents
+---------------
+
+.. py:function:: test_cluster()
+
+   test cluster.
+
+
+.. py:function:: test_crisp()
+
+   test crisp cluster.
+
+
+.. py:function:: test_fuzzy()
+
+   test fuzzy cluster.
+
+
