+++ _sources/autoapi/pygmi/test/index.rst
@@ -1,0 +1,24 @@
+pygmi.test
+==========
+
+.. py:module:: pygmi.test
+
+.. autoapi-nested-parse::
+
+   Tests for PyGMI.
+
+
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 1
+
+   /autoapi/pygmi/test/test_cluster/index
+   /autoapi/pygmi/test/test_gravity/index
+   /autoapi/pygmi/test/test_mag/index
+   /autoapi/pygmi/test/test_pfmod/index
+   /autoapi/pygmi/test/test_raster/index
+
+
