### 2025-05-08 #7019 @boisgera
> 2. It would be good to add a github workflow for code testing, but I'm not sure how memory intensive that is. Is it possible to add a subset of really small test cases to test out individual components? This usually helps new users to see if a certain commit broke the package or not. I'm not sure if this is a necessity for publication though - [@boisgera](https://github.com/boisgera) ?

JOSS provides somes guidelines here: https://joss.readthedocs.io/en/latest/review_criteria.html#tests.

There are some references to tests in the documentation ([here](https://patrick-cole.github.io/pygmi/autoapi/pygmi/test/index.html#module-pygmi.test)) and each subdirectory (such as [this one](https://patrick-cole.github.io/pygmi/autoapi/pygmi/test/test_cluster/index.html)) has basic instructions on how to run the tests. So AFAICT, that qualifies as OK.

Automated tests are not mandatory for publication, but as you point out they would be a good thing for the project. @AnkitBarik I would suggest that you add an issue to the PyGMI project as a reminder of this (to be addressed later, but not blocking for publication).
