### 2025-05-08 Patrick-Cole/pygmi#26 @AnkitBarik
Feature request : Github workflow for unit tests
This is a feature request to add a GitHub workflow to run unit tests automatically after a push to the repository. An example of a workflow script can be found [here](https://github.com/AnkitBarik/planetMagFields/blob/main/.github/workflows/main.yml).

Cheers,
Ankit
