### 2025-05-08 #7019 @AnkitBarik
Good point. I have added an [issue](https://github.com/Patrick-Cole/pygmi/issues/26) to the repository. I have checked off almost everything on the checklist. @Patrick-Cole @boisgera - it would be good to have one example documented that uses a dataset in the repository, but the given example(s) satisfy the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#example-usage). So, I'll go ahead and check that off as well.

@Patrick-Cole I was able to evaluate the CLI interface using @boisgera 's `pixi` trick. It would be eventually good to have some backward compatibility, at least till Python 3.10, but it's not necessary to get it published.

This completes my review, I am happy to recommend the publication of this paper in JOSS.
