### 2025-05-12 402f925
Update
Added some different icons for dark themes.
