### 2025-05-12 Patrick-Cole/pygmi#23 @Patrick-Cole
I have updated qt to qt6, which seems to solve this problem.
