+++ requirements.txt
@@ -14,7 +14,6 @@ pyqt6>=6.9.0
 pytest>=8.3.5
 pyvista>=0.45.0
 pyvistaqt>=0.11.2
-pywin32>=310
 rasterio>=1.4.3
 rioxarray>=0.19.0
 scikit-learn>=1.6.1
