### 2025-05-12 #7019 @boisgera
Hi @Patrick-Cole!

Could you have a look at https://github.com/Patrick-Cole/pygmi/pull/27 please?

That should fix the warnings wrt DOIs provided in https://github.com/openjournals/joss-reviews/issues/7019#issuecomment-287174057.
