+++ .github/workflows/python-app.yml
@@ -36,6 +36,8 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+        # Ignore some errors
+        flake8 . --ignore=C901
     - name: Test with pytest
       run: |
         cd ${{github.workspace}}/pygmi/test
