+++ .github/workflows/python-app.yml
@@ -33,11 +33,9 @@ jobs:
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics --ignore=C901
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-        # Ignore some errors
-        flake8 . --ignore=C901
     - name: Test with pytest
       run: |
         cd ${{github.workspace}}/pygmi/test
