+++ .github/workflows/python-app.yml
@@ -25,8 +25,8 @@ jobs:
         python-version: "3.10"
     - name: Install dependencies
       run: |
-        sudo apt-get update
-        sudo apt-get -y install freeglut3-dev
+        # sudo apt-get update
+        # sudo apt-get -y install freeglut3-dev
         python -m pip install --upgrade pip
         pip install flake8 pytest
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
@@ -35,7 +35,7 @@ jobs:
         # stop the build if there are Python syntax errors or undefined names
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exit-zero --max-line-length=127 --statistics
+        flake8 . --count --exit-zero --max-line-length=127 --statistics --ignore=F841
     - name: Test with pytest
       run: |
         cd ${{github.workspace}}/pygmi/test
