+++ pygmi/clust/fuzzy_clust.py
@@ -881,5 +881,6 @@ def _testfn():
     plt.colorbar()
     plt.show()
 
+
 if __name__ == "__main__":
     _testfn()
