+++ pygmi/clust/menu.py
@@ -81,8 +81,7 @@ class MenuWidget():
         self.menu.addAction(self.action_segmentation)
         self.action_segmentation.triggered.connect(self.segmentation)
 
-        self.action_super_class = QtGui.QAction("Supervised "
-                                                    "Classification")
+        self.action_super_class = QtGui.QAction("Supervised Classification")
         self.menu.addAction(self.action_super_class)
         self.action_super_class.triggered.connect(self.super_class)
 
@@ -95,8 +94,7 @@ class MenuWidget():
 # Context menus
         context_menu['Cluster'].addSeparator()
 
-        self.action_cluster_statistics = QtGui.QAction('Cluster '
-                                                           'Statistics')
+        self.action_cluster_statistics = QtGui.QAction('Cluster Statistics')
         context_menu['Cluster'].addAction(self.action_cluster_statistics)
         self.action_cluster_statistics.triggered.connect(self.cluster_stats)
 
@@ -109,13 +107,13 @@ class MenuWidget():
         self.action_show_class_range.triggered.connect(self.show_class_range)
 
         self.action_show_membership_data = QtGui.QAction("Show Membership "
-                                                             "Data")
+                                                         "Data")
         context_menu['memCluster'].addAction(self.action_show_membership_data)
         self.action_show_membership_data.triggered.connect(self.show_membership_data)
 
         self.action_show_objvrcncexbigraphs = QtGui.QAction("Show OBJ, "
-                                                                "VRC, NCE, "
-                                                                "XBI Graphs")
+                                                            "VRC, NCE, "
+                                                            "XBI Graphs")
         context_menu['objCluster'].addAction(self.action_show_objvrcncexbigraphs)
         self.action_show_objvrcncexbigraphs.triggered.connect(self.show_vrc_etc)
 
