+++ pygmi/bholes/graphs.py
@@ -628,7 +628,7 @@ def chkname(iname):
 def _testfn():
     """Test routine."""
     import sys
-    from matplotlib import rcParams
+    # from matplotlib import rcParams
     from pygmi.bholes.iodefs import get_CGS
 
     # rcParams['font.size'] = 4
