+++ pygmi/bholes/__init__.py
@@ -2,4 +2,3 @@
 These routines cover the import and display of Council for Geoscience borehole
 data. This data can also be used in modelling in other modules.
 """
-
