+++ .github/workflows/python-app.yml
@@ -35,7 +35,7 @@ jobs:
         # stop the build if there are Python syntax errors or undefined names
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exit-zero --max-line-length=127 --statistics --ignore=F841
+        flake8 . --count --exit-zero --max-line-length=127 --statistics --ignore=F841,E226,E241
     - name: Test with pytest
       run: |
         cd ${{github.workspace}}/pygmi/test
