+++ .github/workflows/python-app.yml
@@ -39,4 +39,4 @@ jobs:
     - name: Test with pytest
       run: |
         cd ${{github.workspace}}/pygmi/test
-        # pytest
+        pytest
