### 2025-05-12 ee0235e
Update grvmag3d.py
