+++ pygmi/pfmod/grvmag3d.py
@@ -41,8 +41,10 @@ from PyQt6 import QtWidgets, QtCore, QtGui
 
 import numpy as np
 from numba import jit, prange
-from matplotlib import colormaps
 import matplotlib as mpl
+mpl.use('QtAgg')
+
+from matplotlib import colormaps
 import matplotlib.pyplot as plt
 
 from pygmi.raster.reproj import data_reproject
@@ -50,7 +52,7 @@ from pygmi.pfmod.datatypes import LithModel
 from pygmi.misc import PTime
 from pygmi.misc import frm
 
-mpl.use('QtAgg')
+
 
 
 class GravMag():
