### 2025-05-12 8144274
Update grvmag3d.py
