+++ .github/workflows/python-app.yml
@@ -28,7 +28,7 @@ jobs:
         # sudo apt-get update
         # sudo apt-get -y install libegl1
         python -m pip install --upgrade pip
-        pip install flake8 pytest
+        pip install flake8 pytest pywin32
         pip install -r requirements.txt
     - name: Lint with flake8
       run: |
