+++ .github/workflows/python-app.yml
@@ -15,7 +15,7 @@ permissions:
 jobs:
   build:
 
-    runs-on: windows-latest
+    runs-on: ubuntu-latest
 
     steps:
     - uses: actions/checkout@v4
@@ -25,11 +25,11 @@ jobs:
         python-version: "3.12"
     - name: Install dependencies
       run: |
-        # sudo apt-get update
-        # sudo apt-get -y install libegl1
+        sudo apt-get update
+        sudo apt-get -y install libegl1
         python -m pip install --upgrade pip
-        pip install flake8 pytest pywin32
-        pip install -r requirements.txt
+        pip install flake8 pytest
+        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
