+++ .github/workflows/python-app.yml
@@ -23,6 +23,7 @@ jobs:
       uses: actions/setup-python@v3
       with:
         python-version: "3.12"
+        test-env: "PyQt6~=6.9.0 PyQt6-Qt6~=6.9.0"
     - name: Install dependencies
       run: |
         sudo apt-get update
@@ -31,12 +32,12 @@ jobs:
         python -m pip install --upgrade pip
         pip install flake8 pytest
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-    - name: Lint with flake8
-      run: |
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exit-zero --max-line-length=127 --statistics --ignore=F841,E226,E241
+    # - name: Lint with flake8
+    #   run: |
+    #     # stop the build if there are Python syntax errors or undefined names
+    #     flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+    #     # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+    #     flake8 . --count --exit-zero --max-line-length=127 --statistics --ignore=F841,E226,E241
     - name: Test with pytest
       run: |
         cd ${{github.workspace}}/pygmi/test
