+++ pygmi/test/test_raster.py
@@ -1,0 +1,557 @@
+# -----------------------------------------------------------------------------
+# Name:        test_raster.py (part of PyGMI)
+#
+# Author:      Patrick Cole
+# E-Mail:      pcole@geoscience.org.za
+#
+# Copyright:   (c) 2019 Council for Geoscience
+# Licence:     GPL-3.0
+#
+# This file is part of PyGMI
+#
+# PyGMI is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# PyGMI is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+# -----------------------------------------------------------------------------
+"""
+These are tests. Run pytest on this file from within this directory to do
+the tests.
+"""
+
+import os
+import glob
+# import sys
+import tempfile
+# from PyQt6 import QtWidgets
+import numpy as np
+from pyproj.crs import CRS
+import pytest
+
+from pygmi.raster.datatypes import Data
+from pygmi.raster import cooper, dataprep, equation_editor, ginterp, graphs
+from pygmi.raster import iodefs, normalisation, smooth
+from pygmi.raster.misc import aspect2, check_dataid
+
+# APP = QtWidgets.QApplication(sys.argv)  # Necessary to test Qt Classes
+
+
+def test_gradients():
+    """test directional derivative."""
+    datin = [[1, 2], [1, 2]]
+    dat2 = [[-1, -1], [-1, -1]]
+    dat = cooper.gradients(datin, 0, 1, 1)
+    np.testing.assert_array_equal(dat, dat2)
+
+
+def test_dratio():
+    """test derivative ratio."""
+    datin = [[1, 2], [1, 2]]
+    dat2 = [[-np.pi/2, -np.pi/2], [-np.pi/2, -np.pi/2]]
+    dat = cooper.derivative_ratio(datin, 90, 1)
+    np.testing.assert_array_equal(dat, dat2)
+
+
+def test_thgrad():
+    """test total horizontal gradient."""
+    datin = [[1, 2], [1, 2]]
+    dat2 = [[0.1, 0.1], [0.1, 0.1]]
+    dat = cooper.thgrad(datin, 10, 10)
+    np.testing.assert_array_equal(dat, dat2)
+
+
+def test_vertical():
+    """test vertical derivative."""
+    datin = np.array([[1, 2], [1, 2]])
+    dat2 = np.array([[-0.90757121,  0.90757121],
+                     [-0.90757121,  0.90757121]])
+    dat = cooper.vertical(datin, 10)
+    np.testing.assert_array_almost_equal(dat, dat2)
+
+
+def test_viz():
+    """test visibility."""
+    datin = np.ma.array([[1, 2], [1, 2]])
+    vtot2 = [[4., 4.], [4., 4.]]
+    vstd2 = [[0.5345224838248488, 0.5345224838248488],
+             [0.5345224838248488, 0.5345224838248488]]
+    vsum2 = [[2.613125929752753, 2.613125929752753],
[…]
