+++ .github/workflows/python-app.yml
@@ -23,12 +23,11 @@ jobs:
       uses: actions/setup-python@v3
       with:
         python-version: "3.12"
-        test-env: "PyQt6~=6.9.0 PyQt6-Qt6~=6.9.0"
     - name: Install dependencies
       run: |
         sudo apt-get update
-        sudo apt-get -y install libgl1
-        sudo apt-get -y install libegl1
+        sudo apt-get -y install xorg libxkbcommon0
+        # sudo apt-get -y install libegl1
         python -m pip install --upgrade pip
         pip install flake8 pytest
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
