### 2025-05-12 e3c655e
Update grvmag3d.py
