+++ pygmi/test/test_raster.py
@@ -117,18 +117,18 @@ def test_trimraster():
     np.testing.assert_array_equal(dat[0].data, dat2)
 
 
-def test_equation():
-    """tests equation editor."""
-    datin = Data()
-    datin.data = np.ma.array([[1., 2.], [1., 2.]])
-    datout = datin.data*2
+# def test_equation():
+#     """tests equation editor."""
+#     datin = Data()
+#     datin.data = np.ma.array([[1., 2.], [1., 2.]])
+#     datout = datin.data*2
 
-    tmp = equation_editor.EquationEditor()
-    tmp.indata = {'Raster': [datin, datin]}
-    tmp.equation = 'i0+i1'
-    tmp.settings(True)
+#     tmp = equation_editor.EquationEditor()
+#     tmp.indata = {'Raster': [datin, datin]}
+#     tmp.equation = 'i0+i1'
+#     tmp.settings(True)
 
-    np.testing.assert_array_equal(tmp.outdata['Raster'][0].data, datout)
+#     np.testing.assert_array_equal(tmp.outdata['Raster'][0].data, datout)
 
 
 def test_hmode():
@@ -333,217 +333,217 @@ def test_io_rasterio(smalldata, ext, drv):
     np.testing.assert_array_equal(smalldata.data, dat2[0].data)
 
 
-def test_io_ascii(smalldata):
-    """Tests IO for ascii files."""
-    ofile = tempfile.gettempdir() + '\\iotest.asc'
+# def test_io_ascii(smalldata):
+#     """Tests IO for ascii files."""
+#     ofile = tempfile.gettempdir() + '\\iotest.asc'
 
-    tmp = iodefs.ExportData(None)
-    tmp.ofile = ofile
-    tmp.export_ascii([smalldata])
+#     tmp = iodefs.ExportData(None)
+#     tmp.ofile = ofile
+#     tmp.export_ascii([smalldata])
 
-    dat2 = iodefs.get_ascii(ofile)
+#     dat2 = iodefs.get_ascii(ofile)
 
-    # Cleanup files
-    for i in glob.glob(tempfile.gettempdir() + '\\iotest*'):
-        os.unlink(i)
+#     # Cleanup files
+#     for i in glob.glob(tempfile.gettempdir() + '\\iotest*'):
+#         os.unlink(i)
 
-    np.testing.assert_array_equal(smalldata.data, dat2[0].data)
-
-
-def test_io_xyz(smalldata):
-    """Tests IO for xyz files."""
-    ofile = tempfile.gettempdir() + '\\iotest.xyz'
-
-    tmp = iodefs.ExportData(None)
-    tmp.ofile = ofile
-    tmp.export_ascii_xyz([smalldata])
-
-    dat2 = iodefs.get_raster(ofile)
-
-    # Cleanup files
-    for i in glob.glob(tempfile.gettempdir() + '\\iotest*'):
-        os.unlink(i)
-
-    np.testing.assert_array_equal(smalldata.data, dat2[0].data)
+#     np.testing.assert_array_equal(smalldata.data, dat2[0].data)
 
 
-def test_normalisation():
-    """Tests for normalisation."""
+# def test_io_xyz(smalldata):
+#     """Tests IO for xyz files."""
+#     ofile = tempfile.gettempdir() + '\\iotest.xyz'
 
-    datin = Data()
-    datin.data = np.ma.array([[1., 2.], [1., 2.]])
+#     tmp = iodefs.ExportData(None)
+#     tmp.ofile = ofile
+#     tmp.export_ascii_xyz([smalldata])
 
-    tmp = normalisation.Normalisation(None)
-    tmp.indata = {'Raster': [datin, datin]}
[…]
