+++ .github/workflows/python-app.yml
@@ -25,9 +25,9 @@ jobs:
         python-version: "3.12"
     - name: Install dependencies
       run: |
-        # sudo apt-get update
+        sudo apt-get update
         # sudo apt-get -y install xorg libxkbcommon0
-        # sudo apt-get -y install libegl1
+        sudo apt-get -y install libegl1
         python -m pip install --upgrade pip
         pip install flake8 pytest
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
