+++ pygmi/test/test_mag.py
@@ -1,0 +1,137 @@
+# -----------------------------------------------------------------------------
+# Name:        test_mag.py (part of PyGMI)
+#
+# Author:      Patrick Cole
+# E-Mail:      pcole@geoscience.org.za
+#
+# Copyright:   (c) 2019 Council for Geoscience
+# Licence:     GPL-3.0
+#
+# This file is part of PyGMI
+#
+# PyGMI is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# PyGMI is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+# -----------------------------------------------------------------------------
+"""
+These are tests. Run pytest on this file from within this directory to do
+the tests.
+"""
+
+# import sys
+# from PyQt6 import QtCore
+import numpy as np
+from pyproj.crs import CRS
+
+from pygmi.raster.datatypes import Data
+from pygmi.mag import dataprep
+from pygmi.mag import igrf, tiltdepth
+
+
+def test_tilt1():
+    """test tilt angle."""
+    datin = np.ma.array([[1, 2], [1, 2]])
+    t12 = [[-1.2626272556789115, 1.2626272556789115],
+           [-1.2626272556789115, 1.2626272556789115]]
+    th2 = [[0.0, 0.0],
+           [0.0, 0.0]]
+    t22 = [[0.0, 0.0],
+           [0.0, 0.0]]
+    ta2 = [[1.8572654805528055e-17, 1.8572654805528055e-17],
+           [1.8572654805528055e-17, 1.8572654805528055e-17]]
+    tdx2 = [[0.30816907111598496, 0.30816907111598496],
+            [0.30816907111598496, 0.30816907111598496]]
+    t1, th, t2, ta, tdx, _, _ = dataprep.tilt1(datin, 90, 0)
+
+    np.testing.assert_array_equal(t1, t12)
+    np.testing.assert_array_equal(th, th2)
+    np.testing.assert_array_equal(t2, t22)
+    np.testing.assert_array_equal(ta, ta2)
+    np.testing.assert_array_equal(tdx, tdx2)
+
+
+def test_rtp():
+    """Test RTP."""
+    datin = Data()
+    datin.data = np.ma.array([[1, 2], [1, 2]])
+    dat2 = [[0.7212671143002998, 1.9651600796627182],
+            [1.060458126573062, 1.8041542185243205]]
+
+    dat = dataprep.rtp(datin, 60, 30)
+
+    np.testing.assert_array_equal(dat.data, dat2)
+
+
+# def test_IGRF():
+#     """Tests IGRF Calculation."""
+#     dat = Data()
+#     dat.data = np.ma.array([[29000., 29000.], [29000., 29000.]],
+#                            mask=[[0, 0], [0, 0]])
+
+#     dat.set_transform(1, 25, 1, -27)
+#     dat.crs = CRS.from_epsg(4326)
+#     dat.dataid = 'mag'
+
+#     datin2 = Data()
+#     datin2.data = np.ma.array([[0., 0.], [0., 0.]], mask=[[0, 0], [0, 0]])
+
+#     datin2.set_transform(1, 25, 1, -27)
+#     datin2.crs = CRS.from_epsg(4326)
[…]
