### 2025-05-12 1667ef5
Update test_mag.py
