### 2025-05-13 8e81966
Update igrf.py
