+++ pygmi/mag/dataprep.py
@@ -218,48 +218,49 @@ def tilt1(data, azi, s, k=2):
     """
     dmin = data.min()
     dmax = data.max()
-    dm = 0.5*(dmin+dmax)
+    dm = 0.5 * (dmin + dmax)
     data.data[data.mask] = dm
     data[np.isnan(data)] = dm
     data[np.isinf(data)] = dm
 
     if s > 0:
-        se = np.ones((s, s))/(s*s)
+        se = np.ones((s, s)) / (s * s)
         data2 = signal.convolve2d(data, se, 'valid')  # smooth
         mask = np.ma.getmaskarray(data.data)
         mask = signal.convolve2d(mask, se, 'valid')
         data = np.ma.array(data2, mask=mask)
 
     nr, nc = data.shape
-    dtr = np.pi/180.0
-    azi = azi*dtr
+    dtr = np.pi / 180.0
+    azi = azi * dtr
 
     dy, dx = np.gradient(data)
-    dxtot = np.ma.sqrt(dx*dx+dy*dy)
+    dxtot = np.ma.sqrt(dx * dx + dy * dy)
     nmax = np.max([nr, nc])
     npts = int(2**nextpow2(nmax))
     dz = vertical(data, npts, 1)
-    t1 = np.ma.arctan(dz/dxtot)
-    th = np.real(np.arctanh(np.nan_to_num(dz/dxtot)+(0+0j)))
+    t1 = np.ma.arctan(dz / dxtot)
+    th = np.real(np.arctanh(np.nan_to_num(dz / dxtot) + (0 + 0j)))
 
-    tdx = np.real(np.ma.arctan(dxtot/abs(dz)))
+    tdx = np.real(np.ma.arctan(dxtot / abs(dz)))
 
-    dx1 = dx*np.cos(azi)+dy*np.sin(azi)  # Standard directional derivative
-    dx2 = dx*np.cos(azi+np.pi/2)+dy*np.sin(azi+np.pi/2)
-    dxz = np.ma.sqrt(dx2*dx2+dz*dz)
-    ta = np.ma.arctan(dx1/dxz)         # Tilt directional derivative
+    # Standard directional derivative
+    dx1 = dx * np.cos(azi) + dy * np.sin(azi)
+    dx2 = dx * np.cos(azi + np.pi / 2) + dy * np.sin(azi + np.pi / 2)
+    dxz = np.ma.sqrt(dx2 * dx2 + dz * dz)
+    ta = np.ma.arctan(dx1 / dxz)         # Tilt directional derivative
 
     # 2nd order Tilt angle
 
     s = max(s, 3)
-    se = np.ones([s, s])/(s*s)
+    se = np.ones([s, s]) / (s * s)
     ts = signal.convolve2d(t1.filled(t1.mean()), se, 'same')
     ts = np.ma.array(ts, mask=t1.mask)
 
     [dxs, dys] = np.gradient(ts)
     dzs = vertical(ts, npts, 1)
-    dxtots = np.ma.sqrt(dxs*dxs+dys*dys)
-    t2 = np.ma.arctan(dzs/dxtots)
+    dxtots = np.ma.sqrt(dxs * dxs + dys * dys)
+    t2 = np.ma.arctan(dzs / dxtots)
 
     # Standard tilt angle, hyperbolic tilt angle, 2nd order tilt angle,
     # Tilt Based Directional Derivative, Total Derivative
@@ -267,17 +268,17 @@ def tilt1(data, azi, s, k=2):
     data = dxtot
     nr, nc = data.shape
     dy, dx = np.gradient(data)
[…]
