+++ pygmi/raster/iodefs.py
@@ -337,7 +337,7 @@ class ImportRGBData(BasicModule):
         if dat[0].data.dtype == np.uint16:
             iidat = np.iinfo(dat[0].data.dtype)
             dat[0].data = dat[0].data.astype(float)
-            dat[0].data = (dat[0].data-iidat.min)/(iidat.max-iidat.min)
+            dat[0].data = (dat[0].data - iidat.min) / (iidat.max - iidat.min)
 
         self.outdata[output_type] = dat
 
@@ -378,7 +378,7 @@ def clusterprep(dat):
             i.metadata['Cluster']['memdat'] = [[]] * numclus
             for j in dat:
                 if 'Membership' in j.dataid and i.dataid in j.dataid:
-                    cnt = int(j.dataid.split(':')[0].split()[-1])-1
+                    cnt = int(j.dataid.split(':')[0].split()[-1]) - 1
                     i.metadata['Cluster']['memdat'][cnt] = j.data
             dat2.append(i)
 
@@ -427,14 +427,14 @@ def get_ascii(ifile):
             nval = float(nval)
 
         ulxmap = float(adata[5])
-        ulymap = float(adata[7])+ydim*nrows
+        ulymap = float(adata[7]) + ydim * nrows
         if 'center' in adata[4].lower():
-            ulxmap = ulxmap - xdim/2
+            ulxmap = ulxmap - xdim / 2
         if 'center' in adata[6].lower():
-            ulymap = ulymap - ydim/2
+            ulymap = ulymap - ydim / 2
         adata = adata[12:]
     else:
-        with open(ifile[:-3]+'hdr', 'r', encoding='utf-8') as hfile:
+        with open(ifile[:-3] + 'hdr', 'r', encoding='utf-8') as hfile:
             tmp = hfile.readlines()
 
         xdim = float(tmp[0].split()[-1])
@@ -536,12 +536,12 @@ def get_raster(ifile, *, nval=None, piter=None, showlog=print,
     # ENVI Case
     if ext == 'hdr':
         ifile = ifile[:-4]
-        if os.path.exists(ifile+'.dat'):
-            ifile = ifile+'.dat'
-        elif os.path.exists(ifile+'.raw'):
-            ifile = ifile+'.raw'
-        elif os.path.exists(ifile+'.img'):
-            ifile = ifile+'.img'
+        if os.path.exists(ifile + '.dat'):
+            ifile = ifile + '.dat'
+        elif os.path.exists(ifile + '.raw'):
+            ifile = ifile + '.raw'
+        elif os.path.exists(ifile + '.img'):
+            ifile = ifile + '.img'
         elif not os.path.exists(ifile):
             return None
     # ER Mapper case
@@ -552,7 +552,7 @@ def get_raster(ifile, *, nval=None, piter=None, showlog=print,
                 clong = metadata.split('STMLO')[1][:2]
 
                 if 'CAPE' in metadata:
-                    custom_wkt = ('PROJCS["Cape / TM'+clong+'",'
+                    custom_wkt = ('PROJCS["Cape / TM' + clong + '",'
                                   'GEOGCS["Cape",'
                                   'DATUM["Cape",'
                                   'SPHEROID["Clarke 1880 (Arc)",'
@@ -566,7 +566,8 @@ def get_raster(ifile, *, nval=None, piter=None, showlog=print,
                                   'AUTHORITY["EPSG","4222"]],'
                                   'PROJECTION["Transverse_Mercator"],'
                                   'PARAMETER["latitude_of_origin",0],'
-                                  'PARAMETER["central_meridian",'+clong+'],'
+                                  'PARAMETER["central_meridian",' +
+                                  clong + '],'
[…]
