### 2025-05-14 56133b1
Update
cluster crisp and fuzzy can now be called as functions
