+++ pygmi/pfmod/mvis3d.py
@@ -1472,7 +1472,7 @@ def _testfn():
     """Test function."""
     from pygmi.pfmod.iodefs import ImportMod3D
 
-    app = QtWidgets.QApplication(sys.argv)
+    _ = QtWidgets.QApplication(sys.argv)
     app.setStyle(QtWidgets.QStyleFactory.create('Fusion'))
 
     ifile = r'C:/Workdata/modelling/Magmodel_Upper22km_AveAll_diapir_withDeepDens_newdens.npz'
@@ -1494,7 +1494,7 @@ def _testfn2():
     """Test function."""
     from pygmi.pfmod.iodefs import ImportMod3D
 
-    app = QtWidgets.QApplication(sys.argv)
+    _ = QtWidgets.QApplication(sys.argv)
     app.setStyle(QtWidgets.QStyleFactory.create('Fusion'))
 
     ifile = r'C:/Workdata/modelling/Magmodel_Upper22km_AveAll_diapir_withDeepDens_newdens.npz'
