+++ pygmi/clust/fuzzy_clust.py
@@ -23,10 +23,11 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 # -----------------------------------------------------------------------------
 """
-Fuzzy clustering is a set of clustering routines, making use of fuzzy logic.
+Fuzzy clustering is a set of clustering routines.
+
+This makes use of fuzzy logic.
 """
 
-import os
 from PyQt6 import QtWidgets
 import numpy as np
 
@@ -206,10 +207,36 @@ class FuzzyClust(BasicModule):
             if temp == 0:
                 return False
 
-            self.parent.process_is_active()
-        self.acceptall()
+            if self.parent is not None:
+                self.parent.process_is_active()
 
-        if not nodialog:
+        data = [i.copy() for i in self.indata['Raster']]
+        self.update_vars()
+
+        ifiles = []
+        if self.init_type == 'manual':
+            ext = ('ASCII matrix (*.txt);;'
+                   'ASCII matrix (*.asc);;'
+                   'ASCII matrix (*.dat)')
+            filename = QtWidgets.QFileDialog.getOpenFileName(
+                self.parent, 'Read Cluster Centers', '.', ext)
+            if filename == '':
+                return False
+            ifiles.append(filename)
+
+            filename = QtWidgets.QFileDialog.getOpenFileName(
+                self.parent, 'Read Cluster Center Constraints', '.', ext)
+            ifiles.append(filename)
+
+        dat_out = fuzzyclust(data, self.cltype, self.min_cluster,
+                             self.max_cluster, self.constrain, self.runs,
+                             self.max_iter, self.fexp, self.term_thresh,
+                             self.init_type, ifiles, self.showlog, self.piter)
+
+        self.outdata['Cluster'] = dat_out
+        self.outdata['Raster'] = self.indata['Raster']
+
+        if not nodialog and self.parent is not None:
             self.parent.process_is_active(False)
             self.parent.pbar.to_max()
 
@@ -268,28 +295,28 @@ class FuzzyClust(BasicModule):
         self.constrain = self.dsb_constraincluster.value()
         self.fexp = self.dsb_fuzzynessexp.value()
 
-    def acceptall(self):
-        """
-        Run.
+        if self.rb_datadriven.isChecked():
+            self.init_type = 'data driven'
+        elif self.rb_manual.isChecked():
+            self.init_type = 'manual'
+        else:
+            self.init_type = 'random'
 
-        Returns
-        -------
-        None.
 
-        """
-        data = [i.copy() for i in self.indata['Raster']]
-        self.update_vars()
+def fuzzyclust(data, cltype='fuzzy c-means', min_cluster=5, max_cluster=5,
+               cov_constr=0., no_runs=1, max_iter=100, expo=1.5,
+               term_thresh=0.00001, init_type='random', ifiles=None,
+               showlog=print, piter=iter):
+    """
+    Run.
 
-        cltype = self.cltype
-        cov_constr = self.constrain
-        no_runs = self.runs
-        max_iter = self.max_iter
-        term_thresh = self.term_thresh
-        no_clust = np.array([self.min_cluster, self.max_cluster])
-        de_norm = self.denorm
-        expo = self.fexp
+    Returns
+    -------
+    None.
 
-        self.showlog('Fuzzy Clustering started')
+    """
+    no_clust = np.array([min_cluster, max_cluster])
+    showlog('Fuzzy Clustering started')
 
 # #############################################################################
[…]
