+++ pygmi/mag/igrf.py
@@ -880,7 +880,7 @@ def _testfn():
 
     dat = get_raster(ifile)
 
-    app = QtWidgets.QApplication(sys.argv)
+    _ = QtWidgets.QApplication(sys.argv)
 
     igrf = IGRF()
     igrf.indata['Raster'] = dat

[…]
