### 2025-05-14 4584ee4
Update
Normalization and equation editor can now be called via functions.
