+++ pygmi/raster/equation_editor.py
@@ -165,188 +165,6 @@ class EquationEditor(BasicModule):
         if txt != '':
             self.lbl_bands.setText(': ' + self.bands[txt])
 
-    def eq_fix(self, indata):
-        """
-        Corrects names in equation to variable names.
-
-        Parameters
-        ----------
-        indata : list of pygmi.raster.datatypes.Data.
-            PyGMI raster dataset.
-
-        Returns
-        -------
-        neweq : str
-            Corrected equation.
-
-        """
-        neweq = str(self.equation)
-        neweq = neweq.replace('ln', 'log')
-        neweq = neweq.replace('^', '**')
-        neweq = neweq.replace('nodata', str(indata[0].nodata))
-
-        if 'log' in neweq:
-            self.showlog('Warning, if you have invalid log values, they will '
-                         'be masked out.')
-
-        if 'sqrt' in neweq:
-            self.showlog('Warning, if you have invalid sqrt values, they will '
-                         'be masked out.')
-
-        neweq = neweq.strip()
-
-        return neweq
-
-    def mean(self, eq, localdict):
-        """
-        Get mean pixel value of all input bands.
-
-        Parameters
-        ----------
-        eq : str
-            Equation with std command.
-        localdict : dictionary
-            Dictionary of data.
-
-        Returns
-        -------
-        findat : numpy array
-            Output array.
-
-        """
-        idx = eq.index('mean(') + 5
-        eq2 = eq[idx:]
-        idx = eq2.index(')')
-        eq2 = eq2[:idx]
-        eq2 = eq2.replace(' ', '')
-        eq3 = eq2.split(',')
-
-        stack = []
-        mask = None
-        for i in localdict:
-            if i not in eq3:
-                continue
-            if mask is None:
-                mask = localdict[i].mask
-            else:
-                mask = np.logical_and(mask, localdict[i].mask)
-
-            if i == 'iall':
-                stack.append(localdict[i])
-            else:
-                stack.append([localdict[i]])
-
-        stack = np.ma.vstack(stack)
-        findat = np.ma.mean(stack, 0)
-        findat.mask = mask
-
-        return findat
-
-    def std(self, eq, localdict):
-        """
-        Get standard deviation pixel value of all input bands.
-
-        Parameters
-        ----------
-        eq : str
-            Equation with std command.
-        localdict : dictionary
-            Dictionary of data.
-
-        Returns
-        -------
-        findat : numpy array
-            Output array.
-
-        """
-        idx = eq.index('std(') + 4
-        eq2 = eq[idx:]
-        idx = eq2.index(')')
-        eq2 = eq2[:idx]
-        eq2 = eq2.replace(' ', '')
-        eq3 = eq2.split(',')
-
-        stack = []
-        mask = None
-        for i in localdict:
-            if i not in eq3:
-                continue
-            if mask is None:
-                mask = localdict[i].mask
-            else:
-                mask = np.logical_and(mask, localdict[i].mask)
-
-            if i == 'iall':
-                stack.append(localdict[i])
-            else:
-                stack.append([localdict[i]])
-
-        stack = np.ma.vstack(stack)
-        findat = np.ma.std(stack, 0)
-        findat.mask = mask
-
-        return findat
-
-    def mosaic(self, eq, localdict):
-        """
-        Mosaics data into a single band dataset.
-
-        Parameters
-        ----------
-        eq : str
-            Equation with mosaic command.
-        localdict : dictionary
-            Dictionary of data.
-
-        Returns
-        -------
[…]
