### 2025-05-15 1317fb9
Update
Smoothing can now be called via a function.
