### 2025-05-15 Patrick-Cole/pygmi#26 @Patrick-Cole
I have added this feature.
