+++ pygmi/rsense/landsat_composite.py
@@ -28,7 +28,7 @@ import os
 import glob
 from datetime import datetime
 import numpy as np
-from PyQt6 import QtWidgets
+from PySide6 import QtWidgets
 
 from pygmi.rsense.iodefs import get_data
 from pygmi.raster.misc import lstack
