+++ pygmi/rsense/iodefs.py
@@ -35,7 +35,7 @@ import datetime
 from collections import defaultdict
 import warnings
 
-from PyQt6 import QtWidgets
+from PySide6 import QtWidgets
 import numpy as np
 import numexpr as ne
 import pandas as pd
