+++ pygmi/pfmod/misc.py
@@ -25,7 +25,7 @@
 """These are miscellaneous functions for the program."""
 
 import time
-from PyQt6 import QtWidgets, QtCore, QtGui
+from PySide6 import QtWidgets, QtCore, QtGui
 import numpy as np
 import rasterio
 
