+++ pygmi/bholes/graphs.py
@@ -29,7 +29,7 @@ import re
 import textwrap
 import numpy as np
 import pandas as pd
-from PyQt6 import QtWidgets, QtCore
+from PySide6 import QtWidgets, QtCore
 from matplotlib.path import Path
 import matplotlib.patches as mpatches
 from matplotlib.backends.backend_qtagg import FigureCanvasQTAgg
