+++ pygmi/rsense/transforms.py
@@ -26,7 +26,7 @@
 
 import os
 import numpy as np
-from PyQt6 import QtWidgets
+from PySide6 import QtWidgets
 from sklearn.decomposition import IncrementalPCA
 import numexpr as ne
 import matplotlib.pyplot as plt
