+++ pyproject.toml
@@ -47,7 +47,6 @@ dependencies = [
                 'pytest>=8.3.5',
                 'pyvista>=0.45.0',
                 'pyvistaqt>=0.11.2',
-                'pywin32>=310; os_name == "nt"',
                 'rasterio>=1.4.3',
                 'rioxarray>=0.19.0',
                 'scikit-learn>=1.6.1',
