### 2025-05-21 #7019 @Patrick-Cole
H @boisgera ,

I merged the pull request, are the DOI warnings okay now?
