### 2025-05-21 2eeafbe
Batch export ternary bugfix
