### 2025-05-21 0363980
Update - added scroll bar to Advanced viewer options
