### 2025-05-23 801f2e2
Update
Added scrollbar to magnetic inversion
Fixed a bug when loading in projects
