### 2025-05-23 c175de8
add progress colour for context menus
