+++ pygmi/rsense/dataprep.py
@@ -427,7 +427,7 @@ def c_correction(data, dem, azimuth, zenith, *, showlog=print, piter=iter):
         m, b = np.polyfit(x, y, 1)
         c = b / m
 
-        print(f'zenith:{zenith} azimuth:{azimuth} c:{c}')
+        # print(f'zenith:{zenith} azimuth:{azimuth} c:{c}')
         # plt.figure(dpi=200)
         # plt.plot(x, y, '.')
         # trendpoly = np.poly1d((m, b))
