### 2025-05-26 de63340
Update
Added selected mtpy functions to solve conda problem.
Fixed occam display bug when apparent resistivities are all zero.
