+++ pygmi/mt/mtpy/analysis/distortion.py
@@ -1,0 +1,399 @@
+#!/usr/bin/env python
+
+"""
+mtpy/analysis/distortion.py
+
+Contains functions for the determination of (galvanic) distortion of impedance tensors.
+The methods used follow Bibby et al 2005.
+As it has been pointed out in that paper, there are various possibilities for
+constraining the solution, esp. in the 2D case.
+
+Here we just implement the 'most basic' variety for the calculation of the 
+distortion tensor. Other methods can be implemented, but since the optimal assumptions and
+constraints depend on the application, the actual place for further functions
+is in an independent, personalised module.
+
+Algorithm Details:
+Finding the distortion of a Z array. Using the phase tensor
+so, Z arrays are transformed into PTs first), following Bibby et al. 2005.
+
+First, try to find periods that indicate 1D. From them determine D incl.
+the g-factor by calculatiing a weighted mean. The g is assumed in order to
+cater for the missing unknown in the system, it is here set to det(X)^0.5.
+After that is found, the function no_distortion from the Z module can be
+called to obtain the unperturbated regional impedance tensor.
+
+Second, if there are no 1D sections: Find the strike angle, then rotate the
+Z to the principal axis. In order to do that, use the rotate(-strike) method
+of the Z module. Then take the real part of the rotated Z. As in the 1D case,
+we need an assumption to get rid of the (2) unknowns:
+set det(D) = P and det(D) = T, where P,T can be chosen. Common choice is to
+set  one of P,T to an arbitrary value (e.g. 1). Then check, for which values
+of the other parameter  S^2 = T^2+4*P*X_12*X_21/det(X) > 0 holds.
+
+@UofA, 2013  (LK)
+
+Edited by JP, 2016
+
+"""
+
+import copy
+
+import numpy as np
+
+import pygmi.mt.mtpy.analysis.geometry as MTge
+import pygmi.mt.mtpy.core.z as MTz
+import pygmi.mt.mtpy.utils.calculator as MTcc
+import pygmi.mt.mtpy.utils.exceptions as MTex
+
+
+def find_distortion(z_object, g='det', num_freq=None, lo_dims=None):
+    """
+    find optimal distortion tensor from z object
+
+    automatically determine the dimensionality over all frequencies, then find
+    the appropriate distortion tensor D
+
+    Parameters
+    ----------
+
+        **z_object** : mtpy.core.z object
+
+        **g** : [ 'det' | '01' | '10 ]
+                type of distortion correction
+                *default* is 'det'
+
+        **num_freq** : int
+                       number of frequencies to look for distortion from 
+                       the index 0
+                       *default* is None, meaning all frequencies are used
+
+        **lo_dims** : list
+                      list of dimensions for each frequency
+                      *default* is None, meaning calculated from data
+
+    Returns
+    -------
+
+        **distortion** : np.ndarray(2, 2)
+                         distortion array all real values
+
+        **distortion_err** : np.ndarray(2, 2)
+                             distortion error array
+
+
+    Examples
+    --------
+
+        :Estimate Distortion: ::
+
+            >>> import pygmi.mt.mtpy.analysis.distortion as distortion
+            >>> dis, dis_err = distortion.find_distortion(z_obj, num_freq=12)
+
+    """
+
+    if num_freq is not None:
[…]
