+++ pygmi/mt/mtpy/core/jfile.py
@@ -1,0 +1,328 @@
+# -*- coding: utf-8 -*-
+"""
+.. module:: JFile
+   :synopsis: Deal with J-Files of the format propsed by Alan Jones 
+
+.. moduleauthor:: Jared Peacock <jpeacock@usgs.gov>
+"""
+
+# ==============================================================================
+import numpy as np
+import os
+
+import pygmi.mt.mtpy.core.z as mtz
+# ==============================================================================
+# Class to read j_file
+# ==============================================================================
+
+
+class JFile(object):
+    """
+    be able to read and write a j-file
+    """
+
+    def __init__(self, j_fn=None):
+        self.j_fn = j_fn
+        self.header_dict = None
+        self.metadata_dict = None
+        self.Z = None
+        self.Tipper = None
+
+        if self.j_fn is not None:
+            self.read_j_file()
+
+    def _validate_j_file(self):
+        """
+        change the lat, lon, elev lines to something machine readable,
+        if they are not.
+        """
+
+        if not os.path.isfile(self.j_fn):
+            raise NameError('Could not find {0}, check path'.format(self.j_fn))
+
+        with open(self.j_fn, 'r', errors='replace') as fid:
+            j_lines = fid.readlines()
+
+        for variable in ['lat', 'lon', 'elev']:
+            for ii, line in enumerate(j_lines):
+                if variable in line.lower():
+                    name = line.split('=')[0]
+                    try:
+                        value = float(line.split('=')[1].strip())
+                    except ValueError:
+                        value = 0.0
+                        print('changed {0} to 0.0'.format(name[1:]))
+                    j_lines[ii] = '{0} = {1}\n'.format(name, value)
+                    break
+
+        return j_lines
+
+    def _read_header_line(self, line):
+        """
+        read a header line
+        """
+        line = ' '.join(line[1:].strip().split())
+
+        new_line = ''
+
+        # need to restructure the string so its readable, at least the way
+        # that birrp outputs the file
+        e_find = 0
+        for ii in range(len(line)):
+            if line[ii] == '=':
+                e_find = ii
+                new_line += line[ii]
+            elif line[ii] == ' ':
+                if abs(e_find-ii) == 1:
+                    pass
+                else:
+                    new_line += ','
+            else:
+                new_line += line[ii]
+
+        # now that we have a useful line, split it into its parts
+        line_list = new_line.split(',')
+
+        # try to split up the parts into a key=value setup
+        # and try to make the values floats if they can be
+        l_dict = {}
+        key = 'null'
+        for ll in line_list:
+            ll_list = ll.split('=')
+            if len(ll_list) == 1:
+                continue
+
+            # some times there is just a list of numbers, need a way to read
+            # that.
+            if len(ll_list) != 2:
+                if type(l_dict[key]) is not list:
+                    l_dict[key] = list([l_dict[key]])
+                try:
+                    l_dict[key].append(float(ll))
+                except ValueError:
+                    l_dict[key].append(ll)
+            else:
+                key = ll_list[0]
+                try:
+                    value = float(ll_list[1])
+                except ValueError:
+                    value = ll_list[1]
+
+                l_dict[key] = value
+
+        return l_dict
+
+    def read_header(self, j_lines=None):
+        """
+        Parsing the header lines of a j-file to extract processing information.
+
+        Input:
+        - j-file as list of lines (output of readlines())
+
+        Output:
+        - Dictionary with all parameters found
+
+        """
+        if j_lines is None:
+            j_lines = self._validate_j_file()
[…]
