+++ pygmi/mt/mtpy/utils/exceptions.py
@@ -1,0 +1,73 @@
+#!/usr/bin/env python
+
+"""
+Specific exceptions for MTpy.
+
+
+
+@UofA, 2013
+(LK)
+
+"""
+
+#=================================================================
+
+
+class MTpyError_float(Exception):
+    pass
+
+
+class MTpyError_inputarguments(Exception):
+    pass
+
+
+class MTpyError_value(Exception):
+    pass
+
+
+class MTpyError_ts_data(Exception):
+    pass
+
+
+class MTpyError_config_file(Exception):
+    pass
+
+
+class MTpyError_file_handling(Exception):
+    pass
+
+
+class MTpyError_edi_file(Exception):
+    pass
+
+
+class MTpyError_EDI(Exception):
+    pass
+
+
+class MTpyError_Z(Exception):
+    pass
+
+
+class MTpyError_Tipper(Exception):
+    pass
+
+
+class MTpyError_PT(Exception):
+    pass
+
+
+class MTpyError_occam(Exception):
+    pass
+
+
+class MTpyError_parameter_number(Exception):
+    pass
+
+
+class MTpyError_processing(Exception):
+    pass
+
+
+class MTpyError_module_import(Exception):
+    pass

[…]
