+++ pygmi/version.py
@@ -1,1 +1,1 @@
-__version__ = '3.2.9.6'
+__version__ = '3.2.9.7'
