+++ requirements.txt
@@ -20,4 +20,5 @@ scikit-image>=0.25.2
 shapelysmooth>=0.2.1
 simpeg>=0.24.0
 beautifulsoup4>=4.13.4
+pyyaml>=6.0.2
 
