+++ MANIFEST.in
@@ -5,6 +5,7 @@ include pygmi/mag/IGRF13.COF
 include pygmi/seis/descriptions.txt
 include pygmi/mt/mtpy/LICENSE.txt
 include pygmi/mt/mtpy/README.rst
+include pygmi/mt/mtpy/utils/epsg.npy
 recursive-include pygmi *.py *.png *.ico *.html
 prune pygmi/cgs
 
