+++ pyproject.toml
@@ -53,6 +53,7 @@ dependencies = [
                 'shapelysmooth>=0.2.1',
                 'simpeg>=0.24.0',
                 'beautifulsoup4>=4.13.4',
+                'pyyaml>=6.0.2',
                 ]
 
 [project.urls]
