### 2025-05-28 18de601
Added command line version information
