+++ pygmi/clust/super_class.py
@@ -29,7 +29,7 @@ import sys
 
 import numpy as np
 from PySide6 import QtWidgets, QtCore
-import matplotlib.pyplot as plt
+# import matplotlib.pyplot as plt
 from matplotlib.figure import Figure
 from matplotlib.patches import Polygon as mPolygon
 from matplotlib.lines import Line2D
@@ -209,6 +209,40 @@ class GraphMap(FigureCanvasQTAgg):
         self.im1.rgbmode = self.manip
         self.im1.set_data(data)
         self.im1.set_extent(extent)
+
+        self.ax1.xaxis.set_major_formatter(frm)
+        self.ax1.yaxis.set_major_formatter(frm)
+
+        self.figure.canvas.draw()
+
+    def update_class(self, dat):
+        """
+        Update plot.
+
+        Parameters
+        ----------
+        dat : dict
+            PyGMI dataset/s (pygmi.raster.datatypes.Data) in a dictionary.
+
+        Returns
+        -------
+        None.
+
+        """
+        clippercu = 0
+        clippercl = 0
+
+        data = dat
+        lclip, uclip = np.percentile(data.compressed(),
+                                     [clippercl, 100 - clippercu])
+
+        self.im1.set_clim(lclip, uclip)
+
+        # extent = dat.extent
+
+        self.im1.rgbmode = 'Single Colour Map'
+        self.im1.set_data(data)
+        # self.im1.set_extent(extent)
 
         self.ax1.xaxis.set_major_formatter(frm)
         self.ax1.yaxis.set_major_formatter(frm)
@@ -493,6 +527,7 @@ class SuperClass(BasicModule):
         self.c = [0, 1, 0]
         self.df = None
         self.data = {}
+        self.zonal = None
 
         self.map = GraphMap(self)
         self.dpoly = QtWidgets.QPushButton('Delete Polygon')
@@ -508,6 +543,8 @@ class SuperClass(BasicModule):
         self.cmb_band2 = QtWidgets.QComboBox()
         self.cmb_band3 = QtWidgets.QComboBox()
         self.cmb_manip = QtWidgets.QComboBox()
+        self.rb_data = QtWidgets.QRadioButton('Data')
+        self.rb_results = QtWidgets.QRadioButton('Results')
 
         self.mpl_toolbar = NavigationToolbar2QT(self.map, self.parent)
 
@@ -547,10 +584,16 @@ class SuperClass(BasicModule):
         self.cmb_manip.addItems(actions)
 
         vbl_1b.addWidget(self.cmb_manip)
+        vbl_1b.addWidget(self.rb_data)
+        vbl_1b.addWidget(self.rb_results)
+
+        self.rb_data.setChecked(True)
+        self.rb_results.setEnabled(False)
 
         loadshape = QtWidgets.QPushButton('Load Class Shapefile')
         saveshape = QtWidgets.QPushButton('Save Class Shapefile')
-        calcmetrics = QtWidgets.QPushButton('Calculate and Display Metrics')
+        calcmetrics = QtWidgets.QPushButton('Display Metrics')
+        pb_calc = QtWidgets.QPushButton('Calculate')
 
         self.setWindowTitle('Supervised Classification')
         self.tablewidget.setRowCount(0)
@@ -584,7 +627,7 @@ class SuperClass(BasicModule):
         gl_right.addWidget(self.tablewidget, 1, 0, 3, 2)
         gl_right.addWidget(self.apoly, 1, 2, 1, 1)
         gl_right.addWidget(self.dpoly, 2, 2, 1, 1)
-        gl_right.addWidget(calcmetrics, 3, 2, 1, 1)
+        # gl_right.addWidget(calcmetrics, 3, 2, 1, 1)
[…]
