+++ pygmi/raster/dataprep.py
@@ -37,7 +37,7 @@ import rasterio.merge
 from pyproj.crs import CRS
 from rasterio.warp import calculate_default_transform
 import geopandas as gpd
-from shapely import LineString
+from shapely import LineString, unary_union
 
 from pygmi.raster.datatypes import Data
 from pygmi.misc import ContextModule, BasicModule
@@ -1587,10 +1587,45 @@ def merge_max(merged_data, new_data, merged_mask, new_mask, index=None,
     merged_data[:] = tmp1
 
 
+def merge_order(ifiles, igeoms):
+    """Sort data in an order which ensures overlaps."""
+
+    # from shapely.plotting import plot_polygon
+    # import matplotlib.pyplot as plt
+
+    # fig = plt.figure()
+    # ax = plt.gca()
+    # plot_polygon(glist[0])
+    # plot_polygon(glist[1])
+    # plot_polygon(glist[2])
+    # plt.show()
+
+    ofiles = []
+    ogeoms = []
+
+    ofiles.append(ifiles.pop(0))
+    ogeoms.append(igeoms.pop(0))
+
+    while igeoms:
+        areas = []
+        master = unary_union(ogeoms)
+        for i in igeoms:
+            areas.append(master.intersection(i).area)
+
+        for i, area in enumerate(areas):
+            if area == max(areas):
+                ofiles.append(ifiles.pop(i))
+                ogeoms.append(igeoms.pop(i))
+                break
+    pass
+
+    return ofiles
+
+
 def mosaic(dat, *, idir=None, bfile=None, bandstofiles=False, piter=iter,
            showlog=print, singleband=False, forcetype=None,
            shifttomedian=False, tmpdir=None, nodata=None, method='first',
-           res=None):
+           res=None, ifiles=None):
     """
     Merge files with different numbers of bands and/or band order.
 
@@ -1599,7 +1634,8 @@ def mosaic(dat, *, idir=None, bfile=None, bandstofiles=False, piter=iter,
     Parameters
     ----------
     dat : list
-        List of PyGMI data bands to be merged. Can be empty if idir is provided.
+        List of PyGMI data bands to be merged. Can be empty if idir is
+        provided.
     idir : str, optional
         Directory where file to be mosaiced are found. The default is None.
     bfile : str, optional
@@ -1622,7 +1658,7 @@ def mosaic(dat, *, idir=None, bfile=None, bandstofiles=False, piter=iter,
         Nodata value. The default is None.
     method : str, optional
         Mosaic method. Can be 'first', 'last', 'merge_min', 'merge_max' or
-        'merge_median. The default is 'first'.
+        'merge_median'. The default is 'first'.
     res : float, optional
         Output resolution. Can be a tuple. The default is None.
 
@@ -1648,10 +1684,11 @@ def mosaic(dat, *, idir=None, bfile=None, bandstofiles=False, piter=iter,
         for i in dat['RasterFileList']:
             indata += get_from_rastermeta(i, piter=iter, metaonly=True)
 
-    if idir is not None:
-        ifiles = []
-        for ftype in ['*.tif', '*.hdr', '*.img', '*.ers']:
-            ifiles += glob.glob(os.path.join(idir, ftype))
+    if idir is not None or ifiles is not None:
+        if ifiles is None:
+            ifiles = []
+            for ftype in ['*.tif', '*.hdr', '*.img', '*.ers']:
+                ifiles += glob.glob(os.path.join(idir, ftype))
 
         if not ifiles:
             showlog('No input files in that directory')
@@ -1731,6 +1768,7 @@ def mosaic(dat, *, idir=None, bfile=None, bandstofiles=False, piter=iter,
 
[…]
