### 2025-06-13 22f02b3
Update
Fixed using some hyperspectral files with feature detection
