### 2025-06-13 abfea2b
Update CHANGES.rst
