+++ CHANGES.rst
@@ -2,7 +2,7 @@
 Changelog
 =========
 
-v3.2.9.6, 26 May 2025
+v3.2.9.7, 26 May 2025
 ---------------------
 * Added selected mtpy functions to solve conda problem.
 * Fixed occam display bug when apparent resistivities are all zero.
