### 2025-06-18 #7019 @boisgera
> I have updated the Zenodo title and created a 3.2.9.7 release on Github

Great, thank you.

@Patrick-Cole thanks a lot for this project and your reactivity during the review! :pray: @AnkitBarik, @jmtemmos thank you for your thoughtful reviews :pray:. I'll now recommend the acceptance of the paper and the Editor in Chief will take charge of the process.
