### 2025-06-18 #7019 @boisgera
OK, in the paper bibliography, there is *probably* an issue with the item:

```
@article{Horn1981,
   author = {Berthold K. P. Horn},
   doi = {10.1109/PROC.1981.11918},
   isbn = {0018-9219},
   issn = {00189219},
   issue = {1},
   journal = {Proceedings of the IEEE},
   keywords = {Hill shading,Reflectance map,Shaded overlays,To},
   pages = {14-47},
   title = {HILL SHADING AND THE REFLECTANCE MAP},
   volume = {69},
   url = {http://ieeexplore.ieee.org/lpdocs/epic03/wrapper.htm?arnumber=1456186},
   year = {1981},
}
```

AFAICT, what is registered as isbn is probably the issn (and I don't know what the isbn is).

@Patrick-Cole could you have a look at this?
