### 2025-06-19 a2f2aa7
Update paper.bib
