+++ paper/paper.bib
@@ -209,14 +209,13 @@ doi = {10.1186/s40623-015-0228-9}
 @article{Horn1981,
    author = {Berthold K. P. Horn},
    doi = {10.1109/PROC.1981.11918},
-   isbn = {0018-9219},
    issn = {00189219},
-   issue = {1},
    journal = {Proceedings of the IEEE},
    keywords = {Hill shading,Reflectance map,Shaded overlays,To},
    pages = {14-47},
-   title = {HILL SHADING AND THE REFLECTANCE MAP},
+   title = {Hill shading and the reflectance map},
    volume = {69},
+   number = {1},
    url = {http://ieeexplore.ieee.org/lpdocs/epic03/wrapper.htm?arnumber=1456186},
    year = {1981},
 }
