### 2025-06-30 525527e
update bib
