+++ paper/paper.bib
@@ -39,7 +39,7 @@
 }
 
 @article{COLE2021106219,
-  title = {Geometry of the Bushveld Complex from 3D potential field modelling},
+  title = {Geometry of the {B}ushveld {C}omplex from 3D potential field modelling},
   journal = {Precambrian Research},
   volume = {359},
   pages = {106219},
@@ -172,7 +172,7 @@ doi = {10.1186/s40623-015-0228-9}
    issue = {2},
    journal = {Economic Geology},
    pages = {209-228},
-   title = {Quantitative mineralogy from infrared spectroscopic data. I. Validation of mineral abundance and composition scripts at the rocklea channel iron deposit in Western Australia},
+   title = {Quantitative mineralogy from infrared spectroscopic data. I. Validation of mineral abundance and composition scripts at the {R}ocklea channel iron deposit in {W}estern {A}ustralia},
    volume = {107},
    year = {2012},
 }
@@ -185,7 +185,7 @@ doi = {10.1186/s40623-015-0228-9}
    journal = {Geophysics},
    month = {7},
    pages = {J25-J32},
-   title = {New standards for reducing gravity data: The North American gravity database},
+   title = {New standards for reducing gravity data: The {N}orth {A}merican gravity database},
    volume = {70},
    url = {https://library.seg.org/doi/10.1190/1.1988183},
    year = {2005},
@@ -224,7 +224,7 @@ doi = {10.1186/s40623-015-0228-9}
   author = {Musekiwa, C., Doggart, S., Cole, J., Janse van Rensburg, G., Phikiso, Z., Cole, P., Dudumashe, N., Sogayise, S.},
   year = {2023},
   month = {9},
-  title = {Multispectral and hyperspectral remote sensing for geology mapping: Namaqualand region, South Africa and Namibia.},
+  title = {Multispectral and hyperspectral remote sensing for geology mapping: {N}amaqualand region, {S}outh {A}frica and {N}amibia.},
   booktitle = {29th Colloquium of African Geology}, 
   city = {Windhoek, Namibia},
  }
@@ -248,7 +248,7 @@ author = {Pedregosa, Fabian and Varoquaux, Gael and Gramfort, Alexandre and Mich
 year = {2012},
 month = {01},
 pages = {},
-title = {Scikit-learn: Machine Learning in Python},
+title = {Scikit-learn: Machine Learning in {P}ython},
 volume = {12},
 journal = {Journal of Machine Learning Research}
 }
@@ -302,7 +302,7 @@ year = {2019}
 @software{gillies_2019,
   author =    {Sean Gillies and others},
   organization = {Mapbox},
-  title =     {Rasterio: geospatial raster I/O for {Python} programmers},
+  title =     {Rasterio: geospatial raster {I/O} for {Python} programmers},
   year =      {2013},
   url = "https://github.com/rasterio/rasterio"
 }
