+++ pygmi/mag/matchedfilt.py
@@ -33,15 +33,17 @@ import winsound
 
 import matplotlib.pyplot as plt
 import numpy as np
-import scipy.stats as stats
+from scipy.optimize import curve_fit
+
 
 from pygmi.raster.iodefs import get_raster
-from pygmi.raster.dataprep import fftprep, fft_getkxy
+from pygmi.raster.dataprep import fftprep
+from pygmi.vector.minc import minc
 
 
 def main():
     """Start of program."""
-    ifile = r"D:\workdata\PyGMI Test Data\Magnetics\IGRF\MAGMICROLEVEL.ers"
+    ifile = r"c:\work\PyGMI Test Data\Magnetics\IGRF\MAGMICROLEVEL.ers"
 
     dat = get_raster(ifile)
 
@@ -50,15 +52,24 @@ def main():
     xdim = data.xdim
     ydim = data.ydim
 
-    ndat, rdiff, cdiff, datamedian = fftprep(data)
+    ndat, rdiff, cdiff, datamedian = fftprepminc(data)
+    ndat2, rdiff, cdiff, datamedian = fftprep(data)
+    plt.subplot(121)
+    plt.imshow(ndat)
+    plt.subplot(122)
+    plt.imshow(ndat2)
+    plt.show()
 
     # Calculate the radially averaged power spectrum
     radial_bins, radial_mean = radial_average_power_spectrum(ndat, xdim, ydim)
-
+    radial_bins2, radial_mean2 = radial_average_power_spectrum(
+        ndat2, xdim, ydim)
     # Plot the result
     plt.figure()
     plt.plot(radial_bins, radial_mean,
              label="Radially Averaged Power Spectrum")
+    plt.plot(radial_bins2, radial_mean2,
+             label="Radially Averaged Power Spectrum old")
     plt.xlabel("Wavenumbers")
     plt.ylabel("Power")
     plt.title("Radially Averaged Power Spectrum")
@@ -68,27 +79,67 @@ def main():
     plt.show()
 
 
-def compute_1d_power_spectrum(data):
-    # Step 1: Perform 2D Fourier Transform
-    fft2_result = np.fft.fft2(data)
-    # Shift zero frequency to the center
-    fft2_shifted = np.fft.fftshift(fft2_result)
+def fftprepminc(data):
+    """
+    FFT preparation.
 
-    # Step 2: Compute the power spectrum
-    power_spectrum_2d = np.abs(fft2_shifted) ** 2
+    Parameters
+    ----------
+    data : pygmi.raster.datatypes.Data
+        Input dataset.
 
-    # Step 3: Create radial bins for averaging
-    ny, nx = data.shape
-    y, x = np.indices((ny, nx))
-    center_y, center_x = ny // 2, nx // 2
-    r = np.sqrt((x - center_x) ** 2 + (y - center_y) ** 2)
-    r = r.astype(int)
+    Returns
+    -------
+    zfin : numpy array.
+        Output prepared data.
+    rdiff : int
+        rows divided by 2.
+    cdiff : int
+        columns divided by 2.
+    datamedian : float
+        Median of data.
 
-    # Step 4: Radially average the power spectrum
-    radial_profile = np.bincount(
-        r.ravel(), weights=power_spectrum_2d.ravel()) / np.bincount(r.ravel())
+    """
+    datamedian = np.ma.median(data.data)
+    ndat = data.data - datamedian
 
-    return radial_profile
+    nr, nc = data.data.shape
[…]
