+++ pygmi/vector/show_table.py
@@ -158,7 +158,7 @@ def _testfn():
 
     sfile = r'D:\Work\Programming\geochem\all_geochem.shp'
 
-    _ = QtWidgets.QApplication(sys.argv)
+    app = QtWidgets.QApplication(sys.argv)
 
     tmp1 = ImportVector()
     tmp1.ifile = sfile
