### 2025-07-08 31f1edd
Update
Added CG-6 support to gravity processing
