+++ pygmi/grav/iodefs.py
@@ -48,9 +48,9 @@ class ImportCG5(BasicModule):
     def __init__(self, parent):
         super().__init__(parent)
 
-        self.df_cg5 = None
         self.df_gps = None
         self.is_import = True
+        self.filt = None
 
         self.cmb_line = QtWidgets.QComboBox()
         self.cmb_station = QtWidgets.QComboBox()
@@ -81,7 +81,7 @@ class ImportCG5(BasicModule):
         lbl_ychan = QtWidgets.QLabel('Latitude:')
         lbl_zchan = QtWidgets.QLabel('Ellipsoid (GPS) Elevation:')
         lbl_bthres = QtWidgets.QLabel('Minimum Base Station Number:')
-        pb_cg5 = QtWidgets.QPushButton('Load CG-5 File')
+        pb_cg5 = QtWidgets.QPushButton('Load CG-5/CG-6 File')
         pb_gps = QtWidgets.QPushButton('Load GPS File')
 
         pixmapi = QtWidgets.QStyle.StandardPixmap.SP_DialogOpenButton
@@ -97,7 +97,7 @@ class ImportCG5(BasicModule):
         self.cmb_ychan.setEnabled(False)
         self.cmb_zchan.setEnabled(False)
 
-        self.setWindowTitle(r'Import CG-5 Data')
+        self.setWindowTitle(r'Import CG-5/CG-6 Data')
 
         gl_main.addWidget(self.le_cg5file, 0, 1, 1, 1)
         gl_main.addWidget(pb_cg5, 0, 0, 1, 1)
@@ -175,11 +175,10 @@ class ImportCG5(BasicModule):
         cren[self.cmb_ychan.currentText()] = 'latitude'
         cren[self.cmb_zchan.currentText()] = 'elevation'
 
-        self.df_gps.rename(columns=cren, inplace=True)
-
         dfmerge = merge_gpsmag(self.le_cg5file.text(),
                                self.le_gpsfile.text(),
-                               float(self.le_basethres.text()), self.showlog)
+                               float(self.le_basethres.text()), cren,
+                               self.showlog)
 
         if dfmerge is False:
             return False
@@ -231,10 +230,10 @@ class ImportCG5(BasicModule):
         None.
 
         """
-        ext = 'CG-5 ASCII (*.txt *.xyz)'
+        ext = 'CG-5/CG-6 ASCII (*.txt *.xyz *.dat)'
 
         if filename == '':
-            filename, _ = QtWidgets.QFileDialog.getOpenFileName(
+            filename, self.filt = QtWidgets.QFileDialog.getOpenFileName(
                 self.parent, 'Open File', '.', ext)
             if filename == '':
                 return
@@ -327,7 +326,8 @@ def get_cg5(filename):
 
     Returns
     -------
-    None.
+    df_cg5 : Pandas DataFrame
+        Gravity data
 
     """
     os.chdir(os.path.dirname(filename))
@@ -359,6 +359,42 @@ def get_cg5(filename):
     return df_cg5
 
 
+def get_cg6(filename):
+    """
+    Get CG-6 filename and load data.
+
+    Parameters
+    ----------
+    filename : str
+        CG-6 filename.
+
+    Returns
+    -------
+    df : Pandas DataFrame
+        Gravity data
+
+    """
+    os.chdir(os.path.dirname(filename))
+
+    with open(filename, encoding='utf-8') as fno:
+        tmp = fno.readlines()
+
+    i = 0
+    while r'/Station' not in tmp[i]:
+        i += 1
+
[…]

[…]
